% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_humidity.R
\name{absolute_humidity}
\alias{absolute_humidity}
\title{Function to calculate absolute humidity from temperature and relative 
humidity.}
\usage{
absolute_humidity(temp, rh)
}
\arguments{
\item{temp}{Air temperature in degrees Celcius.}

\item{rh}{Relative humidity in percentage.}
}
\value{
Numeric vector, absolute humidity in g.m-3
}
\description{
Function to calculate absolute humidity from temperature and relative 
humidity.
}
\author{
Stuart K. Grange
}
