% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_by_id_and_range.R
\name{add_by_id_and_range}
\alias{add_by_id_and_range}
\title{Function to add a variable by testing an identifier and a range.}
\usage{
add_by_id_and_range(df, test, df_map, by, min, max, add)
}
\arguments{
\item{df}{Data frame to test.}

\item{test}{Variable in \code{df} for the range test.}

\item{df_map}{Mapping data frame containing \code{by}, \code{min}, \code{max},
and \code{add} variables.}

\item{by}{An identifying variable (or up to three) in both \code{df} and 
\code{df_map} which is to be used for the matching test.}

\item{min, max}{Variables in \code{df_map} which will be used for the range
test, i.e., is \code{test} between \code{min} and \code{max}?}

\item{add}{Variable in \code{df_map} to add to \code{df}. Generally, this 
will be an integer key.}
}
\value{
Tibble.
}
\description{
Function to add a variable by testing an identifier and a range.
}
\examples{

\dontrun{

# Add a site variable based on sensor_id and date ranges to an observation
# table containing time series
add_by_id_and_range(
  data_observations,
  test = "date",
  df_map = data_site_ranges,
  by = "sensor_id",
  min = "date_start",
  max = "date_end",
  add = "site"
)

}

}
\author{
Stuart K. Grange
}
