% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_monday_based_weeks.R
\name{add_monday_based_weeks}
\alias{add_monday_based_weeks}
\title{Function to add Monday-based week number and date ranges to a tibble.}
\usage{
add_monday_based_weeks(df)
}
\arguments{
\item{df}{A tibble/data frame including parsed dates. The date 
variable/column must be named \code{date}.}
}
\value{
\code{df} with four new variables located after the \code{date} 
variable.
}
\description{
Function to add Monday-based week number and date ranges to a tibble.
}
\author{
Stuart K. Grange
}
