% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_numbers.R
\name{add_row_numbers}
\alias{add_row_numbers}
\title{Function to add a row number variable.}
\usage{
add_row_numbers(df, name = "row_number")
}
\arguments{
\item{df}{Data frame to be transformed.}

\item{name}{Name of variable to use.}
}
\description{
Convert row numbers to a variable which is an integer, unlike 
\code{dplyr::add_rownames} which adds a character variable.
}
\examples{
\dontrun{
# Add row numbers
data_zones <- add_row_numbers(data_zones)

# Or change the name
data_zones <- add_row_numbers(data_zones, name = "order")

}

}
\author{
Stuart K. Grange
}

