% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streak.R
\name{add_steak_variables}
\alias{add_steak_variables}
\title{Function to add streak variables to a tibble.}
\usage{
add_steak_variables(df)
}
\arguments{
\item{df}{Tibble or data frame containing a logical \code{event} variable.}
}
\value{
Tibble, \code{df} with additional streak variables.
}
\description{
Function to add streak variables to a tibble.
}
\seealso{
\code{\link{streak}}, \code{\link{streak_max}}, 
\code{\link{streak_group}}
}
\author{
Stuart K. Grange
}
