% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_duplicated.R
\name{any_duplicated}
\alias{any_duplicated}
\title{Function to test if a vector contains duplicate elements.}
\usage{
any_duplicated(x)
}
\arguments{
\item{x}{A vector to test for duplicate elements.}
}
\value{
Logical vector with a length of \code{1}.
}
\description{
Unlike \code{\link{anyDuplicated}} which returns the index value of the first
duplicated element, \code{any_duplicated} is simpler and returns a logical
vector.
}
\author{
Stuart K. Grange
}
