% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hampel_identifier.R
\name{apply_hampel_identifier}
\alias{apply_hampel_identifier}
\title{Function to apply the Hampel identifier, a median absolute deviation (MAD) 
outlier detector.}
\usage{
apply_hampel_identifier(x, k = 60, t0 = 3, values = FALSE, na_preserve = TRUE)
}
\arguments{
\item{x}{A numeric vector, typically a time series.}

\item{k}{An integer representing the window length of \code{x} that is 
evaluated for the outlier detection.}

\item{t0}{The outlier threshold. The default of 3 is Ron Pearson's 
three-sigma edit rule.}

\item{values}{Should the filtered verson of the \code{x} vector be returned 
rather than a logical vector representing the identification of outliers?}

\item{na_preserve}{If \code{value} is \code{TRUE}, should the missing 
elements of \code{x} be propagated into the returned numeric vector?}
}
\value{
Numeric or logical vector depending on \code{values}.
}
\description{
The Hampel identifier is robust against outliers and omits all missing 
(\code{NA}) elements before calculation.
}
\seealso{
\url{https://github.com/cran/pracma}
}
\author{
Hans W. Borchers and Stuart K. Grange
}
