% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_left.R
\name{arrange_left}
\alias{arrange_left}
\title{Function to arrange variables to the left of a data frame.}
\usage{
arrange_left(df, variable)
}
\arguments{
\item{df}{Data frame which is to be transformed.}

\item{variable}{Variable(s) to be moved to the left side of \code{df}.}
}
\description{
\code{arrange_left} is useful when some variables/columns want to be moved to
the left-side of a data frame. \code{arrange_left} will silently skip 
variables if they do not have a match in \code{names(df)}.
}
\examples{
\dontrun{

# Move a single variable to the first position
data_ozone <- arrange_left(data_ozone, "date")

# Move many variables to left hand side of data frame
data_ozone <- arrange_left(data_ozone, c("date", "site", "site_name", "sensor"))

}

}
\author{
Stuart K. Grange
}

