% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.binary.R
\name{as.binary}
\alias{as.binary}
\title{Function to convert integers to binary strings.}
\usage{
as.binary(x, n = 32)
}
\arguments{
\item{x}{Integer to convert into its binary form.}

\item{n}{Number of digits/bits to return. Default is 32.}
}
\description{
\code{as.binary} converts integers to their binary representation.
\code{as.binary} is vectorised.
}
\details{
\code{as.binary} does not work for integers which are larger than 32 bit.
To-do: Fix this.

\url{http://stackoverflow.com/questions/6614283/converting-decimal-to-binary-in-r}
}
\examples{
\dontrun{

as.binary(5)
"00000000000000000000000000000101"

vector <- sample(1:1000, 5)
vector
721 409 703 538 806

as.binary(vector)
"00000000000000000000001011010001" "00000000000000000000000110011001"
"00000000000000000000001010111111" "00000000000000000000001000011010"
"00000000000000000000001100100110"

}
}
\author{
Stuart K. Grange
}

