% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_binary.R
\name{as_binary}
\alias{as_binary}
\title{Function to convert integers to binary strings.}
\usage{
as_binary(x, n = 32)
}
\arguments{
\item{x}{Integer to convert into its binary from.}

\item{n}{Number of digits/bits to return. Default is 32.}
}
\description{
\code{as_binary} converts integers to their binary representation. 
\code{as_binary} does not work for integers which are larger than 32 bit.
}
\examples{
\dontrun{

# Vector with length of one
as_binary(5L)

# Sampled vector 
as_binary(sample(1L:1000L, 1000))

}

}
\author{
Stuart K. Grange
}
\seealso{
\href{http://stackoverflow.com/questions/6614283/converting-decimal-to-binary-in-r}{stackoverflow}
}

