% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_zero_row_tibble.R
\name{build_zero_row_tibble}
\alias{build_zero_row_tibble}
\title{Function to create a zero row tibble, usually for pre-allocation of values.}
\usage{
build_zero_row_tibble(names)
}
\arguments{
\item{names}{Vector of names for the tibble.}
}
\value{
Tibble with names and zero rows.
}
\description{
Function to create a zero row tibble, usually for pre-allocation of values.
}
\author{
Stuart K. Grange
}
