% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ci.R
\name{calculate_ci}
\alias{calculate_ci}
\title{Function to calculate confidence intervals (CIs), usually based on population
standard deviation.}
\usage{
calculate_ci(x, level = 0.95)
}
\arguments{
\item{x}{Numeric vector.}

\item{level}{Confidence interval to calculate. This is a decimal, 
\strong{i.e.}, 0.95 is the 95 \% CI.}
}
\value{
Tibble.
}
\description{
The CI calculations use the \code{MeanCI} function written by 
Andri Signorell.
}
\examples{

# Calculate confidence intervals
x <- c(44617L, 7066L, 17594L, 2726L, 1178L, 18898L, 5033L, 37151L,  4514L, 4000L)
calculate_ci(x)

}
\author{
Stuart K. Grange and Andri Signorell.
}
