% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_control_limits.R
\name{calculate_control_limits}
\alias{calculate_control_limits}
\title{Function to calculate control limits, usually used for \code{XmR} control 
charts.}
\usage{
calculate_control_limits(x, control_constant = 1.128, control_multiplier = 3)
}
\arguments{
\item{x}{A numeric vector.}

\item{control_constant}{Constant used to transform the mean delta to 
sequential deviation. Through simulations, this is usually set as 1.128.}

\item{control_multiplier}{Multiplier for sequential deviation to get the 
lower and upper control limits, typically 3.}
}
\value{
Tibble.
}
\description{
Function to calculate control limits, usually used for \code{XmR} control 
charts.
}
\seealso{
\code{\link{plot_control_chart}}
}
\author{
Stuart K. Grange
}
