% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlations.R
\name{calculate_correlations}
\alias{calculate_correlations}
\title{Function to calculate correlations and return a tidy output.}
\usage{
calculate_correlations(df, type = "pearson")
}
\arguments{
\item{df}{Input data frame/tibble. Only numeric variables in \code{df} will
be used for the calculation.}

\item{type}{Type of correlation statistic: \code{"pearson"} or 
\code{"spearman"}.}
}
\value{
Tibble.
}
\description{
Function to calculate correlations and return a tidy output.
}
\author{
Stuart K. Grange
}
