% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_date_range.R
\name{calculate_date_range}
\alias{calculate_date_range}
\title{Function to summarise a date vector.}
\usage{
calculate_date_range(date)
}
\arguments{
\item{date}{A \code{POSIXct} or \code{Date} vector.}
}
\value{
A tibble with one row.
}
\description{
Function to summarise a date vector.
}
\seealso{
\code{\link{calculate_ci}}, \code{\link{calculate_quantiles}},
\code{\link{calculate_range}}
}
\author{
Stuart K. Grange.
}
