% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_date_summaries_wide.R
\name{calculate_date_summaries_wide}
\alias{calculate_date_summaries_wide}
\title{Function to calculate summaries of dates for a continuous time series when
the data are in "wide" format.}
\usage{
calculate_date_summaries_wide(
  df,
  ...,
  interval = "hour",
  use_data_table = TRUE,
  progress = FALSE
)
}
\arguments{
\item{df}{Input tibble. \code{df} must contain a \code{date} variable and at
least one variable with a a \code{numeric} data type.}

\item{...}{Grouping variables/specification for the summaries.}

\item{interval}{What interval should the aggregation be? English names and 
multiples of time intervals are accepted. For example, a valid \code{interval}
is \code{"5 mins"}.}

\item{use_data_table}{Should the \code{data.table} backend be used for the
aggregation calculations?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Tibble.
}
\description{
\code{calculate_date_summaries_wide} will pad the time series before 
aggregation to ensure all date steps are present in the return.
}
\seealso{
\code{\link{calculate_date_summaries}}
}
\author{
Stuart K. Grange
}
