% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deltas.R
\name{calculate_deltas}
\alias{calculate_deltas}
\title{Function to calculate deltas or differences between two vectors.}
\usage{
calculate_deltas(x, y, absolute = FALSE, round = NA)
}
\arguments{
\item{x, y}{Numeric vectors.}

\item{absolute}{Should deltas be represented absolutely (all positive)?}

\item{round}{Number of digits to round deltas to.}
}
\value{
Numeric vector
}
\description{
Function to calculate deltas or differences between two vectors.
}
\examples{

# Calculate errors 
calculate_deltas(sample(1:5), sample(1:5))

# Absolute
calculate_deltas(sample(1:5), sample(1:5), absolute = TRUE)

}
\author{
Stuart K. Grange.
}
