% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dew_point.R
\name{calculate_dew_point}
\alias{calculate_dew_point}
\title{Function to calculate the dew point from temperature and relative 
humidity.}
\usage{
calculate_dew_point(temp, rh)
}
\arguments{
\item{temp}{Air temperature in degrees Celsius.}

\item{rh}{Relative humidity in percentage.}
}
\value{
Numeric vector, dew point temperature in degrees Celsius.
}
\description{
Function to calculate the dew point from temperature and relative 
humidity.
}
\examples{

# Calculate dew point with temperature in degrees C and relative humidity
# in percent
calculate_dew_point(20, 65)

}
\author{
Stuart K. Grange
}
