% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dew_point.R
\name{calculate_mixing_ratio}
\alias{calculate_mixing_ratio}
\title{Function to calculate the mixing ratio of water vapour.}
\usage{
calculate_mixing_ratio(temp, pressure = 1013.25)
}
\arguments{
\item{temp}{Temperature degrees Celsius.}

\item{pressure}{Pressure in hectopascals (hPa).}
}
\value{
Numeric vector, mixing ratio in \code{g.kg-1 in dry air}.
}
\description{
The mixing ratio is the mass of water vapour in grams mixed into a kilogram 
of dry air.
}
\examples{

# Calculate the mixing ratio
calculate_mixing_ratio(temp = 20, pressure = 1013.25)

}
\seealso{
\code{\link{absolute_humidity}}, \code{\link{calculate_dew_point}},
\code{\link{calculate_vapour_pressure}}, \code{\link{specific_humidity}}
}
\author{
Stuart K. Grange
}
