% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pace_splits.R
\name{calculate_pace_splits}
\alias{calculate_pace_splits}
\title{Function to calculate splits for a running or cycling event.}
\usage{
calculate_pace_splits(distance, time, interval = 1)
}
\arguments{
\item{distance}{Target distance in kilometres.}

\item{time}{Target time to cover \code{distance} as a string. Use a format 
like \code{"01:30:00"} for one hour and thirty minutes.}

\item{interval}{Interval to calculate splits for in kilometres. Default is 
\code{1}.}
}
\value{
Data frame.
}
\description{
Function to calculate splits for a running or cycling event.
}
\examples{
\dontrun{

# Pace for a half marathon
# Every km
calculate_pace_splits(half_marathon(), "01:30:00", 1)

# Every 5 km
calculate_pace_splits(half_marathon(), "01:30:00", 5)


# Pace for 10 km 
calculate_pace_splits(10, "30:00")

# Pace for 5 km 
calculate_pace_splits(5, "18:15")

}


}
\author{
Stuart K. Grange
}

