% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_range.R
\name{calculate_range}
\alias{calculate_range}
\title{Function to summarise a numeric vector by calculating and minimum, maximum, 
range, and other summary statistics.}
\usage{
calculate_range(x, inf_to_na = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{inf_to_na}{Should any \code{Inf} or \code{-Inf} values be transformed
to \code{NA}? \code{Inf} values occur if there are no non-missing elements
in the vector and can cause issues plotting.}
}
\value{
A tibble with one row.
}
\description{
Function to summarise a numeric vector by calculating and minimum, maximum, 
range, and other summary statistics.
}
\examples{

# Calculate a few summaries with this vector
x <- c(44617L, 7066L, 17594L, 2726L, 1178L, 18898L, 5033L, 37151L,  4514L, 4000L)

# Calculate the summaries
calculate_range(x)

}
\seealso{
\code{\link{calculate_ci}}, \code{\link{calculate_quantiles}}
}
\author{
Stuart K. Grange.
}
