% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pace.R
\name{calculate_speed}
\alias{calculate_speed}
\alias{calculate_pace}
\title{Functions for calculating speed in m s-1 and pace in min km-1.}
\usage{
calculate_speed(seconds, metres)

calculate_pace(seconds, metres, round = NA)
}
\arguments{
\item{seconds}{Numeric vector of time in seconds.}

\item{metres}{Numeric vector of metres covered within the time period 
in \code{seconds}.}

\item{round}{For \code{calculate_pace}, the number of digits to round the 
output too.}
}
\value{
Numeric vector of speed in m s-1 or pace in min km-1.
}
\description{
Functions for calculating speed in m s-1 and pace in min km-1.
}
\examples{

# Calculate some speeds
calculate_speed(9.58, 100)
ms_to_km_h(calculate_speed(9.58, 100))

# Calculate pace of a run, the input takes seconds and metres
# Use seconds for time
calculate_pace(5400, half_marathon() * 1000)

# Or the hms data type
calculate_pace(parse_time("01:30:00"), half_marathon() * 1000)

}
\seealso{
\code{\link{decimal_minute_to_string}}
}
\author{
Stuart K. Grange
}
\keyword{cycling}
\keyword{pace,}
\keyword{running,}
\keyword{speed,}
\keyword{sport,}
