% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_time_span.R
\name{calculate_time_span}
\alias{calculate_time_span}
\title{Function to calculate time span between dates.}
\usage{
calculate_time_span(date_one, date_two, format = c("seconds", "hms",
  "duration", "period", "days"), round = NA, as.character = FALSE)
}
\arguments{
\item{date_one}{First date.}

\item{date_two}{Second date.}

\item{format}{Type of return.}

\item{round}{Number of digits to round return to.}

\item{as.character}{Should the return be a character vector.}
}
\value{
Vector length of 1. Can be of type numeric, hms, duration, period, 
or character.
}
\description{
Function to calculate time span between dates.
}
\examples{

# Define some dates
date_one <- lubridate::ymd_h("2019-02-03 5", tz = "UTC")
date_two <- lubridate::ymd_hms("2019-04-12 6:54:55.98565", tz = "UTC")

# Find spans in different formats
# Seconds as default
calculate_time_span(date_one, date_two)

# Other formats
calculate_time_span(date_one, date_two, format = "hms")
calculate_time_span(date_one, date_two, format = "duration")
calculate_time_span(date_one, date_two, format = "period")
calculate_time_span(date_one, date_two, format = "days")

# Use round to reduce precision
calculate_time_span(date_one, date_two, format = "period", round = 0)

}
\author{
Stuart K. Grange.
}
