% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_water_partial_pressure.R
\name{calculate_water_partial_pressure}
\alias{calculate_water_partial_pressure}
\title{Function to calculate the water partial pressure/vapour pressure of water 
using the August-Roche-Magnus (or Magnus-Tetens or Magnus) equation.}
\usage{
calculate_water_partial_pressure(temp, rh)
}
\arguments{
\item{temp}{Temperature in degrees Celsius -- usually air temperature.}

\item{rh}{Relative humidity in percent.}
}
\value{
Numeric vector with the units of hectopascal/millibar.
}
\description{
Function to calculate the water partial pressure/vapour pressure of water 
using the August-Roche-Magnus (or Magnus-Tetens or Magnus) equation.
}
