% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_water_volume_mixing_ratio.R
\name{calculate_water_volume_mixing_ratio}
\alias{calculate_water_volume_mixing_ratio}
\title{Function to calculate the mixing ratio of water.}
\usage{
calculate_water_volume_mixing_ratio(temp, rh, pressure, as_percent = FALSE)
}
\arguments{
\item{temp}{Temperature in degrees Celsius -- usually air temperature.}

\item{rh}{Relative humidity in percent.}

\item{pressure}{Pressure (usually air pressure) in the units of 
hectopascal/millibar.}

\item{as_percent}{Should the units used for the return be percentage rather 
than a ratio?}
}
\value{
Numeric vector with the units of a ratio or percentage if 
\code{as_percent} is \code{TRUE}.
}
\description{
\code{calculate_water_volume_mixing_ratio} uses the August-Roche-Magnus 
equation to calculate the partial pressure of air and this is divided by
air pressure to calculate the mixing ratio of water in dry air.
}
\examples{

# Calculate the mixing ratio of water from some observations in Zürich, at an
# elevation of ~ 430 m, therefore air pressure is relatively low
calculate_water_volume_mixing_ratio(13.3, 92.5, 976.5)

# Or as a percent
 calculate_water_volume_mixing_ratio(13.3, 92.5, 976.5, as_percent = TRUE)

}
\seealso{
\code{\link{calculate_water_partial_pressure}}
}
