% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculatePace.R
\name{calculatePace}
\alias{calculatePace}
\title{Function for calculating pace, usually in min km-1}
\usage{
calculatePace(time = time, distance = distance)
}
\description{
Function for calculating pace, usually in min km-1
}
\author{
Stuart Grange
}

