% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Function to chunk a vector into pieces of \emph{n} length.}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{Vector or list to be chunked into pieces.}

\item{n}{Length of chunk elements.}
}
\value{
List.
}
\description{
Function to chunk a vector into pieces of \emph{n} length.
}
\seealso{
\url{https://stackoverflow.com/questions/3318333/split-a-vector-into-chunks-in-r}
}
\author{
Harlan
}
