% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk_n}
\alias{chunk_n}
\title{Function to chunk a vector into \emph{n} pieces.}
\usage{
chunk_n(x, n)
}
\arguments{
\item{x}{Vector to be chunked.}

\item{n}{Number of chunks to split \code{x} into.}
}
\description{
Function to chunk a vector into \emph{n} pieces.
}
\author{
Joachim Schork
}
