% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clc.R
\name{clc}
\alias{clc}
\title{Function to clear the R console.}
\usage{
clc()
}
\description{
Found here: \url{http://stackoverflow.com/questions/14260340/function-to-clear-the-console-in-r}
}
\author{
Hadley Wickham
}

