% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gps_coordinates.R
\name{clean_gps_coordinates}
\alias{clean_gps_coordinates}
\title{Function to clean and convert non decimal degrees to decimal degrees.}
\usage{
clean_gps_coordinates(x, round = 6)
}
\arguments{
\item{x}{Vector containing coordinates in this format: 
\code{53 deg 37\' 56.49\" N}, usually from \code{\link{file_metadata}}.}

\item{round}{Number of decimal points to round the coordinates to.}
}
\description{
Function to clean and convert non decimal degrees to decimal degrees.
}
\seealso{
\code{\link{file_metadata}}, \code{\link{dms_to_decimal}}
}
\author{
Stuart K. Grange
}
