% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_message.R
\name{cli_n}
\alias{cli_n}
\title{Function to print the length of an object, generally for printing with 
\strong{cli} messaging.}
\usage{
cli_n(x, sep = FALSE)
}
\arguments{
\item{x}{Object to print length of.}

\item{sep}{Should the length be formatted with a thousands separator?}
}
\description{
\code{cli_n} will return the number of rows of a data frame, not the number 
of columns.
}
\author{
Stuart K. Grange
}
