% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours_for_plotting.R
\name{colour_byzantine_blue}
\alias{colour_byzantine_blue}
\alias{colour_crimson}
\alias{colour_inferno_purple}
\alias{colour_inferno_orange}
\alias{colour_inferno_peach}
\alias{colours_ggpubr}
\alias{colours_plotly}
\title{Functions to access common colours used for plotting.}
\usage{
colour_byzantine_blue()

colour_crimson()

colour_inferno_purple()

colour_inferno_orange()

colour_inferno_peach()

colours_ggpubr()

colours_plotly()
}
\value{
Character vector of various length.
}
\description{
Functions to access common colours used for plotting.
}
\author{
Stuart K. Grange
}
