% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours_brewer.R
\name{colours_brewer}
\alias{colours_brewer}
\title{Function to generate a vector of colours from ColorBrewer palettes.}
\usage{
colours_brewer(
  palette = c("Set1", "Set2", "Set3", "Pastel1", "Pastel2", "Dark2")
)
}
\arguments{
\item{palette}{ColorBrewer palette name.}
}
\value{
Character vector.
}
\description{
Function to generate a vector of colours from ColorBrewer palettes.
}
\seealso{
\code{\link[RColorBrewer]{brewer.pal}}, \url{http://colorbrewer.org}
}
\author{
Stuart K. Grange
}
