% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combineGeoJson.R
\name{combineGeoJson}
\alias{combineGeoJson}
\title{Function to read many \code{geojson} files and combine them as a
JavaScript variable.}
\usage{
combineGeoJson(dir.input = "~/Desktop/", var = "counties_data",
  dir.export = "~/Desktop/", remove = TRUE)
}
\description{
Useful for when using \code{geojson} for leafet.
}
\author{
Stuart Grange
}

