% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_growth.R
\name{compound_growth}
\alias{compound_growth}
\title{Function to calculate compound growth.}
\usage{
compound_growth(start, rate, steps = 1)
}
\arguments{
\item{start}{The starting condition where compound growth begins.}

\item{rate}{The growth rate.}

\item{steps}{The number of discrete growth iterations.}
}
\value{
Numeric vector.
}
\description{
Function to calculate compound growth.
}
\examples{

# Start at 15 and grow by 10 \% ten times
compound_growth(start = 15, rate = 0.1, steps = 10)

# Start at 15 and grow by 50 \% ten times
compound_growth(start = 15, rate = 0.5, steps = 10)

}
\author{
Stuart K. Grange
}
