% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertCoordinateSystem.R
\name{convertCoordinateSystem}
\alias{convertCoordinateSystem}
\title{Function to convert coordinate systems.}
\usage{
convertCoordinateSystem(x, to = "+init=epsg:4326")
}
\description{
Function to convert coordinate systems.
}
\author{
Stuart K. Grange
}

