% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wind_direction_bins.R
\name{create_wind_direction_bins}
\alias{create_wind_direction_bins}
\title{Function to transform numeric wind direction values (0--360 degrees) into 
bins.}
\usage{
create_wind_direction_bins(wd)
}
\arguments{
\item{wd}{A numeric vector representing wind direction values.}
}
\value{
A factor vector with levels.
}
\description{
Function to transform numeric wind direction values (0--360 degrees) into 
bins.
}
\examples{

# Bind wind directions
create_wind_direction_bins(seq(from = 0, to = 360, by = 10))

}
\author{
Stuart K. Grange
}
