% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_latex_table.R
\name{data_frame_to_latex_table}
\alias{data_frame_to_latex_table}
\title{Function to format a data frame to a LaTeX table.}
\usage{
data_frame_to_latex_table(
  df,
  caption = "Example caption...",
  label = "tab:",
  caption_location = "top",
  digits = NULL
)
}
\arguments{
\item{df}{Input data frame.}

\item{caption}{Text for LaTeX's table caption.}

\item{label}{Label for LaTeX's table.}

\item{caption_location}{Where to place the table's caption.}

\item{digits}{Number of digits to print for numeric variables.}
}
\value{
Invisible character vector.
}
\description{
\code{data_frame_to_latex_table} uses the \strong{xtable} package.
}
\examples{

# Format a data frame 
data_frame_to_latex_table(ggplot2::mpg)

}
\author{
Stuart K. Grange
}
