% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_timeseries.R
\name{data_frame_to_timeseries}
\alias{data_frame_to_timeseries}
\title{Function to promote a data frame to a time-series object.}
\usage{
data_frame_to_timeseries(df, tz = "UTC")
}
\arguments{
\item{df}{Data frame containing time-series data with a variable named 
\code{"date"}.}

\item{tz}{Time-zone of created time-series object. Default is \code{"UTC"}.}
}
\value{
Named list of time-series objects.
}
\description{
Function to promote a data frame to a time-series object.
}
\author{
Stuart K. Grange
}

