% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_centre.R
\name{date_centre}
\alias{date_centre}
\title{Function to calculate the centre point of two dates.}
\usage{
date_centre(date_start, date_end)
}
\arguments{
\item{date_start}{Start date.}

\item{date_end}{End date.}
}
\value{
\code{POSIXct} vector with a length of one.
}
\description{
Function to calculate the centre point of two dates.
}
\examples{

# Define some dates
date_one <- lubridate::ymd_h("2019-02-03 5", tz = "UTC")
date_two <- lubridate::ymd_hms("2019-04-12 6:54:55.98565", tz = "UTC")

# Calculate centre point
date_centre(date_one, date_two)

}
\author{
Stuart K. Grange.
}
