% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_message.R
\name{date_message}
\alias{date_message}
\alias{cli_date}
\alias{message_date_prefix}
\title{Functions to get a formatted date string, usually used as prefixes to 
messages.}
\usage{
date_message()

cli_date()

message_date_prefix()
}
\value{
Character vector.
}
\description{
Functions to get a formatted date string, usually used as prefixes to 
messages.
}
\examples{

# The message prefix
date_message()

# Within a message
message(date_message(), "Loading...")

}
\author{
Stuart K. Grange
}
