% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date_sequence.R
\name{date_sequence}
\alias{date_sequence}
\title{Function to generate a sequence of past or future dates.}
\usage{
date_sequence(days = -1, character = TRUE, today = TRUE)
}
\arguments{
\item{days}{Number of days of the sequence from system's date. Use negative
numbers to go back in time.}

\item{today}{Should system date be included in the returned sequence?}

\item{character.}{Should the date sequence be returned as a character
vector rather than a date vector? Default is TRUE.}
}
\description{
date_sequence is useful for generating patterns which match file-names.
}
\examples{
# Make a date sequence of the last three days and today
pattern <- date_sequence(days = -3)

# Display
pattern

"2015-06-22" "2015-06-23" "2015-06-24" "2015-06-25"
}
\author{
Stuart K. Grange
}

