% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_as_numeric.R
\name{dates_as_numeric}
\alias{dates_as_numeric}
\title{Function to convert all date variables in a data frame to their numeric 
representation.}
\usage{
dates_as_numeric(df, include_times = TRUE)
}
\arguments{
\item{df}{Data frame or tibble.}

\item{include_times}{Should times (\code{hms}) also be converted?}
}
\value{
Data frame or tibble.
}
\description{
Function to convert all date variables in a data frame to their numeric 
representation.
}
\author{
Stuart K. Grange
}
