% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_arrange_variables.R
\name{db_arrange_variables}
\alias{db_arrange_variables}
\title{Function to arrange variables in a data frame to match those found in a 
database table.}
\usage{
db_arrange_variables(con, table, df)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table name in database.}

\item{df}{Data frame.}
}
\description{
\code{db_arrange_variables} is useful for preparing data for insert and when
look-up tables are being created.
}
\author{
Stuart K. Grange
}

