% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_simple_wrappers.R
\name{db_clear_results}
\alias{db_clear_results}
\title{Function to clear results for a database connection.}
\usage{
db_clear_results(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Function to clear results for a database connection.
}

