% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_simple_wrappers.R
\name{db_commit}
\alias{db_commit}
\title{Function to commit transactions for a database connection.}
\usage{
db_commit(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Function to commit transactions for a database connection.
}

