% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Function to create a database connection with a JSON configuration file.}
\usage{
db_connect(file, database)
}
\arguments{
\item{file}{\code{JSON} file or string containing database connection 
details.}

\item{database}{The name of the database within \code{file} to use to create
a database connection to. If only one entry is in \code{file}, this argument 
is not needed and will be ignored if used.}
}
\description{
\code{db_connect} uses a \code{JSON} configuration file to create a database
connection. This configuration file will generally exist outside a code 
package so database credentials are not accidentally transmitted or shared.
}
\details{
If only one entry is in the \code{JSON} file, the \code{database} argument is
not needed.

MySQL and PostgreSQL connections are currently supported.
}
\examples{
\dontrun{
# Connect to an air quality database
db <- db_connect("connections.json", "air_quality")


# Use a json confg file which looks similar to this
string <- '
{
  "driver": "MySQL",
  "host": "172.31.4.159",
  "database_name": "database_seven",
  "user": "web",
  "password": "read_password"
}'

# Connect, no need for second argument when one connection is present in 
# configuration file
db_seven <- db_connect(string)

}

}
\author{
Stuart K. Grange
}

