% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_contents.R
\name{db_contents}
\alias{db_contents}
\title{Function to quickly return tables, variables, and values in a database.}
\usage{
db_contents(db, limit = 1)
}
\arguments{
\item{db}{Database connection}

\item{limit}{Maximum number rows to return in each table. Default is 1. Use 
\code{NA} to read entire table; use with caution.}
}
\description{
\code{db_contents} will return a tidy data frame with three variables, "table",
"variable", and "value". \code{db_contents} is useful to explore a database's
structure.
}
\examples{
\dontrun{
# Only one value will be returned
data_single <- db_contents(db)

# Entire table will be returned
data_all <- db_contents(db, limit = NA)
}

}
\author{
Stuart K. Grange
}

