% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_file_size.R
\name{db_file_size}
\alias{db_file_size}
\title{Function to get the file size of a SQLite database.}
\usage{
db_file_size(con, unit = "mb")
}
\arguments{
\item{con}{SQLite database connection.}

\item{unit}{Measurement unit. Default is \code{"mb"} for megabytes.}
}
\description{
Function to get the file size of a SQLite database.
}
\author{
Stuart K. Grange
}

