% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_activities.R
\name{db_list_activities}
\alias{db_list_activities}
\title{Function to list activities/processes for a database.}
\usage{
db_list_activities(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
\code{db_list_activities} only supports PostgreSQL databases at present.
}
\author{
Stuart K. Grange
}

