% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_constraints.R
\name{db_list_constraints}
\alias{db_list_constraints}
\title{Function to list table constraints for a database.}
\usage{
db_list_constraints(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
\code{db_list_constraints} only supports PostgreSQL databases at present.
}
\author{
Stuart K. Grange
}

