% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_indices.R
\name{db_list_indices}
\alias{db_list_indices}
\title{Function to list indices for a database tables.}
\usage{
db_list_indices(con, table = NA)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table name. If unused, all tables will be queried.}
}
\description{
Only SQLite databases are currently supported.
}
\author{
Stuart K. Grange
}

