% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_simple_wrappers.R
\name{db_list_results}
\alias{db_list_results}
\title{Function to list results for a database connection.}
\usage{
db_list_results(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Function to list results for a database connection.
}

