% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert.R
\name{db_list_variables}
\alias{db_list_variables}
\title{Function to list all variables/columns/fields in a database table.}
\usage{
db_list_variables(con, table)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table to list variables/columns/fields.}
}
\description{
\code{db_list_variables} is a wrapper for \code{DBI::dbListFields}.
}

