% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_optimise_table.R
\name{db_optimise_table}
\alias{db_optimise_all_tables}
\alias{db_optimise_table}
\title{Functions to optimise MySQL tables.}
\usage{
db_optimise_table(con, table, clear = TRUE)

db_optimise_all_tables(con, progress = "text")
}
\description{
\code{db_optimise_table} uses MySQL's \code{OPTIMIZE TABLE} function to 
analyse, vacuum, and optimise a table. \code{db_optimise_all_tables} is 
a vectorised version which will use the same \code{OPTIMIZE TABLE} function
and apply it to all tables in a database.
}
\author{
Stuart K. Grange
}

