% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_send.R
\name{db_send}
\alias{db_send}
\title{Function to send a statements to a database.}
\usage{
db_send(con, statement, progress = "none")
}
\arguments{
\item{con}{Database connection.}

\item{statement}{Statement to send to \code{con}.}

\item{progress}{Type of progress bar to display. Default is \code{"none"}.}
}
\description{
\code{db_send} is a wrapper for \code{DBI::dbSendQuery} but is vectorised 
over \code{statement}
}

