% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_table_names.R
\name{db_table_names}
\alias{db_table_names}
\title{Function to get the names of database table and produce a data frame with
zero rows.}
\usage{
db_table_names(db, table)
}
\arguments{
\item{db}{Database connection.}

\item{table}{Database table.}
}
\description{
\code{db_table_names} is useful when preparing a data frame for a SQL insert.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{dbWriteTable}}, \code{\link{rbind.fill}}
}

