% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_table_status.R
\name{db_table_status}
\alias{db_table_status}
\title{Function to get a table's status from a MySQL database.}
\usage{
db_table_status(db, table)
}
\arguments{
\item{db}{MySQL database connection.}

\item{table}{Database table.}
}
\description{
\code{db_table_status} queries a MySQL database for a table's status. 
See \url{https://dev.mysql.com/doc/refman/5.1/en/show-table-status.html} for 
the variables which are returned.
}
\author{
Stuart K. Grange
}

