% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_use_sql.R
\name{db_use_sql}
\alias{db_use_sql}
\title{Function to load SQL script and send it to a database.}
\usage{
db_use_sql(con, file)
}
\arguments{
\item{con}{Database connection.}

\item{file}{File name of SQL script.}
}
\description{
Function to load SQL script and send it to a database.
}
\author{
Stuart K. Grange
}

