% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_vacuum.R
\name{db_vacuum}
\alias{db_optimise}
\alias{db_vacuum}
\title{Functions to vacuum, analyse, and optimise database tables.}
\usage{
db_vacuum(con, table)

db_optimise(con, table)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Database table.}
}
\description{
The PostgreSQL statement uses \code{VACUUM (VERBOSE)} and the MySQL statement
uses \code{OPTIMIZE}.
}
\examples{
\dontrun{
db_vacuum_analyse(con, "hysplit_data")
db_optimise(con, "malta_data")
}

}
\author{
Stuart K. Grange
}

