% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_variables.R
\name{db_variables}
\alias{db_variables}
\title{Function to return all variable names in all tables in a SQL database.}
\usage{
db_variables(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
If there are complexities around schemas, PostGIS tables, or temporary tables, 
the table name is returned with a single variable of \code{NA}.
}
\examples{
\dontrun{
table_variables <- db_variables(con)

}

}
\author{
Stuart K. Grange
}

