% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_count.R
\name{decimal_count}
\alias{decimal_count}
\title{Function to count the number of digits on the right of a decimal point (the
mantissa).}
\usage{
decimal_count(x)
}
\description{
Function to count the number of digits on the right of a decimal point (the
mantissa).
}
\author{
Stuart K. Grange
}
\seealso{
\url{http://stackoverflow.com/questions/5173692/how-to-return-number-of-decimal-places-in-r}
}

