% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_minute_to_string.R
\name{decimal_minute_to_string}
\alias{decimal_minute_to_string}
\title{Function to transform decimal minutes into a formatted \code{"HH:MM:SS"} or
\code{"MM:SS"} string.}
\usage{
decimal_minute_to_string(x, hour = TRUE)
}
\arguments{
\item{x}{Numeric vector of decimal minutes.}

\item{hour}{Should the hour also be included in the formatted string?}
}
\description{
Function to transform decimal minutes into a formatted \code{"HH:MM:SS"} or
\code{"MM:SS"} string.
}
\examples{

decimal_minute_to_string(5)

decimal_minute_to_string(5.5)

decimal_minute_to_string(5.34)

decimal_minute_to_string(5.954226)

# Without the hour piece
decimal_minute_to_string(5.954226, hour = FALSE)


}
\author{
Stuart K. Grange
}
