% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_averaging_period.R
\name{detect_averaging_period}
\alias{detect_averaging_period}
\title{Function to determine averaging period in a date vector.}
\usage{
detect_averaging_period(date, skip = 1, n = 100)
}
\description{
Function to determine averaging period in a date vector.
}
\author{
Stuart K. Grange
}

