% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_number_plate.R
\name{detect_number_plate}
\alias{detect_number_plate}
\title{Function to detect number plates from an image.}
\usage{
detect_number_plate(file, region = "gb", warn = FALSE)
}
\arguments{
\item{file}{Vector of file names.}

\item{region}{Region of number plates.}

\item{warn}{Should the function raise a warning if unable to detect a number
plate?}
}
\description{
\code{detect_number_plate} is built on-top of 
(\href{https://github.com/openalpr/openalpr}{\code{openalpr} }) and this 
needs to be installed before this function will work.
}
\author{
Stuart K. Grange
}
