% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distance_by_haversine.R
\name{distance_by_haversine}
\alias{distance_by_haversine}
\title{Function to calculate distances between two points in metres or kilometres.}
\usage{
distance_by_haversine(latitude, longitude, latitude.2, longitude.2,
  unit = "metres", radius = 6371)
}
\arguments{
\item{latitude}{A vector of latitude values in decimal degrees.}

\item{longitude}{A vector of longitude values in decimal degrees.}

\item{latitude.2}{A vector of latitude values in decimal degrees.}

\item{longitude.2}{A vector of longitude values in decimal degrees.}

\item{unit}{The unit of the returned distance values. Can be metres or
kilometres. Metres is the default.}

\item{radius}{The radius of the Earth in kilometres. The default value is
6371, but I have seen versions of similar functions which use slightly
different values (plus or minus two or three kilometres) so this value can
be altered.}
}
\description{
\code{distance_by_haversine} is used when distances between latitude and
longitude-pairs need to be represented in metre or kilometre units.
}
\details{
\code{distance_by_haversine} is not a spatial function, it uses the
Haversine Formula which does not compensate for Earth's slightly
non-spherical shape. The Haversine Formula is most appropriate when used for
short-distance calculations and will overestimate distances in the polar
regions but underestimate distances near the equator. For more information
see: \url{https://en.wikipedia.org/wiki/Haversine_formula}.
}
\examples{
\dontrun{
# What is the distance between Oxford's Radcliffe Camera and Cliveden House
in Buckinghamshire?
# In metres
distance_by_haversine(51.753447, -1.254024, 51.558170, -0.688250)
44662.57

# In kilometres
distance_by_haversine(51.753447, -1.254024, 51.558170, -0.688250, "km")
44.66257

# or because we are in a locle with a different distance unit, in miles
distance_by_haversine(51.753447, -1.254024, 51.558170, -0.688250, "km") * 0.62
27.69079


# Data frame usage
# Distances of many air quality monitoring sites from the City of London
data.air$distance <- distance_by_haversine(
  data.air$latitude, data.air$longitude, 51.513468, -0.089133, "km")

# Arrange by distance, closest at the top
data.air <- plyr::arrange(data.air, distance)

}
}
\author{
Stuart K. Grange
}
\seealso{
See \code{\link{gDistance}}
}

