% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{download_file}
\alias{download_file}
\title{Function to download many files to a directory.}
\usage{
download_file(file, directory = NA, quiet = TRUE, progress = "none")
}
\arguments{
\item{file}{File name/URL of files to download.}

\item{directory}{Name of destination directory where files are to be downloaded
to. If \code{directory} is not used, the current working directory will be 
the destination.}

\item{quiet}{Should \code{download.file}'s messages be suppressed? Default is
\code{TRUE}.}

\item{progress}{Type of progress bar to display. Default is none, but it can
be \code{"text"} or \code{"time"} too.}
}
\description{
\code{download_file} uses \code{download.file} but can download many files 
and write them to a single directory. \code{download_file} will create the
destination directory if it does not exist.
}
\examples{
\dontrun{

# Download a single file
url <- "http://cdr.eionet.europa.eu/lv/eu/aqd/d/envvfk_0q/REP_D-LV_LEGMC_20150911_D-001.xml"
download_file(url, "downloaded/")

# Download many files with information
url <- c("http://cdr.eionet.europa.eu/lv/eu/aqd/d/envvfk_0q/REP_D-LV_LEGMC_20150911_D-001.xml",
         "http://cdr.eionet.europa.eu/lv/eu/aqd/g/envvnkijg/REP_D-LV_LEGMC_20151222_G-001.xml")
         
download_file(url, "downloaded/", quiet = FALSE)

}

}
\author{
Stuart K. Grange
}

