% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{download_ftp_file}
\alias{download_ftp_file}
\title{Function to download files from an FTP or SFTP server.}
\usage{
download_ftp_file(file, credentials = "", directory = NA, curl = TRUE,
  progress = "text")
}
\arguments{
\item{credentials}{Credentials for a FTP or SFTP server. Do not use 
\code{credentials} if the server does not require authentication. 
\code{credentials} takes the format: \code{"username:password"}.}

\item{directory}{Directory where the files are to be downloaded to. If 
\code{directory} does not exist, it will be created.}

\item{curl}{Should \strong{RCurl} be used to download the files or base R's 
\code{download.file}? Default is \code{TRUE}.}

\item{progress}{Progress bar type. Default is \code{"text"}, for no bar, use
\code{"none"}.}

\item{url}{The url(s) of the file(s) which are to be downloaded.}
}
\description{
Function to download files from an FTP or SFTP server.
}
\examples{
\dontrun{

# Set credentials in this format
credentials <- "username:password"

# Get the file "hour_5" in the directory "test" from a FTP server
download_ftp_file("ftp://195.174.23.76/test/hour_5.csv", 
                  credentials, "ftp_data/")
                  

# List files on a ftp server
list_files_ftp("ftp://195.174.23.76/test", credentials)
"ftp://195.174.23.76/test/hour_5.csv"

}

}
\seealso{
\strong{\link{RCurl}}, \code{\link{list_files_ftp}}, 
\code{\link{upload_ftp_file}}
}
\author{
Stuart K. Grange
}
