% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_to_temporary}
\alias{download_to_temporary}
\title{Function to download a file to R's temporary directory.}
\usage{
download_to_temporary(url, quiet = TRUE)
}
\arguments{
\item{url}{URL of a file to download.}

\item{quiet}{Should the function be quiet about its activities? Default is 
\code{TRUE}.}
}
\description{
Function to download a file to R's temporary directory.
}
\author{
Stuart K. Grange
}
