% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_with_scp.R
\name{download_with_scp}
\alias{download_with_scp}
\title{Function to download files from a remote system with \code{scp} (secure copy).}
\usage{
download_with_scp(host, file_remote, file_local, user, password,
  progress = "none")
}
\arguments{
\item{host}{Host name of remote system.}

\item{file_remote}{A vector of remote files names to download.}

\item{file_local}{A vector of file names which are the destination files for
\code{file_remote}.}

\item{user}{User name for \code{scp}.}

\item{password}{Password for the user for \code{scp}.}

\item{Type}{of progress bar to display.}
}
\value{
Invisible.
}
\description{
\code{download_with_scp} offers an alternative to the \code{curl} based 
functions which can be troublesome to use with the \code{sftp} protocol.
}
\examples{
\dontrun{

# Get a file
download_with_scp(
  host = "192.1.1.1",
  file_remote = "/network_storage/r_files/sock_data.rds", 
  file_local = "~/Desktop/sock_data_copied.rds",
  user = "username",
  password = "password_for_username"
)

}

}
\seealso{
\code{\link{list_files_scp}}
}
\author{
Stuart K. Grange
}
