% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drop_na_columns.R
\name{drop_na_columns}
\alias{drop_na_columns}
\title{Function to drop columns in a data frame which only contain NA.}
\usage{
drop_na_columns(df)
}
\description{
See: http://stackoverflow.com/questions/2643939/remove-columns-from-dataframe-where-all-values-are-na
}
\author{
teucer
}

