% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_rows.R
\name{duplicate_rows}
\alias{duplicate_rows}
\title{Function to return duplicated rows in a data frame.}
\usage{
duplicate_rows(df, variable)
}
\arguments{
\item{df}{Data frame.}

\item{variable}{A vector of variables to test for uniqueness.}
}
\description{
\code{duplicate_rows} can be though as the inverse of 
\code{\link{dplyr::distinct}}.
}
\author{
Stuart K. Grange
}

