% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_rows.R
\name{duplicated_rows}
\alias{duplicated_rows}
\title{Function to return duplicated rows in a data frame.}
\usage{
duplicated_rows(df, variable, invert = FALSE)
}
\arguments{
\item{df}{Data frame.}

\item{variable}{A vector of variables to test for uniqueness.}

\item{invert}{Should the function exclude duplicated. This is the same as
\code{\link{distinct}}.}
}
\description{
\code{duplicated_rows} can be though as the inverse of 
\code{\link{dplyr::distinct}}.
}
\author{
Stuart K. Grange
}

