% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_text.R
\name{evaluate_text}
\alias{evaluate_text}
\title{Function to evaluate and execute an R expression stored as text (a character
vector).}
\usage{
evaluate_text(text)
}
\arguments{
\item{text}{Text that is valid R code.}
}
\description{
\code{evaluate_text} uses \code{eval(parse(text = x))}.
}
\examples{

# Evaluate a simple expression
evaluate_text("5 + 5")

}
\author{
Stuart K. Grange
}
