% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_read.R
\name{excel_read}
\alias{excel_read}
\title{Function to wrap \code{readxl::read_excel} and then do some formating.}
\usage{
excel_read(file, sheet = 1, range = NULL, col_names = TRUE,
  col_types = NULL, na = "", trim_ws = TRUE, skip = 0, n_max = Inf,
  guess_max = min(1000, n_max), clean = TRUE, convert = FALSE,
  comment_character = NULL)
}
\arguments{
\item{file}{File name of \code{.xls} or \code{.xlsx} file.}

\item{sheet}{Sheet within \code{file} to read. \code{sheet} can be a name or 
an integer which indicates position.}

\item{col_names}{Use the first row as the data frame names?}

\item{col_types}{Data-types of columns.}

\item{na}{Missing value indicator. Default is \code{""}.}

\item{skip}{Number of rows to skip before reading tabular data.}

\item{n_max}{Maximum number of data rows to read.}

\item{guess_max}{Maximum number of data rows to use for guessing column types.}

\item{clean}{Should the names of the returned data frame be cleaned? Default
is \code{TRUE} and will return lower-case and underscore seperated names.}

\item{convert}{Should \code{type.convert} be run on the returned data frame? 
This can help to get correct data-types.}

\item{comment_character}{A single character which marks observations (rows)
which should be dropped.}
}
\description{
Function to wrap \code{readxl::read_excel} and then do some formating.
}
\author{
Stuart K. Grange
}
