% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_read.R
\name{excel_read_all}
\alias{excel_read_all}
\title{Function to read all sheets in a Microsoft Excel workbook.}
\usage{
excel_read_all(file, sheet = NA, convert = FALSE)
}
\description{
\code{excel_read_all} will only work well if the sheets have a consistent 
structure.
}

