% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_growth.R
\name{exponential_growth}
\alias{exponential_growth}
\title{Function to calculate exponential growth.}
\usage{
exponential_growth(start, rate, steps = 1)
}
\arguments{
\item{start}{The starting condition where exponential growth begins.}

\item{rate}{The growth rate or exponent to represent growth.}

\item{steps}{The number of discrete growth iterations.}
}
\value{
Numeric vector.
}
\description{
Exponential growth is also called geometric growth and has the property of 
the rate of change is proportional to the quantity itself.
}
\examples{

# Starting condition of 10000 with a low value for the exponential growth
exponential_growth(start = 10000, rate = 1.1, steps = 6)

# The same starting condition but with twice the growth rate, a huge change
exponential_growth(start = 10000, rate = 1.1 * 2, steps = 6)

}
\author{
Stuart K. Grange
}
