% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ggplot.R
\name{export_ggplot}
\alias{export_ggplot}
\title{Function to export \strong{ggplot2} plot, usually as a \code{.pdf} file.}
\usage{
export_ggplot(
  file = NA,
  plot = ggplot2::last_plot(),
  width = 6,
  height = 5,
  crop = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{file}{File name to export plot to, usually as a \code{.pdf} file.}

\item{plot}{\strong{ggplot2} object. If not used, the last plot will be 
retrieved.}

\item{width}{Plot width in inches.}

\item{height}{Plot height in inches.}

\item{crop}{Should the output be cropped? \strong{systemr} needs to be 
installed if \code{TRUE}.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \strong{ggplot2} object.
}
\description{
Function to export \strong{ggplot2} plot, usually as a \code{.pdf} file.
}
\author{
Stuart K. Grange
}
