% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_information.R
\name{file_information}
\alias{file_information}
\title{Function to get file information.}
\usage{
file_information(files, drop = FALSE)
}
\arguments{
\item{files}{Vector of file names.}
}
\value{
Data frame.
}
\description{
\code{file_infomation} uses \code{file.info} but cleans up the returned data
frame a little.
}
\author{
Stuart K. Grange
}
