% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{file_metadata}
\alias{file_metadata}
\title{Function to extract file metadata with \code{exiftool}.}
\usage{
file_metadata(file, progress = FALSE)
}
\arguments{
\item{file}{A vector of files to extract metadata from.}

\item{progress}{Should a progress bar be displayed?}
}
\description{
\code{file_metadata} uses Phil Harvey's Perl \code{exiftool} to extract a 
file's metadata and then formats the output into a table which can be 
conveniently used within R.
}
\details{
\code{file_metadata} can be useful for media files and the exploration of 
metadata. 

All variables which are supported and returned by \code{exiftool} will be 
present. Watch out for very large numbers of variables when imputing large
XML or JSON documents with a large numbers of nested elements.
}
\examples{

\dontrun{
# Get metadata file a single file
data_metadata <- file_metadata("music.mp3")
}

}
\seealso{
\href{http://www.sno.phy.queensu.ca/~phil/exiftool/}{exiftool}
}
\author{
Stuart K. Grange
}
