% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_size.R
\name{file_size}
\alias{file_size}
\title{Function to get file size.}
\usage{
file_size(file, unit = "mb")
}
\arguments{
\item{file}{File name.}

\item{unit}{Measurement unit. Default is \code{"mb"} for megabytes.}
}
\description{
Function to get file size.
}
\author{
Stuart K. Grange
}

