% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatHyperlink.R
\name{formatHyperlink}
\alias{formatHyperlink}
\title{Function to format a hyperline (href) string so when used in DT:dataTable or
googleVis::table, the link is active and click-able}
\usage{
formatHyperlink(x, check = TRUE)
}
\description{
Function to format a hyperline (href) string so when used in DT:dataTable or
googleVis::table, the link is active and click-able
}
\author{
Stuart K. Grange
}

