% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{french_verbs}
\alias{french_verbs}
\title{Conjugation of two common French verbs.}
\format{A data frame with 6 observations and three rows.}
\source{
\url{https://www.duolingo.com/}, within the first plural lesson.
}
\usage{
french_verbs
}
\description{
Conjugation of two common French verbs.
}
\keyword{datasets}

