% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_functions.R
\name{gam_smoother}
\alias{gam_smoother}
\title{Function to smooth an \emph{x}-\emph{y} pair with a generalised additive 
model (GAM).}
\usage{
gam_smoother(x, y, method = "GCV.Cp")
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{method}{Smoothing estimation method.}
}
\value{
Numeric vector with the length of \code{x} or \code{y}.
}
\description{
Function to smooth an \emph{x}-\emph{y} pair with a generalised additive 
model (GAM).
}
\author{
Stuart K. Grange
}
