% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_volume_to_mass.R
\name{gas_volume_to_mass}
\alias{gas_volume_to_mass}
\alias{gas_mass_to_volume}
\alias{milligram_to_microgram}
\alias{microgram_to_milligram}
\alias{ppb_to_ppm}
\alias{ppm_to_ppb}
\alias{ppt_to_ppb}
\alias{ppb_to_ppt}
\alias{convert_water_percent_to_mass}
\alias{convert_absolute_humidity_to_percent}
\title{Functions to convert gas concentrations between volume- and mass-units.}
\usage{
gas_volume_to_mass(
  volume,
  gas,
  molecular_mass = NA,
  unit_input = "ppb",
  unit_output = "ug_m3",
  temp = 0,
  pressure = 101325
)

gas_mass_to_volume(
  mass,
  gas,
  molecular_mass = NA,
  unit_input = "ug_m3",
  unit_output = "ppb",
  temp = 0,
  pressure = 101325
)

milligram_to_microgram(x)

microgram_to_milligram(x)

ppb_to_ppm(x)

ppm_to_ppb(x)

ppt_to_ppb(x)

ppb_to_ppt(x)

convert_water_percent_to_mass(
  percent,
  molecular_mass = 18.01528,
  temp = 0,
  pressure = 101325
)

convert_absolute_humidity_to_percent(
  humidity,
  molecular_mass = 18.01528,
  temp = 0,
  pressure = 101325
)
}
\arguments{
\item{volume}{Gas concentration in ppb or ppm.}

\item{gas}{A string such as \code{"o3"}, \code{"co"}, or \code{"co2"} which 
represents the input gas. This is used to find the molecular mass of a gas. 
Every gas-species will never be supported, therefore, use the 
\code{molecular_mass} argument if your gas is not implemented.}

\item{molecular_mass}{Molecular mass of gas in g mol-1. For example, for 
ozone, this is 48 and for carbon monoxide this is 28.01.}

\item{unit_input, unit_output}{Input and output units.}

\item{temp}{Default is 0 degrees Celsius. However, values such as 20 or 25 
degrees Celsius are often used.}

\item{pressure}{Default is 101325 Pa (1 standard atmosphere).}

\item{mass}{Gas concentration in ug m-3 or mg m-3.}

\item{x, percent, humidity}{Vector of gases to convert.}
}
\value{
Numeric vector.
}
\description{
Use \code{gas_volume_to_mass} when a gas concentration is in ppb or ppm and 
ug m-3 or mg m-3 is desired. Use \code{gas_mass_to_volume} when a gas 
concentration is in ug m-3 or mg m-3 and ppb or ppm is desired.
}
\details{
These functions allow for input of temperatures and pressures which change 
the coefficients used for the conversion between the unit systems, they do
not use conversion factors. 

\code{convert_water_percent_to_mass} will return humidity values in g.m-3 
and \code{convert_absolute_humidity_to_percent} assumes the humidty input is
also in g.m-3.
}
\examples{

# Ozone
# From ppb
gas_volume_to_mass(400, "o3")

# To ug m-3
gas_mass_to_volume(800, "o3")

# For monitoring sites in Europe where conversions are done at 20 degrees
gas_mass_to_volume(2.24, "no", temp = 20, unit_output = "ppb")
gas_mass_to_volume(24.53, "no2", temp = 20, unit_output = "ppb")
gas_mass_to_volume(37.74, "o3", temp = 20, unit_output = "ppb")

# CO usually uses different units
gas_mass_to_volume(
  mass = 0.39, 
  gas = "co", 
  unit_input = "mg_m3",
  temp = 20, 
  unit_output = "ppb"
 )

# NOx as NO2
gas_volume_to_mass(14.64, "no2", temp = 20, unit_input = "ppb")

# Convert water concentration in percent to g.m-3 at different temperatures
convert_water_percent_to_mass(1.18, temp = 0)
convert_water_percent_to_mass(1.18, temp = 25)

}
\seealso{
\code{\link{molecular_mass_table}}
}
\author{
Stuart K. Grange
}
