% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_all_combinations.R
\name{generate_all_combinations}
\alias{generate_all_combinations}
\title{Function to generate all combinations of the elements within a vector.}
\usage{
generate_all_combinations(x)
}
\arguments{
\item{x}{A vector to generate all combinations of elements of.}
}
\value{
List, with the length of 2 ^ length(\code{x}) - 1.
}
\description{
Function to generate all combinations of the elements within a vector.
}
\examples{

# Get all combinations of three letters
letters <- c("s", "t", "u")
generate_all_combinations(letters)

}
\seealso{
\code{\link{combn}}
}
\author{
Stuart K. Grange
}
