% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_database_names.R
\name{get_database_names}
\alias{get_database_names}
\title{Function to get the names of database table and produce a data frame with
zero rows.}
\usage{
get_database_names(db, table)
}
\arguments{
\item{db}{Database connection.}

\item{table}{Database table.}
}
\description{
\code{get_database_names} is very useful when preparing a data frame for an
SQL insert.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{dbWriteTable}}, \code{\link{rbind.fill}}
}

