% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_exchange_rate.R
\name{get_exchange_rate}
\alias{get_exchange_rate}
\title{Function get current exchange rates.}
\usage{
get_exchange_rate(currencies)
}
\arguments{
\item{currencies}{Currency-pairs to get.}
}
\description{
get_exchange_rate is a simple wrapper for \code{quantmod::getFX} which returns
a single exchange as a numeric value.
}
\examples{
\dontrun{
# Get exhange rate of the New Zealand Dollar to the British Pound
nzd.to.gbp <- get_exchange_rate("NZD/GBP")

# On 2015-07-08
nzd.to.gbp
0.4289
}
}

