% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixer_data.R
\name{get_fixer_codes}
\alias{get_fixer_codes}
\title{Function to get Fixer.io currency codes.}
\usage{
get_fixer_codes(lower = TRUE)
}
\description{
Function to get Fixer.io currency codes.
}
\author{
Stuart K. Grange
}

