% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixer_data.R
\name{get_fixer_data}
\alias{get_fixer_data}
\title{Function to get exchange rates from \href{http://fixer.io/}{Fixer.io}.}
\usage{
get_fixer_data(from, to, start = NA, end = NA, lower = TRUE,
  progress = "none")
}
\arguments{
\item{from}{Currency code used as base.}

\item{to}{Currency code as a function of \code{from}.}

\item{start}{Start date.}

\item{end}{End date.}

\item{lower}{Should the currency codes be forced to be lower-case?}

\item{progress}{Type of progress bar to display.}
}
\description{
Function to get exchange rates from \href{http://fixer.io/}{Fixer.io}.
}
\author{
Stuart K. Grange
}

