% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixer_data.R
\name{get_fixer_latest}
\alias{get_fixer_latest}
\title{Function to get lastest exchange rate data from 
\href{http://fixer.io/}{Fixer.io}.}
\usage{
get_fixer_latest(from = "usd", lower = TRUE)
}
\description{
Function to get lastest exchange rate data from 
\href{http://fixer.io/}{Fixer.io}.
}
\author{
Stuart K. Grange
}

