% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_version.R
\name{get_package_version}
\alias{get_package_version}
\title{Function to get package versions and return details in a tibble.}
\usage{
get_package_version(package, as_vector = FALSE)
}
\arguments{
\item{package}{Vector of package names to get details for. \code{package} can
also be \code{"R"}.}

\item{as_vector}{Should the package and version be formatted into a string?}
}
\value{
Tibble or character vector.
}
\description{
Function to get package versions and return details in a tibble.
}
\examples{

# Get package details
get_package_version(c("stats", "utils"))

# Get package details along with R's version
get_package_version(c("R", "stats", "utils"))

# Get details and return a vector
get_package_version(c("R", "stats", "utils"), as_vector = TRUE)

}
\author{
Stuart K. Grange
}
