% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precious_metals_prices.R
\name{get_precious_metals_prices}
\alias{get_precious_metals_prices}
\title{Function to get the cost of precious metals from 
\code{\link{http://services.packetizer.com/spotprices/}}.}
\usage{
get_precious_metals_prices()
}
\description{
Function to get the cost of precious metals from 
\code{\link{http://services.packetizer.com/spotprices/}}.
}
\examples{
\dontrun{

# Get precious metal prices
data_metals <- get_precious_metals_prices()

}

}
\author{
Stuart K. Grange
}

