% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_no_grid_lines.R
\name{gg_no_grid_lines}
\alias{gg_no_grid_lines}
\title{Function to remove all grid lines from a ggplot2 plot.}
\usage{
gg_no_grid_lines()
}
\value{
Invisible, a ggplot2 plot.
}
\description{
Function to remove all grid lines from a ggplot2 plot.
}
\author{
Stuart K. Grange
}
