% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotColours.R
\name{ggplot2Colours}
\alias{ggplot2Colours}
\title{Function to generate a vector of colours which emulate ggplot2's colours.}
\usage{
ggplot2Colours(n = 2)
}
\description{
http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
\author{
John Colby
}

