% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good_friday.R
\name{good_friday}
\alias{good_friday}
\alias{easter_sunday}
\alias{easter_monday}
\title{Functions to calculate the dates of shifting holidays.}
\usage{
good_friday(year = lubridate::year(lubridate::now()))

easter_sunday(year = lubridate::year(lubridate::now()))

easter_monday(year = lubridate::year(lubridate::now()))
}
\arguments{
\item{year}{Years to calculate dates for.}
}
\value{
\code{POSIXct} vector.
}
\description{
Functions to calculate the dates of shifting holidays.
}
\examples{

# Calculate some holiday dates
good_friday()
easter_sunday()
easter_monday()

}
