% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_vector.R
\name{\%within_vector\%}
\alias{\%within_vector\%}
\title{Function to vectorise \code{lubridate::`\%within\%`}.}
\usage{
date \%within_vector\% interval
}
\arguments{
\item{date}{Date vector.}

\item{interval}{Interval vector.}
}
\value{
Logical vector with the length of \code{date}.
}
\description{
Function to vectorise \code{lubridate::`\%within\%`}.
}
\examples{
\dontrun{

# Test dates within a data frame
data_temp$test <- interval_test(data_temp$date, interval)

}

}
\author{
Stuart K. Grange
}

