% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepl_all.R
\name{grepl_all}
\alias{grepl_all}
\title{Function to apply \code{grepl} across many variables in a data frame.}
\usage{
grepl_all(df, pattern, ignore.case = FALSE)
}
\arguments{
\item{df}{Data frame.}

\item{pattern}{Pattern to match with \code{\link{grepl}}.}

\item{ignore.case}{Use case sensitive matching? Default is \code{FALSE}.}
}
\value{
Logical vector with the length of \code{ncol(df)}.
}
\description{
\code{grepl_all} is useful for searching an entire data frame for a pattern
and can be used for filtering.
}
\author{
Stuart K. Grange
}
