% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{has_sub_seconds}
\alias{has_sub_seconds}
\title{Function to test a date (\code{}) vector for sub second accuracy.}
\usage{
has_sub_seconds(date)
}
\arguments{
\item{date}{A date vector.}
}
\value{
A logical vector with the length of 1.
}
\description{
Function to test a date (\code{}) vector for sub second accuracy.
}
\author{
Stuart K. Grange
}
