% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.binary.R
\name{integer_to_binary}
\alias{integer_to_binary}
\title{Main function to convert integers to binary
No export}
\usage{
integer_to_binary(x, n)
}
\description{
Main function to convert integers to binary
No export
}

