% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locf.R
\name{interpolate_na}
\alias{interpolate_na}
\title{Function for linearly interpolating \code{NA}s.}
\usage{
interpolate_na(x, na.rm = FALSE)
}
\arguments{
\item{x}{Input vector.}

\item{na.rm}{Should leading NAs be removed? Default is \code{FALSE}.}
}
\value{
Vector length of \code{x}.
}
\description{
Function for linearly interpolating \code{NA}s.
}
\author{
Stuart K. Grange
}
