% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_test.R
\name{interval_test}
\alias{interval_test}
\title{Function to test if a date is within a vector of intervals.}
\usage{
interval_test(df, df_look)
}
\arguments{
\item{df}{Input data frame.}

\item{df_look}{Invalidation data frame containing \code{"date_start"} and 
\code{"date_end"} variables.}
}
\value{
A logical vector.
}
\description{
Function to test if a date is within a vector of intervals.
}
\examples{
\dontrun{
# Load invalidation table and parse dates
data_invalid <- read.csv("weather_underground_invalidation_table.csv") \%>\% 
  mutate(date_start = ymd_hm(date_start),
         date_end = ymd_hm(date_end))
         
# Test dates
data_nelson$interval_test <- interval_test(data_nelson, data_invalid)

}

}
\author{
Stuart K. Grange
}

