% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Function to identify outliers in a numeric vector by uisng the same logic 
used for R's base boxplots.}
\usage{
is_outlier(x, coef = 10)
}
\arguments{
\item{x}{Numeric vector.}

\item{coef}{A coefficient which determines the classification of an outlier. 
Use 0 for no outliers to be identified.}
}
\value{
Logical vector with the length of \code{x}.
}
\description{
Function to identify outliers in a numeric vector by uisng the same logic 
used for R's base boxplots.
}
\seealso{
\code{\link{boxplot.stats}}.
}
\author{
Stuart K. Grange.
}
