% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outside_quantiles.R
\name{is_outside_quantiles}
\alias{is_outside_quantiles}
\title{Function to test elements of a vector are within a given quantile.}
\usage{
is_outside_quantiles(x, probabilities = c(0.01, 0.99), type = 7, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{probabilities}{A numeric vector with the length of \code{2} to use as 
the quantile limits.}

\item{type}{Method to use to calculate the quantiles.}

\item{na.rm}{Should \code{NA} values be omitted? The \code{\link{quantile}} 
function cannot be applied to vectors with \code{NA} values.}
}
\value{
Logical vector with length of \code{x}.
}
\description{
Function to test elements of a vector are within a given quantile.
}
\seealso{
\code{\link{quantile}}, \code{\link{calculate_quantiles}}
}
\author{
Stuart K. Grange
}
