% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_many.R
\name{join_many}
\alias{join_many}
\title{Function for joining many data frames together. Also called a multi-merge.}
\usage{
join_many(dfs, all = TRUE)
}
\arguments{
\item{dfs}{A \emph{list} of data frames.}

\item{all}{\code{merge}'s argument for a left-join.}
}
\description{
Function for joining many data frames together. Also called a multi-merge.
}
\examples{
\dontrun{
# Join three data frames together by a common key variable
data_join <- join_many(list(data_dates, data_min, data_max))
}

}
\author{
Stuart K. Grange
}

