% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_word_count.R
\name{latex_word_count}
\alias{latex_word_count}
\title{Function to count words and characters in a \code{LaTeX} (\code{.tex}) 
document.}
\usage{
latex_word_count(file, json = FALSE)
}
\arguments{
\item{file}{File name of \code{.tex} file(s).}

\item{json}{Should the return be pretty printed JSON?}
}
\value{
Data frame or pretty printed JSON.
}
\description{
Function to count words and characters in a \code{LaTeX} (\code{.tex}) 
document.
}
\author{
Stuart K. Grange
}
