% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_between.R
\name{left_join_between}
\alias{left_join_between}
\title{Function to join two tibbles/data frames together using a between clause.}
\usage{
left_join_between(df, df_y, by, match_fun, clean = TRUE, drop_between = TRUE)
}
\arguments{
\item{df, df_y}{Data frames/tibbles to join.}

\item{by}{What variables to use for the join.}

\item{match_fun}{The functions used for the joining. See examples for clarity.}

\item{clean}{Should the joined output be cleaned before being returned?}

\item{drop_between}{Should the variables used for the between/range test be
dropped?}
}
\value{
Tibble.
}
\description{
\code{left_join_between} is build upon the \strong{fuzzyjoin} package and 
\code{\link{findInterval}} function. \code{left_join_between} is also slow.
}
\examples{

\dontrun{

# Build a look-up table
data_invalidations <- dplyr::tribble(
  ~station_id, ~date_start, ~date_end,
  8208, "2019-12-07 16:00", "2019-12-07 17:15"
) \%>\% 
  mutate(valid = FALSE) \%>\% 
  mutate_at(c("date_start", "date_end"), ymd_hms, tz = "UTC", truncated = 3)

# Some time series data
data_time_series <- dplyr::tribble(
  ~station_id, ~date,                     ~value,
  8208L,       "2019-12-07 15:45:00 UTC", 1.753, 
  8208L,       "2019-12-07 16:00:00 UTC", 1.745, 
  8208L,       "2019-12-07 16:15:00 UTC", 0.075, 
  8208L,       "2019-12-07 16:30:00 UTC", 0.076, 
  8208L,       "2019-12-07 16:45:00 UTC", 0.075, 
  8208L,       "2019-12-07 17:00:00 UTC", 0.076, 
  8208L,       "2019-12-07 17:15:00 UTC", 1.701, 
  8208L,       "2019-12-07 17:30:00 UTC", 1.692, 
  8208L,       "2019-12-07 17:45:00 UTC", 1.683, 
  8208L,       "2019-12-07 18:00:00 UTC", 1.674, 
  8208L,       "2019-12-07 18:15:00 UTC", 1.668
) \%>\%
  mutate(date = as.POSIXct(date, tz = "UTC"))

# Do the join
data_time_series \%>\% 
  left_join_between(
  data_invalidations,
  by = c(
    "station_id" = "station_id",
    "date" = "date_start",
    "date" = "date_end"
  ),
  match_fun = list(`==`, `>=`, `<=`)
) 

}


}
\seealso{
\code{fuzzy_left_join}, \code{\link{findInterval}}
}
\author{
Stuart K. Grange
}
