% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_conditional.R
\name{left_join_conditional}
\alias{left_join_conditional}
\title{Function to join two data frames together based on a range condition.}
\usage{
left_join_conditional(x, y, x_test, y_min, y_max, keep = FALSE)
}
\arguments{
\item{x, y}{Data frames to join.}

\item{x_test}{Variable name in \code{x} which should be tested for a range.}

\item{y_min, y_max}{Variables in \code{y} which define the range for the test
against \code{x_test}.}

\item{keep}{Should the \code{y_min} and \code{y_max} be kept in the joined 
return?}
}
\description{
\code{left_join_conditional} will test if a variable in table \code{x} is 
within a range of two variables in table \code{y} and join the other 
variables if \code{TRUE}.
}
\examples{
\dontrun{

# Code urban area population into defined groups
data_join <- left_join_conditional(data_urban, data_urban_class, 
                                   "population", "size_min", "size_max")

}

}
\author{
Stuart K. Grange
}

