% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_functions.R
\name{linear_fit}
\alias{linear_fit}
\title{Function to predict the linear function of \emph{x}-\emph{y} pair.}
\usage{
linear_fit(x, y, robust = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{robust}{Should robust linear regression be used rather than ordinary 
least squares?}
}
\value{
Numeric vector with the length of \code{x} or \code{y}.
}
\description{
Function to predict the linear function of \emph{x}-\emph{y} pair.
}
\author{
Stuart K. Grange
}
