% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{Function to list files in a directory.}
\usage{
list_files(
  path = ".",
  pattern = NULL,
  all.files = FALSE,
  full.names = TRUE,
  recursive = FALSE,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = FALSE,
  fs_path = FALSE
)
}
\arguments{
\item{path}{Path of directory.}

\item{pattern}{Pattern for regular expression matching.}

\item{all.files}{Logical to force the return of hidden files.}

\item{full.names}{Logical to force the full file path to be returned. The 
default is \code{TRUE}.}

\item{recursive}{Logical to operate recursively.}

\item{ignore.case}{Logical for \code{pattern}.}

\item{include.dirs}{Should sub-directory names be included?}

\item{no..}{Should \code{"."} and \code{".."} be dropped from return?}

\item{fs_path}{Should the return be an \strong{fs} path? Useful for tidying 
paths and colour printing.}
}
\value{
Vector of file names or paths.
}
\description{
\code{list_files} is a wrapper for \code{\link{list.files}} but 
\code{full.names} has the default value of \code{TRUE}.
}
\author{
Stuart K. Grange
}
