% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{list_files_ftp}
\alias{list_files_ftp}
\title{Function to list files on an FTP or SFTP server.}
\usage{
list_files_ftp(url, credentials = "")
}
\description{
Function to list files on an FTP or SFTP server.
}

