% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_with_scp.R
\name{list_files_scp}
\alias{list_files_scp}
\title{Function to list files and directories on a remote system with \code{scp} 
(secure copy).}
\usage{
list_files_scp(host, directory_remote, user, password)
}
\arguments{
\item{host}{Host name of remote system.}

\item{directory_remote}{A remote directory to list files from.}

\item{user}{User name for \code{scp}.}

\item{password}{Password for the user for \code{scp}.}
}
\value{
Character vector.
}
\description{
\code{list_files_scp} offers an alternative to the \code{curl} based 
functions which can be troublesome to use with the \code{sftp} protocol on 
Ubuntu systems. \code{download_with_scp} needs the \code{sshpass} system 
programme to be installed.
}
\examples{
\dontrun{

# List contents of a directory
list_files_scp(
  host = "192.1.1.1", 
  directory_remote = "/network_storage/r_files/",
  user = "username",
  password = "password_for_username"
)

}

}
\seealso{
\code{\link{download_with_scp}}, \href{https://gist.github.com/arunoda/7790979}{sshpass},
\code{\link{upload_with_scp}}
}
\author{
Stuart K. Grange
}
