% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_functions.R
\name{loess_smoother}
\alias{loess_smoother}
\title{Function to smooth an \emph{x}-\emph{y} pair with loess (local regression).}
\usage{
loess_smoother(x, y, span = 0.75, degree = 2, surface = "interpolate")
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{span}{Smoother span.}

\item{degree}{Degree of the polynomials to be used.}

\item{surface}{A \code{loess.control} option. Can take the value 
\code{"interpolate"} or \code{"direct"}, If \code{"direct"}, the loess model
will predict outside its training space.}
}
\value{
Numeric vector with the length of \code{x} or \code{y}.
}
\description{
Function to smooth an \emph{x}-\emph{y} pair with loess (local regression).
}
\author{
Stuart K. Grange
}
