% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_to_yes_no.R
\name{logical_to_yes_no}
\alias{logical_to_yes_no}
\title{Function to convert a logical vector to a "Yes/No" vector.}
\usage{
logical_to_yes_no(x, as_factor = TRUE, na_as_false = FALSE)
}
\arguments{
\item{x}{Logical vector}

\item{as_factor}{Should the return be a factor?}

\item{na_as_false}{Should missing elements (\code{NA}) be as regarded as 
\code{FALSE}?}
}
\value{
Ordered factor or character vector.
}
\description{
Function to convert a logical vector to a "Yes/No" vector.
}
\examples{

# A vector
x <- c(TRUE, TRUE, TRUE, FALSE, TRUE, NA, FALSE, TRUE, TRUE, TRUE)

# To yes/no
logical_to_yes_no(x)

# Replace NAs with FALSE
logical_to_yes_no(x, na_as_false = TRUE)

# As a character vector
logical_to_yes_no(x, as_factor = FALSE)

}
\author{
Stuart K. Grange
}
