% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_geometric.R
\name{mean_geometric}
\alias{mean_geometric}
\title{Function to calculate the geometric mean of a numeric vector.}
\usage{
mean_geometric(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Should \code{NA}s be removed from the summary?}
}
\value{
Numeric vector with a length of one.
}
\description{
Function to calculate the geometric mean of a numeric vector.
}
\examples{

mean_geometric(sample(1:500, size = 10))

}
\seealso{
\url{https://github.com/cran/EnvStats/blob/master/R/geoMean.R}.
}
\author{
Stuart K. Grange and Steven P. Millard.
}
