% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_wd.R
\name{mean_wd}
\alias{mean_wd}
\title{Function to calculate mean wind direction.}
\usage{
mean_wd(wd, ws, na.rm = FALSE)
}
\arguments{
\item{wd}{Numeric vector representing wind direction in degrees.}

\item{ws}{Numeric vector representing wind speed. If \code{ws} is supplied, 
average wind directions will be weighted by wind speed and therefore will be
the "resultant vector average wind speed". This is usually considered the 
most appropriate aggregation for wind direction.}

\item{na.rm}{Should \code{NA}s be omitted from the calculation?}
}
\value{
Numeric vector with length of \code{1}.
}
\description{
Function to calculate mean wind direction.
}
\author{
Stuart K. Grange
}
