% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_date_formatted.R
\name{message_date_prefix}
\alias{message_date_prefix}
\title{Function to get a formatted date string used to prefix messages.}
\usage{
message_date_prefix()
}
\value{
Character vector
}
\description{
Function to get a formatted date string used to prefix messages.
}
\examples{

# The message prefix
message_date_prefix()

# Within a message
message(message_date_prefix(), "Loading...")

}
\author{
Stuart K. Grange
}
