% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{miles_to_kilometres}
\alias{bar_to_psi}
\alias{celsius_to_fahrenheit}
\alias{cubic_centimetres_to_cubic_inches}
\alias{cubic_inches_to_cubic_centimetres}
\alias{fahrenheit_to_celsius}
\alias{heat_index}
\alias{imperial_gallon_to_litres}
\alias{kilometres_to_miles}
\alias{litres_to_imperial_gallon}
\alias{litres_to_us_gallon}
\alias{miles_to_kilometres}
\alias{mpg_to_km_l}
\alias{mpg_to_l_100_km}
\alias{psi_to_bar}
\alias{us_gallon_to_litres}
\title{Functions to convert unit systems.}
\usage{
miles_to_kilometres(x)

kilometres_to_miles(x)

imperial_gallon_to_litres(x)

litres_to_imperial_gallon(x)

us_gallon_to_litres(x)

litres_to_us_gallon(x)

mpg_to_l_100_km(mpg, type = "imperial")

mpg_to_km_l(mpg, type = "imperial")

cubic_inches_to_cubic_centimetres(x)

cubic_centimetres_to_cubic_inches(x)

bar_to_psi(x)

psi_to_bar(x)

fahrenheit_to_celsius(x)

celsius_to_fahrenheit(x)

heat_index(temp, rh, unit = "c")
}
\description{
Functions to convert unit systems.
}
\author{
Stuart K. Grange
}

