% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{miles_to_km}
\alias{bar_to_psi}
\alias{celsius_to_fahrenheit}
\alias{cubic_centimetres_to_cubic_inches}
\alias{cubic_inches_to_cubic_centimetres}
\alias{fahrenheit_to_celsius}
\alias{foot_pound_to_newton_meter}
\alias{heat_index}
\alias{hp_to_kw}
\alias{imperial_gallon_to_litres}
\alias{km_h_to_km_min}
\alias{km_h_to_knots}
\alias{km_h_to_ms}
\alias{km_to_miles}
\alias{knots_to_km_h}
\alias{kw_to_hp}
\alias{l_100_km_to_100_l}
\alias{litres_to_imperial_gallon}
\alias{litres_to_us_gallon}
\alias{miles_to_km}
\alias{mpg_to_km_l}
\alias{mpg_to_l_100_km}
\alias{ms_to_km_h}
\alias{newton_metre_to_foot_pound}
\alias{newton_metre_to_kg_force}
\alias{psi_to_bar}
\alias{us_gallon_to_litres}
\title{Functions to convert unit systems.}
\usage{
miles_to_km(x)

km_to_miles(x)

knots_to_km_h(x)

km_h_to_knots(x)

ms_to_km_h(x)

km_h_to_ms(x)

km_h_to_km_min(x)

imperial_gallon_to_litres(x)

litres_to_imperial_gallon(x)

us_gallon_to_litres(x)

litres_to_us_gallon(x)

mpg_to_l_100_km(mpg, type = "imperial")

mpg_to_km_l(mpg, type = "imperial")

l_100_km_to_100_l(x)

cubic_inches_to_cubic_centimetres(x)

cubic_centimetres_to_cubic_inches(x)

bar_to_psi(x)

psi_to_bar(x)

fahrenheit_to_celsius(x)

celsius_to_fahrenheit(x)

heat_index(temp, rh, unit = "c")

kw_to_hp(x, metric = FALSE)

hp_to_kw(x, metric = FALSE)

newton_metre_to_kg_force(x)

newton_metre_to_foot_pound(x)

foot_pound_to_newton_meter(x)
}
\description{
Functions to convert unit systems.
}
\author{
Stuart K. Grange
}

