% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Function to calculate mode of a vector}
\usage{
mode(x)
}
\arguments{
\item{x}{Input vector.}
}
\description{
Function to calculate mode of a vector
}
\author{
Ken Williams
}
\seealso{
\href{http://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode}{stackoverflow}
}

