% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_average.R
\name{mode_average}
\alias{mode_average}
\title{Function to calculate mode of a vector.}
\usage{
mode_average(x, na.rm = FALSE)
}
\arguments{
\item{x}{Input vector.}

\item{na.rm}{Should \code{NA}s be omitted? Deafult is \code{FALSE}.}
}
\value{
Vector with a length of one.
}
\description{
Function to calculate mode of a vector.
}
\seealso{
\href{http://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode}{stackoverflow},
\code{\link{mean}}, \code{\link{median}}
}
\author{
Ken Williams with enhancement by Stuart K. Grange
}
