% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_numbers.R
\name{move_variable}
\alias{move_variable}
\title{Function to move a variable (column)  to the first position of a data frame.}
\usage{
move_variable(df, name)
}
\description{
Function to move a variable (column)  to the first position of a data frame.
}
\author{
Stuart K. Grange
}

