% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_extrapolate.R
\name{na_extrapolate}
\alias{na_extrapolate}
\title{Function to extrapolate leading and trailing \code{NA}s.}
\usage{
na_extrapolate(x)
}
\arguments{
\item{x}{Numeric or integer vector.}
}
\value{
Numeric vector.
}
\description{
\code{na_extrapolate} interpolates leading and trailing \code{NA}s with a 
linear function.
}
\examples{

na_extrapolate(c(NA, 1:5, rep(NA, 5)))

}
\seealso{
\code{\link{na_interpolate}}
}
\author{
Stuart K. Grange
}
