% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_locf.R
\name{na_interpolate}
\alias{na_interpolate}
\title{Function for linearly interpolating \code{NA}s.}
\usage{
na_interpolate(x, na.rm = FALSE)
}
\arguments{
\item{x}{Input vector.}

\item{na.rm}{Should leading NAs be removed?}
}
\value{
Vector length of \code{x}.
}
\description{
Function for linearly interpolating \code{NA}s.
}
\author{
Stuart K. Grange
}
