% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Function to normalise a numeric vector's range between 0 and 1.}
\usage{
normalise(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Remove NAs? Default is \code{TRUE}.}
}
\value{
Numeric vector with the length of \code{x}.
}
\description{
Function to normalise a numeric vector's range between 0 and 1.
}
\seealso{
\href{http://stats.stackexchange.com/questions/70801/how-to-normalize-data-to-0-1-range}{stackexchange}
}
\author{
Stuart K. Grange
}
