% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oil_ratio.R
\name{oil_ratio}
\alias{oil_ratio}
\title{Function to calculate oil/mixing ratios.}
\usage{
oil_ratio(ratio, volume_fuel = NA)
}
\arguments{
\item{ratio}{Oil/mixing ratio in fuel:oil order. For example, \code{"25:1"}
or \code{"100:1"}. \code{ratio} can be vector.}

\item{volume_fuel}{Volume of fuel in litres. If \code{volume_fuel} is not used
the a sequence of 1 to 10 litres will be used. \code{volume_fuel} can be 
vector.}
}
\value{
Tidy data frame with metric units.
}
\description{
\code{oil_ratio} is useful for 2-stroke fuel-oil mixing.
}
\examples{
\dontrun{

# Common mixing for garden equipment
oil_ratio(ratio = "50:1", volume_fuel = 5)

# Outboard mixing, depending on oil type and tank size
oil_ratio(ratio = c("50:1", "100:1"), volume_fuel = c(5, 10))

}

}
\author{
Stuart K. Grange
}

