% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_date_arguments.R
\name{parse_date_arguments}
\alias{parse_date_arguments}
\title{Function to parse date arguments for other functions.}
\usage{
parse_date_arguments(date, type, tz = "UTC")
}
\arguments{
\item{date}{Date string to be parsed.}

\item{type}{Type of date: \code{"start"} or \code{"end"}.}

\item{tz}{Time-zone, defaults to \code{"UTC"}.}
}
\value{
POSIXct vector with length of 1.
}
\description{
Function to parse date arguments for other functions.
}
\author{
Stuart K. Grange
}
