% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_elapsed_time.R
\name{parse_elapsed_time}
\alias{parse_elapsed_time}
\title{Function to parse elapsed time into seconds.}
\usage{
parse_elapsed_time(time)
}
\arguments{
\item{time}{String containing elapsed time.}
}
\description{
\code{parse_elapsed_time} parses strings such as \code{"01:24:05"} (hour-
minute-second) and \code{"54:11"} (minute-second). \code{parse_elapsed_time} 
is useful for sports.
}
\examples{
\dontrun{

# With hour
parse_elapsed_time("01:24:05")

# Without hour
parse_elapsed_time("54:11")

# Fractional seconds are preserved
parse_elapsed_time("01:12:11.889")
parse_elapsed_time("20:53.25")

}

}
\author{
Stuart K. Grange
}

