% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{parse_excel_date}
\alias{parse_excel_date}
\title{Function to parse Microsoft Excel's numeric date.}
\usage{
parse_excel_date(x, tz = "UTC", type = "windows")
}
\arguments{
\item{x}{Numeric vector.}

\item{tz}{Time-zone. Default is \code{"UTC"}.}

\item{type}{Type of Microsoft Excel date. Can be \code{"windows"} or 
\code{"os_x_2007"}.}
}
\value{
\code{POSIXct} vector.
}
\description{
Depending on what version of Microsoft Excel, there are two origins used.
}
\seealso{
\href{http://stackoverflow.com/questions/1703505/excel-date-to-unix-timestamp}{stackoverflow},
\code{\link{unix_time_to_excel_date}}
}
\author{
Stuart K. Grange
}
