% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_personal_path.R
\name{parse_personal_path}
\alias{parse_personal_path}
\title{Function to parse and switch common file paths for different systems.}
\usage{
parse_personal_path(path)
}
\arguments{
\item{path}{Path or file to be parsed and switched.}
}
\value{
\code{fs_path} vector.
}
\description{
Function to parse and switch common file paths for different systems.
}
\author{
Stuart K. Grange
}
