% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_time.R
\name{parse_time}
\alias{parse_time}
\title{Function to parse times into the \code{hms} data type.}
\usage{
parse_time(x)
}
\arguments{
\item{x}{Vector of time to be parsed.}
}
\value{
\code{hms} vector.
}
\description{
Function to parse times into the \code{hms} data type.
}
\examples{

# Some different time formats
times <- c(
  "44:41", "44:41.3", "01:38:44.566", "1:38:4", "24:00:00", "24:00", 
  "68:10", "78:12:12", "4:04.1", "4:04", NA, "110:42:00"
)

# Parse
parse_time(times)

# As seconds
as.numeric(parse_time(times))

# parse_time will also handle numeric vectors, there is an assumption that
# the elements are stored as seconds
parse_time(c(55, 555.9, 60))

}
\author{
Stuart K. Grange
}
