% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twos_complement.R
\name{parse_twos_complement}
\alias{parse_twos_complement}
\title{Main function
Parse Two's Complement integer, will convert to-and-from binary}
\usage{
parse_twos_complement(x)
}
\description{
Main function
Parse Two's Complement integer, will convert to-and-from binary
}

