% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_unit_text.R
\name{parse_unit_text}
\alias{parse_unit_text}
\title{Automatic text formatting for openair}
\usage{
parse_unit_text(text)
}
\arguments{
\item{text}{A character vector.}
}
\value{
The function returns an expression for graphical evaluation.
}
\description{
Workhorse function that automatically applies routine text formatting to
common expressions and data names used in openair.
}
\details{
\code{quickText} is routine formatting lookup table. It screens the
supplied character vector \code{text} and automatically applies formatting
to any recognised character sub-series. The function is used in a number of
\code{openair} functions and can also be used directly by users to format
text components of their own graphs (see below).
}
\author{
Karl Ropkins.
}
