% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{parse_unix_time}
\alias{parse_unix_time}
\title{Function to conveniently parse a vector of unix time to a POSIXct date vector.}
\usage{
parse_unix_time(x, tz = "UTC", origin = "1970-01-01")
}
\arguments{
\item{x}{An integer- or numeric-vector of unix times.}

\item{tz}{Time-zone string. \code{parse_unix_time} defaults to \code{"UTC"}.}

\item{origin}{Origin of epoch. By definition, unix time is \code{"1970-01-01"},
but other epochs are in use.}
}
\value{
A POSIXct vector with the length of x.
}
\description{
Function to conveniently parse a vector of unix time to a POSIXct date vector.
}
\examples{

# A vector
unix_time_vector <- c(1460034000, 1460034703)

# Parse time, will be in UTC time-zone
parse_unix_time(unix_time_vector)

# Or in Berlin's time-zone
parse_unix_time(unix_time_vector, tz = "Europe/Berlin")

}
\author{
Stuart K. Grange
}
