% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_write_metadata.R
\name{pdf_write_metadata}
\alias{pdf_write_metadata}
\title{Function to write metadata to a pdf file with \code{exiftool}.}
\usage{
pdf_write_metadata(
  file,
  author = NA,
  title = NA,
  subject = NA,
  producer = NA,
  creator = NA,
  verbose = FALSE
)
}
\arguments{
\item{file}{Vector of \code{.pdf} file names.}

\item{author}{Author to write to \code{file}.}

\item{title}{Title to write to \code{file}.}

\item{subject}{Subject to write to \code{file}.}

\item{producer}{Producer to write to \code{file}.}

\item{creator}{Creator to write to \code{file}.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \code{file}.
}
\description{
Function to write metadata to a pdf file with \code{exiftool}.
}
\examples{

\dontrun{
pdf_write_metadata("~/Desktop/document.pdf")
}

}
\author{
Stuart K. Grange
}
