% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_change.R
\name{percent_of}
\alias{percent_of}
\title{Function to calculate the value of a percentage of another value.}
\usage{
percent_of(value, percent)
}
\arguments{
\item{value}{Numeric vector of values.}

\item{percent}{Percentage of \code{value} to calculate. \code{percent} is not
a decimal, \emph{i.e.} five percent is represented by 5.}
}
\value{
Numeric vector with length of \code{value}.
}
\description{
Function to calculate the value of a percentage of another value.
}
\examples{

# Five percent of 33864
percent_of(33864, 5)

}
\author{
Stuart K. Grange
}
