% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_change.R
\name{percentage_change}
\alias{percentage_change}
\title{Function to calculate percentage change between new and old values.}
\usage{
percentage_change(previous, new, as_decimal = FALSE)
}
\arguments{
\item{previous}{Previous or older value.}

\item{new}{Newer value.}

\item{as_decimal}{Should the percentage be represented in a 0 to 1 scale?}
}
\value{
Numeric vector.
}
\description{
Function to calculate percentage change between new and old values.
}
\examples{

# Simple calculations
percentage_change(100, 50)
percentage_change(50, 100)

# Global mean CO2 mole fraction
percentage_change(277, 400)
percentage_change(277, 400, as_decimal = TRUE)

# Percentage change depends on the order of the inputs because it depends on
# the denominator
percentage_change(104, 130)
percentage_change(130, 104)

}
\seealso{
\code{\link{percentage_difference}}
}
\author{
Stuart K. Grange
}
