% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_monday.R
\name{period_financial}
\alias{period_financial}
\title{Function to transform date into financial period (a month with an offset).}
\usage{
period_financial(date, start = "july")
}
\arguments{
\item{date}{Vector of parsed dates}

\item{start}{The starting month of the financial year. Currently, the only
option is \code{"july"}.}
}
\value{
Integer vector.
}
\description{
Function to transform date into financial period (a month with an offset).
}
\examples{
\dontrun{

period_financial(data_hours$date)

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{wday}}, \code{\link{week_financial}}, \code{\link{month}}
}

