% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_to_string.R
\name{period_to_date}
\alias{period_to_date}
\title{Function convert a lubridate period to a POSIXct date.}
\usage{
period_to_date(period)
}
\arguments{
\item{period}{A vector of \strong{lubridate}'s periods.}
}
\description{
Function convert a lubridate period to a POSIXct date.
}
\author{
Stuart K. Grange
}
