% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_to_string.R
\name{period_to_string}
\alias{period_to_string}
\title{Function to format \strong{lubridate} periods into a \code{HH:MM:SS} format.}
\usage{
period_to_string(period)
}
\arguments{
\item{period}{A vector of \strong{lubridate}'s periods.}
}
\description{
Function to format \strong{lubridate} periods into a \code{HH:MM:SS} format.
}
\author{
Stuart K Grange
}
