% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_control_chart.R
\name{plot_control_chart}
\alias{plot_control_chart}
\title{Function to plot a control chart.}
\usage{
plot_control_chart(
  df,
  by = as.character(),
  size = 2,
  control_constant = 1.128,
  control_multiplier = 3,
  scales = "fixed"
)
}
\arguments{
\item{df}{Input data frame containing \code{value} and \code{date} variables.}

\item{by}{A character vector indicating the grouping variables within 
\code{df} to calculate control limits for and plot as facets.}

\item{size}{Size of the plots' points.}

\item{control_constant}{Constant used to transform the mean delta to 
sequential deviation. Through simulations, this is usually set as 1.128}

\item{control_multiplier}{Multiplier for sequential deviation to get the 
lower and upper control limits, typically 3.}

\item{scales}{A \strong{ggplot2} argument for setting the scales of the 
facets.}
}
\value{
ggplot2 plot.
}
\description{
Function to plot a control chart.
}
\seealso{
\code{\link{calculate_control_limits}}
}
\author{
Stuart K. Grange
}
