% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image_jpg.R
\name{plot_image_jpg}
\alias{plot_image_jpg}
\title{Function to plot a JPG/JPEG image.}
\usage{
plot_image_jpg(file)
}
\arguments{
\item{file}{Image file name.}
}
\value{
Invisible, a plot.
}
\description{
Function to plot a JPG/JPEG image.
}
\author{
Stuart K. Grange
}
