% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Function to plot time series data.}
\usage{
plot_time_series(
  df,
  colour = "#FC4E07",
  facet_variable = NA,
  size = 0.2,
  scales = "fixed",
  ylim = c(NA, NA)
)
}
\arguments{
\item{df}{Input data frame.}

\item{colour}{Colour of line geometry or which variable in \code{df} to code
for colour.}

\item{facet_variable}{Variable in \code{df} to facet the plot.}

\item{size}{Size/width of line geometry.}

\item{scales}{Should scales be fixed ("fixed", the default), free ("free"), 
or free in one dimension ("free_x", "free_y")?
#'}

\item{ylim}{Limits for y-axes.}
}
\value{
Invisible \strong{ggplot2} plot.
}
\description{
Function to plot time series data.
}
\examples{

# Build data frame with correct variables
df <- data.frame(
  date = lubridate::ymd(c("2019-01-01", "2019-01-25"), tz = "UTC"),
  value = 1:2
)

# Plot
plot_time_series(df)

}
\author{
Stuart K. Grange
}
