% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_variation.R
\name{plot_time_variation}
\alias{plot_time_variation}
\title{Function to plot hourly, weekday, and monthly averages.}
\usage{
plot_time_variation(
  df,
  by = NA,
  n_min = 2,
  colours = NA,
  ylim = c(NA, NA),
  legend_name = NA,
  y_label = NA,
  normalise = FALSE,
  labels = ggplot2::waiver(),
  legend_rows = 1,
  plot = TRUE
)
}
\arguments{
\item{df}{Input data frame with at least two variables, two of which are 
named \code{value} and \code{date}.}

\item{by}{A grouping variable to separately plot different variables.}

\item{n_min}{How many observations are needed to create a valid confidence 
interval? By default, at least two observations are needed to avoid very 
large bounds that results in bad y-axes scales.}

\item{colours}{A vector of colours to use for the plotting.}

\item{ylim}{Limits for y-axes.}

\item{legend_name}{A string to overwrite the default legend label.}

\item{y_label}{A string to overwrite the default y-axes labels.}

\item{normalise}{Should the aggregations be normalised by dividing the 
aggregations by their means? This helps when \code{by} is used and the groups
have different response scales.}

\item{labels}{If labels contain expressions to be parsed, set this argument 
to \code{scales::label_parse()}.}

\item{legend_rows}{How many rows should the legend be? Larger numbers are 
useful for long labels.}

\item{plot}{Should the plot be printed?}
}
\value{
A list object containing two objects -- the first element contains 
the aggregations that are calculated and plotted and the second contains the
plots. The plots are further split into an element that contains the four 
individual plots and the second is the four plots combined into a single 
object.
}
\description{
Function to plot hourly, weekday, and monthly averages.
}
\author{
Stuart K. Grange
}
