% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_palette.R
\name{print_palette}
\alias{print_palette}
\title{Function to plot a vector of colours.}
\usage{
print_palette(x)
}
\arguments{
\item{x}{Vector of colours.}
}
\value{
NULL.
}
\description{
Function to plot a vector of colours.
}
\seealso{
\url{https://github.com/karthik/wesanderson}
}
\author{
Stuart K. Grange and amd Karthik Ram.
}
