% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/properCase.R
\name{properCase}
\alias{properCase}
\title{Function to capitalise the first letter of every word in a string}
\usage{
properCase(x)
}
\description{
Wrapper for stringi::stri_trans_totitle
}
\examples{
\dontrun{
# Some strings
string <- c('auckland', 'wellington', 'berlin')

# Make proper case
properCase(string)
"Auckland"   "Wellington" "Berlin"
}
}
\author{
Stuart k. Grange
}

