% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\alias{gc_quiet}
\title{Function to catch \code{cat} messages and make them invisible.}
\usage{
quiet(x)

gc_quiet()
}
\arguments{
\item{x}{An expression.}
}
\value{
Invisible.
}
\description{
Function to catch \code{cat} messages and make them invisible.
}
