% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_remote.R
\name{read_csv_git_hub}
\alias{read_csv_git_hub}
\title{Function to read a \code{.csv} file and return as a data frame from a 
GitHub repository.}
\usage{
read_csv_git_hub(user = "skgrange", repository = "web.server",
  branch = "master", file)
}
\arguments{
\item{user}{GitHub user.}

\item{repository}{GitHub repository.}

\item{branch}{Repository's branch.}

\item{file}{Repository's file, including directory.}
}
\description{
Function to read a \code{.csv} file and return as a data frame from a 
GitHub repository.
}
\examples{
\dontrun{

# Use the defaults
data_hour <- read_csv_git_hub(file = "data/wacl/2016_kirb_hour_data.csv.bz2")

}


}
\author{
Stuart K. Grange
}
