% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_remote.R
\name{read_csv_remote}
\alias{read_csv_remote}
\title{Function to read a \code{.csv} file and return as a data frame from a 
\code{URL}.}
\usage{
read_csv_remote(url, verbose = FALSE)
}
\arguments{
\item{url}{Remote file to read.}

\item{verbose}{Should the function give messages?}
}
\description{
Function to read a \code{.csv} file and return as a data frame from a 
\code{URL}.
}
\author{
Stuart K. Grange
}
