% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_sheet.R
\name{read_excel_sheet}
\alias{read_excel_sheet}
\title{Function to wrap \code{readxl::read_excel}.}
\usage{
read_excel_sheet(file, sheet = 1, col_names = TRUE, col_types = NULL,
  na = "", skip = 0, clean = TRUE)
}
\description{
Function to wrap \code{readxl::read_excel}.
}

