% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json_lines.R
\name{read_json_lines}
\alias{read_json_lines}
\title{Function to read a new line JSON, also known as \code{ndjson} file.}
\usage{
read_json_lines(file, pagesize = 500, verbose = FALSE)
}
\arguments{
\item{file}{File name of new line JSON file.}

\item{pagesize}{Number of lines to write to \code{file} per iteration.}

\item{verbose}{Should the function give messages?}
}
\value{
An R data object, dependent on contents of \code{file}.
}
\description{
Function to read a new line JSON, also known as \code{ndjson} file.
}
\author{
Stuart K. Grange
}
