% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rdata.R
\name{read_rdata}
\alias{read_rdata}
\title{Function to read \code{RData} files but not their name so standard assignment 
can be used.}
\usage{
read_rdata(file)
}
\description{
If a single object needs to be saved in R's native format, use 
\code{\link{saveRDS}} rather than \code{\link{save}}.
}
\examples{
\dontrun{
# Load an RData file with standard assignment. 
data_air <- read_rdata("ADA16_2015.RData")

}

}
\author{
Stuart K. Grange
}

