% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rds_remote.R
\name{read_rds_remote}
\alias{read_rds_remote}
\title{Function to read an \code{.rds} file from a \code{URL}.}
\usage{
read_rds_remote(file_remote, quiet = TRUE)
}
\arguments{
\item{file_remote}{Remote file name.}

\item{quiet}{Should the function give messages?}
}
\value{
An R data class.
}
\description{
Function to read an \code{.rds} file from a \code{URL}.
}
\author{
Stuart K. Grange
}
