% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Function to read SQL script and store statments as a vector.}
\usage{
read_sql(file)
}
\value{
Character vector.
}
\description{
Mutli-line commenting is not correctly parsed yet.
}
\author{
Stuart K. Grange
}

