% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xl.R
\name{read_xl}
\alias{read_xl}
\title{Function to wrap \code{readxl::read_excel}.}
\usage{
read_xl(file, sheet = 1, col_names = TRUE, col_types = NULL, na = "",
  skip = 0)
}
\description{
Function to wrap \code{readxl::read_excel}.
}

