% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{read_xml}
\alias{read_xml}
\title{Function to load \code{XML} document as a list.}
\usage{
read_xml(file, clean_names = TRUE)
}
\arguments{
\item{file}{File name or character vector of an XML document.}

\item{clean_names}{Should duplicated names be made unique?}
}
\value{
Named list.
}
\description{
Function to load \code{XML} document as a list.
}
\author{
Stuart K. Grange
}
