% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/removeNACols.R
\name{removeNACols}
\alias{removeNACols}
\title{Function to drop columns in a data frame which only contain NA}
\usage{
removeNACols(x)
}
\description{
See: http://stackoverflow.com/questions/2643939/remove-columns-from-
dataframe-where-all-values-are-na
}
\author{
teucer
}

