% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_rows.R
\name{replicate_rows}
\alias{replicate_rows}
\title{Function to replicate rows in a data frame.}
\usage{
replicate_rows(df, n, reset = TRUE)
}
\arguments{
\item{df}{Data frame to be replicated.}

\item{n}{Number of times each row is to be replicated.}

\item{reset}{Should row names be reset? Default is \code{TRUE}.}
}
\description{
Function to replicate rows in a data frame.
}
\examples{
\dontrun{
data_replicated <- replicate_rows(data_test, 10)
}

}
\author{
Stuart K. Grange
}

