% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_row_names.R
\name{reset_row_names}
\alias{reset_row_names}
\title{Function to reset row names in a data frame.}
\usage{
reset_row_names(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Data frame.
}
\description{
\code{reset_row_names} can be used with magrittr's pipe (\code{\%>\%}).
}
\author{
Stuart K. Grange
}
