% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_na_rows.R
\name{rm_na_rows}
\alias{rm_na_rows}
\title{Function to remove rows in a data frame when indices are all \code{NA}.}
\usage{
rm_na_rows(df, index = NA)
}
\arguments{
\item{df}{Data frame to be filtered.}

\item{index}{Column index of \code{df} to be evaluated for \code{NA} testing. 
\code{index} usually takes multiple values and handles negative indices.}
}
\description{
\code{rm_na_rows} is useful when multiple variables/columns want to be 
evaluated for \code{NA}s and if all variables are \code{NA}, remove the row.
\code{rm_na_rows} avoids the use of many \code{!is.na(variable)} filters for 
wide data frames.
}
\examples{
\dontrun{
Remove rows which contain only NAs in all but the first two columns
data.clean <- rm_na_rows(data, -1:-2)
}

}
\author{
Stuart K. Grange
}

