% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\alias{mean_bias}
\alias{mean_bias_normalised}
\alias{correlation}
\alias{n_pairs}
\alias{mae}
\title{Functions to calculate various error statistics.}
\usage{
rmse(value_predict, value)

mean_bias(value_predict, value)

mean_bias_normalised(value_predict, value)

correlation(value_predict, value, method = "pearson")

n_pairs(value_predict, value)

mae(value_predict, value)
}
\arguments{
\item{value_predict}{A numeric vector to test against the "truth" 
(\code{value}).}

\item{value}{A numeric vector that is considered the "truth" or observed.}

\item{method}{For \code{correlation}, what method to use? Usually, this 
will be either \code{"pearson"} or \code{"spearman"}.}
}
\value{
Numeric vector with length of 1.
}
\description{
All functions will drop any missing elements before the calculation is done.
}
\author{
Stuart K. Grange
}
