% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_numeric.R
\name{round_numeric}
\alias{round_numeric}
\alias{factor_coerce}
\title{Function to round all numerical variables in a data frame.}
\usage{
round_numeric(df, round = 1)

factor_coerce(df)
}
\arguments{
\item{df}{A data frame containing numeric variables.}

\item{round}{Number of decimal places, used in \code{round}. Default is 1.}
}
\description{
\code{round_numeric} is useful when many variables are contained within a 
data frame  and all numerical values want to be rounded without individual 
naming.
}
\examples{

# Round all numeric variables to two decimal places
\dontrun{
data_tidy_round <- round_numeric(data_tidy, round = 2)

# or, make all numeric variables integers
data_tidy_integer <- round_numeric(data_tidy, round = 0)

}

}
\author{
Stuart K. Grange
}
