% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roundDateToInterval.R
\name{roundDateToInterval}
\alias{roundDateToInterval}
\title{Function to round dates to arbitrary time intervals such as 5-, 10-, 15-, or
30-minute periods.}
\usage{
roundDateToInterval(date, interval = "15 min")
}
\arguments{
\item{date}{A vector of dates to be rounded. This variable must be named
'date'.}

\item{interval}{A numeric or string value which represents the time interval
in minutes.}
}
\description{
roundDateToInterval rounds dates to the nearest arbitrary time interval
unlike xts::align.time which pushes dates to the next interval. Dates which
fall in the centre of the time interval are rounded down.
}
\details{
roundDateToInterval works by first calculating an origin which is the first
instant of the date's hour. roundDateToInterval then extracts the seconds-
past-the-origin of the date and rounds this value to the interval's multiple.
The rounded value and the origin are then summed and returned.

roundDateToInterval does not make any adjustments for duplicated dates in the
input or output.

roundDateToInterval can be used for rounding dates to the nearest whole
minute or hour, however, lubridate::round_date and related functions may be
more convenient.

The \code{interval} argument can take 'friendly' forms such as
\code{'5 mins'} or \code{'15 min'} aswell as \code{5} or \code{15}.
}
\examples{
\dontrun{
# A messy date vector
date.vector <- c('2015-04-09 06:25:23', '2015-04-09 06:41:27',
                 '2015-04-09 07:02:27', '2015-04-09 07:50:40')

# Parse dates
date.vector <- lubridate::ymd_hms(date.vector)

# Round dates to 15 minute intervals
roundDateToInterval(date.vector, '15 min')
"2015-04-09 06:30:00 UTC" "2015-04-09 06:45:00 UTC" "2015-04-09 07:00:00 UTC"
"2015-04-09 07:45:00 UTC"

# Round dates to 5 minute intervals
roundDateToInterval(date.vector, '5 mins')
"2015-04-09 06:25:00 UTC" "2015-04-09 06:40:00 UTC" "2015-04-09 07:00:00 UTC"
"2015-04-09 07:50:00 UTC"
}
}
\author{
Stuart K. Grange
}

