% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roundNumeric.R
\name{roundNumeric}
\alias{roundNumeric}
\title{Function to round all numerical variables in a data frame}
\usage{
roundNumeric(x, round = 2)
}
\arguments{
\item{x}{A data frame containing numeric vairables}

\item{round}{Number of decimal places, used in \code{base::round}}
}
\description{
Function to round all numerical variables in a data frame
}
\examples{
# Round all numeric variables to two decimal places
\dontrun{
data.tidy.round <- roundNumeric(data.tidy, round = 2)
}

# or, make all numeric variables integers
\dontrun{
data.tidy.integer <- roundNumeric(data.tidy, round = 0)
}
}
\author{
Stuart Grange
}

