% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vapour_pressure.R
\name{saturation_vapour_pressure}
\alias{saturation_vapour_pressure}
\title{Function to calculate the saturation vapour pressure of water vapour.}
\usage{
saturation_vapour_pressure(temp, type = 1L)
}
\arguments{
\item{temp}{Air temperature in degrees Celsius.}

\item{type}{What type of saturation vapour pressure should be calculated? 
Type \code{1} uses the Tetens equation with different logic for temperatures
below and above zero degrees and type \code{2} uses the  August-Roche-Magnus 
formula.}
}
\value{
Numeric vector, vapour pressure in hectopascals (hPa).
}
\description{
The saturation vapour pressure is the pressure at which water vapour is in 
thermodynamic equilibrium with its condensed state.
}
\author{
Stuart K. Grange
}
