% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_paths.R
\name{script_paths}
\alias{script_paths}
\title{Function to return a script's file name and directory from within itself.}
\usage{
script_paths()
}
\value{
Named list with two elements.
}
\description{
\code{script_paths} is most useful when calling a script as a programme.
}
\author{
Stuart K. Grange
}
