% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_wd.R
\name{sd_wind}
\alias{sd_wind}
\title{Function to calculate the standard deviation of wind direction based on 
Yamartino's 1984 method.}
\usage{
sd_wind(wd, na.rm = FALSE)
}
\arguments{
\item{wd}{Numeric vector representing wind direction in degrees.}

\item{na.rm}{Should \code{NA}s be omitted from the calculation?}
}
\value{
Numeric vector with length of \code{1}.
}
\description{
Function to calculate the standard deviation of wind direction based on 
Yamartino's 1984 method.
}
\author{
Stuart K. Grange
}
