% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{seconds_in_a_day}
\alias{seconds_in_a_day}
\alias{seconds_in_an_hour}
\alias{seconds_in_a_minute}
\alias{hours_in_a_year}
\alias{days_in_a_year}
\alias{seconds_in_a_year}
\title{Functions to conveniently access number of seconds in different time periods.}
\usage{
seconds_in_a_day()

seconds_in_an_hour()

seconds_in_a_minute()

hours_in_a_year(leap_year = FALSE)

days_in_a_year(leap_year = FALSE)

seconds_in_a_year(leap_year = FALSE)
}
\value{
Integer vecotor with length of one.
}
\description{
Functions to conveniently access number of seconds in different time periods.
}
\author{
Stuart K. Grange
}
