% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_fs_file_info.R
\name{select_fs_file_info}
\alias{select_fs_file_info}
\title{Function to select priority variables from the return of 
\code{fs::file_info}.}
\usage{
select_fs_file_info(df)
}
\arguments{
\item{df}{Data frame returned by \code{fs::file_info}.}
}
\value{
\code{df} with selected variables.
}
\description{
Function to select priority variables from the return of 
\code{fs::file_info}.
}
\author{
Stuart K. Grange
}
