% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialise_r_object.R
\name{serialise_r_object}
\alias{serialise_r_object}
\alias{unserialise_r_object}
\title{Functions to serialise and unseralise R objects.}
\usage{
serialise_r_object(x, compression = NA)

unserialise_r_object(x, compression = NA)
}
\arguments{
\item{x}{An R object.}

\item{compression}{Optional, a compression type to use for the serialisation
or unseralisation process. See \code{\link{memCompress}} for options.}
}
\value{
Raw vector.
}
\description{
Functions to serialise and unseralise R objects.
}
\seealso{
\code{\link{serialize}}, \code{\link{memCompress}}
}
