% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiftVar.R
\name{shiftVar}
\alias{shiftVar}
\title{Function for pushing elements backwards or forwards, i.e. generate lag
or lead variables
http://ctszkin.com/2012/03/11/generating-a-laglead-variables/}
\usage{
shiftVar(x, shift_by)
}
\description{
Function for pushing elements backwards or forwards, i.e. generate lag
or lead variables
http://ctszkin.com/2012/03/11/generating-a-laglead-variables/
}

