% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_moving_average.R
\name{simple_moving_average}
\alias{simple_moving_average}
\title{Function to apply rolling mean to a vector.}
\usage{
simple_moving_average(x, width, align = "center", na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector. The order of \code{x} should be checked to ensure 
behaviour is as expected.}

\item{width}{Width of margin to apply function.}

\item{align}{Should the index of the result be centered, left- or 
right-aligned? Default is \code{"center"}.}

\item{na.rm}{Should \code{NA}s be removed for the aggregation. Default is 
\code{TRUE}, but setting to \code{FALSE} will be faster.}
}
\description{
Function to apply rolling mean to a vector.
}
\seealso{
\code{\link{mean}}, \code{\link{rollapply}}, \code{\link{rollmean}}
}
\author{
Stuart K. Grange
}
