% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialPolygonsfromTidyData.R
\name{spatialPolygonsfromTidyData}
\alias{spatialPolygonsfromTidyData}
\title{Function for converting tidy data, usually post \code{ggplot2::fortify}
into spatial polygons. Object can be exported as geojson or retained as
an object.}
\usage{
spatialPolygonsfromTidyData(x, latitude = "lat", longitude = "long",
  export = FALSE, dir.output = "~/Desktop/", name = "Name")
}
\description{
spatialPolygonsfromTidyData correctly deals with holes.
}
\author{
Stuart Grange
}

