% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dew_point.R
\name{specific_humidity}
\alias{specific_humidity}
\title{Function to calculate specific humidity.}
\usage{
specific_humidity(temp, rh, pressure = 1013.25)
}
\arguments{
\item{temp}{Temperature degrees Celsius.}

\item{rh}{Relative humidity in percentage.}

\item{pressure}{Pressure in hectopascals (hPa).}
}
\value{
Numeric vector, specific humidity in \code{kg.kg-1}.
}
\description{
Function to calculate specific humidity.
}
\examples{

# Calculate specific humidities at different relative humidities
specific_humidity(20, 95:100)

# Calculate specific humidities at different relative humidities and convert
# to g.kg-1
specific_humidity(20, 95:100) * 1000

}
\seealso{
\code{\link{absolute_humidity}}, \code{\link{calculate_dew_point}},
\code{\link{calculate_vapour_pressure}}
}
\author{
Stuart K. Grange
}
