% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_nrow.R
\name{split_into_equal_parts}
\alias{split_into_equal_parts}
\title{Function to spit a data frame/tibble into equal sized parts/pieces.}
\usage{
split_into_equal_parts(df, n_parts = 10)
}
\arguments{
\item{df}{Data frame or tibble to be split.}

\item{n_parts}{Number of parts/pieces to split \code{df} into.}
}
\value{
A named list with the length of \code{n_parts}.
}
\description{
The final part may have a different number of rows than all other parts 
because of the length of the input cannot always be divided equality.
}
\author{
Stuart K. Grange
}
