% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{standardise}
\alias{standardise}
\title{Function to standardise all variables in a data frame.}
\usage{
standardise(df)
}
\arguments{
\item{df}{Data frame containing numeric variables.}
}
\description{
\code{standardise} will transform all numeric variables so they have a mean 
of 0 and a standard deviation of 1. This is also known as the z-score.
}
\examples{
\dontrun{
data_standardise <- standardise(data_csv)

}

}
\author{
and Stuart K. Grange
}
\seealso{
\link{scale}
}

