% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_utils_re_exports.R
\name{stats functions}
\alias{stats functions}
\title{Pseudo-function to re-export functions from the \strong{stats} package.}
\description{
Pseudo-function to re-export functions from the \strong{stats} package.
}
