% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_utils_re_exports.R
\name{stats grDevices}
\alias{stats grDevices}
\title{Pseudo-function to re-export functions from the \strong{grDevices} package.}
\description{
Pseudo-function to re-export functions from the \strong{grDevices} package.
}
