% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_date_formatted.R
\name{str_date_formatted}
\alias{str_date_formatted}
\title{Function to format a date (usually) for printing.}
\usage{
str_date_formatted(date = NA, time_zone = TRUE,
  fractional_seconds = TRUE)
}
\arguments{
\item{date}{Optional, a date to be formated.}

\item{time_zone}{Should the time zone be printed with the date?}

\item{fractional_seconds}{Should fractional seconds be included when printing
the date?}
}
\value{
Character vector.
}
\description{
Function to format a date (usually) for printing.
}
\author{
Stuart K. Grange
}
