% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_english_date_format.R
\name{str_english_date_format}
\alias{str_english_date_format}
\title{Function to format a date in a verbose, English language character string.}
\usage{
str_english_date_format(x, weekday = TRUE, seconds = TRUE,
  time_zone = TRUE, date_only = FALSE)
}
\arguments{
\item{x}{Date vector.}

\item{weekday}{Should the weekday be returned?}

\item{seconds}{Should seconds be printed?}

\item{time_zone}{Should the time zone string be printed?}

\item{date_only}{Should only the date pieces be returned?}
}
\value{
Character vector.
}
\description{
Function to format a date in a verbose, English language character string.
}
\examples{

# All information with seconds and time zone
str_english_date_format(lubridate::now())

# No seconds
str_english_date_format(lubridate::now(), seconds = FALSE)

# No seconds or time zone
str_english_date_format(lubridate::now(), seconds = FALSE, time_zone = FALSE)

}
\author{
Stuart K. Grange.
}
