% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_xml_attributes.R
\name{str_extract_xml_attributes}
\alias{str_extract_xml_attributes}
\title{Function to extract and format an XML attribute string into a key-value pair.}
\usage{
str_extract_xml_attributes(text)
}
\arguments{
\item{text}{Text to parse.}
}
\value{
Data frame.
}
\description{
Function to extract and format an XML attribute string into a key-value pair.
}
\author{
Stuart K. Grange
}
