% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_hms.R
\name{str_hms}
\alias{str_hms}
\title{Function to format a \code{hms} vector to a clean character vector.}
\usage{
str_hms(x, round = NA)
}
\arguments{
\item{x}{A \code{hms} vector.}

\item{round}{Number of decimal points to round to. The default is \code{NA}
which means no rounding.}
}
\value{
Character vector with length of \code{x}.
}
\description{
Function to format a \code{hms} vector to a clean character vector.
}
\author{
Stuart K. Grange
}
