% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_proper_case.R
\name{str_proper_case}
\alias{str_proper_case}
\title{Function to capitalise the first letter of every word in a string.}
\usage{
str_proper_case(x)
}
\description{
A simple wrapper for stringi::stri_trans_totitle.
}
\examples{
\dontrun{
# Some strings
string <- c('auckland', 'wellington', 'berlin')

# Make proper case
str_proper_case(string)
"Auckland"   "Wellington" "Berlin"
}
}
\author{
Stuart K. Grange
}

