% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_replace_multiple.R
\name{str_replace_multiple}
\alias{str_replace_multiple}
\title{Function to replace multiple vectors with multiple vectors.}
\usage{
str_replace_multiple(x, input = c("a", "b"), output = c("z", "y"),
  boundary = FALSE)
}
\description{
Not complete
}
\author{
Stuart Grange
}

