% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_rm_non_ascii.R
\name{str_rm_non_ascii}
\alias{str_rm_non_ascii}
\title{Function to remove all non-ASCII characters from a string.}
\usage{
str_rm_non_ascii(x)
}
\description{
\code{str_rm_non_ascii} is useful when data cleaning.
}
\author{
Stuart K. Grange
}

