% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_split_vector.R
\name{str_split_vector}
\alias{str_split_vector}
\title{Function to split string and return a length-of-one-vector}
\usage{
str_split_vector(x, pattern = "/", n = 1)
}
\description{
Uses stringr::str_split to split string
Returning more than one element at times
Not complete
}
\author{
Stuart K. Grange
}

