% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_system_date.R
\name{str_system_date}
\alias{str_system_date}
\title{Function to get system time with sytem calls.}
\usage{
str_system_date(tz = TRUE)
}
\arguments{
\item{tz}{Should the time-zone be included? Default is \code{TRUE}.}
}
\description{
Function to get system time with sytem calls.
}
\author{
Stuart K. Grange
}
