% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_re_exports.R
\name{stringr functions}
\alias{stringr functions}
\title{Pseudo-function to re-export \strong{stringr}'s common functions.}
\description{
Pseudo-function to re-export \strong{stringr}'s common functions.
}
