% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_to_observations.R
\name{strings_to_observations}
\alias{strings_to_observations}
\title{Functions to transform a delimited string into observations (rows) in a data
frame.}
\usage{
strings_to_observations(df, variable, pattern, trim = FALSE, reset = TRUE)
}
\arguments{
\item{df}{Data frame which contains a delimited string variable.}

\item{variable}{Name of variable in \code{df} which is, or contains delimited 
strings.}

\item{pattern}{Delimiter to split the string. Common values are \code{","} and
\code{"/"}.}

\item{trim}{Should the new vector be trimmed so there is no leading or 
trailing whitespace? Default is \code{FALSE}.}

\item{reset}{Should the \code{row.names} of the data frame be reset after 
replicating rows. Default is \code{TRUE}.}
}
\description{
\code{strings_to_observations} works by first splitting a string variable by a 
pattern. The split object is then used to replicate all other variables in a
data frame. \code{strings_to_observations} is useful for making data tidy.
}
\examples{
\dontrun{
# Make tidy data
data_tidy <- strings_to_observations(french_verbs, variable = "subject", 
                                     pattern = ",")
}

}
\author{
Stuart K. Grange
}

