% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{symbol_degree}
\alias{symbol_degree}
\alias{symbol_mu_lower}
\alias{symbol_subscript_1}
\alias{symbol_subscript_2}
\alias{symbol_subscript_4}
\alias{symbol_subscript_10}
\alias{symbol_subscript_x}
\alias{symbol_subscript_2.5}
\alias{symbol_subscript_3}
\alias{symbol_superscript_3}
\alias{symbol_superscript_minus}
\title{Functions for commonly used unicode symbols.}
\usage{
symbol_degree()

symbol_mu_lower()

symbol_subscript_1()

symbol_subscript_2()

symbol_subscript_4()

symbol_subscript_10()

symbol_subscript_x()

symbol_subscript_2.5()

symbol_subscript_3()

symbol_superscript_3()

symbol_superscript_minus()
}
\value{
Character vector.
}
\description{
Functions for commonly used unicode symbols.
}
\author{
Stuart K. Grange
}
