% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_system_date.R
\name{sys_unix_time}
\alias{sys_unix_time}
\title{Function to get system date as an numeric/integer value representing unix 
time.}
\usage{
sys_unix_time(integer = FALSE)
}
\arguments{
\item{integer}{Should the date be represented as an integer rather than a 
numeric value. Default is \code{FALSE}.}
}
\value{
Numeric or integer vector with a length of one.
}
\description{
Function to get system date as an numeric/integer value representing unix 
time.
}
\author{
Stuart K. Grange
}
