% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_pad.R
\name{time_pad}
\alias{time_pad}
\title{Function to pad time-series.}
\usage{
time_pad(df, pad.time = "10 min", start.unit = NULL, id.var = NULL,
  dplyr.join = TRUE, remove.final = FALSE)
}
\arguments{
\item{df}{A data frame including parsed dates. The date variable/column must
be named 'date'.}

\item{pad.time}{Frequency of time padding. Some examples could be: 'min'
'hour', 'day', 'month', 'year' but multiples such as '5 min' work too.}

\item{start.unit}{A optional date unit where the padded time-series should
begin and end. Examples are 'hour', 'day', 'month', and 'year'.}

\item{id.var}{Should identifying variables be applied to the data
post-pad?}

\item{dplyr.join}{Should dplyr::left_join be used rather than base:merge for
the date sequence join? dplyr::left_join performs better.}

\item{remove.final}{Should the final observation of the padded time-series be
removed? Sometimes if makes sense to remove the last observation if the
end date has been rounded forwards.}
}
\description{
time_pad use is analogous to openair::timeAverage. However, the aggregation
of values does not occur.
}
\details{
time_pad does not drop non-numerical variables, can include identifiers
post-pad, and can start and end a padded time-series at a 'nice place',
for example, at the beginning of an hour or day.

time_pad pads a time-series by calculating the maximum and minimum dates
within a time-series and then generating a uniform date sequence between the
maximum and minimum dates. This date sequence is then joined to the input
data frame and the missing values are represented as NA.

time_pad uses dplyr::left_join for the date join because testing has
demonstrated this function performs much faster than base::merge. However,
users can still opt to use base::merge because dplyr is reporting many bugs.
}
\examples{
# Pad time-series so every minute is present
\dontrun{
data.nelson.pad <- time_pad(data.nelson, pad.time = 'min', start.unit = 'day')
}
}
\author{
Stuart K. Grange
}
\seealso{
See \code{\link{round_date_interval}},
\code{\link{openair::timeAverage}}
}

