% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pad.R
\name{time_zone}
\alias{time_zone}
\title{Function to extract time-zone from a POSIXct date.}
\usage{
time_zone(x)
}
\description{
Function to extract time-zone from a POSIXct date.
}
\author{
Stuart K. Grange
}

