% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json.R
\name{to_json}
\alias{to_json}
\title{Function to quickly use \code{jsonlite::toJSON}.}
\usage{
to_json(x, pretty = TRUE)
}
\arguments{
\item{x}{An R object.}

\item{pretty}{Should the JSON object be pretty printed? Default is 
\code{TRUE}.}
}
\description{
Function to quickly use \code{jsonlite::toJSON}.
}
\author{
Stuart K. Grange
}

