% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dates_with_time_zone.R
\name{transform_dates_with_time_zone}
\alias{transform_dates_with_time_zone}
\title{Function to transform potentially heterogeneous time zones and force their
representation in the UTC time zone.}
\usage{
transform_dates_with_time_zone(
  df,
  dates_replace = TRUE,
  drop_time_zone = FALSE
)
}
\arguments{
\item{df}{Tibble with \code{date} and \code{time_zone} variables.}

\item{dates_replace}{Should the function replace/overwrite the original 
\code{date} variable(s) with the transformed version?}

\item{drop_time_zone}{Should the helper \code{time_zone} variable be dropped
in the return?}
}
\value{
\code{df} with transformed date variables.
}
\description{
\code{transform_dates_with_time_zone} is useful when time series are 
represented in a single time zone, but different groups are better analysed
in different time zones. \code{transform_dates_with_time_zone} will return 
transformed dates with the UTC time zone, but this may not strictly be true.
}
\author{
Stuart K. Grange
}
