% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_convert.R
\name{type_convert}
\alias{type_convert}
\alias{type_converter}
\title{Function to quickly apply \code{type.convert} to objects.}
\usage{
type_convert(x)

type_converter(x)
}
\arguments{
\item{x}{A vector or data frame.}
}
\value{
Vector or data frame, possibly with different data types.
}
\description{
Function to quickly apply \code{type.convert} to objects.
}
\author{
Stuart K. Grange
}
