% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unixTime.R
\name{unixTime}
\alias{unixTime}
\title{Function to calculate unix time from a date vector.}
\usage{
unixTime(date, tz.overwrite = NULL)
}
\description{
unixTime returns the numeric value from a date which represents seconds
since the unix epoch (1970-01-01 (UTC); excluding leap seconds). unixTime
can be very useful to store because it is time zone independent and is
efficent to parse.
}
\details{
Dates in R are often parsed and interacted without including correct time
zone information. However, if the time zone information for an ipnut date is
incorrect, the return of unixTime will also be incorrect. Rather than forcing
users to correctly parse all dates, unixTime has an argumment
\code{tz.overwrite} where the correct time zone for the date can be supplied
only for this function. \code{OlsonNames()} lists the time zones which can be
used.
}
\author{
Stuart K. Grange
}

