% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest}
\alias{unnest}
\title{Functions to transform a delimited string of nested observations into it's
own rows in a data frame.}
\usage{
unnest(df, variable, pattern, trim = FALSE, reset = TRUE, convert = FALSE)
}
\arguments{
\item{df}{Data frame which contains a nested delimited string variable.}

\item{variable}{Name of variable in \code{df} which is, or contains delimited 
strings.}

\item{pattern}{Delimiter to split the string. Common values are \code{","},
\code{"/"}, and \code{";"}.}

\item{trim}{Should the new vector be trimmed so there is no leading or 
trailing whitespace? Default is \code{FALSE}.}

\item{reset}{Should the \code{row.names} of the data frame be reset after 
replicating rows. Default is \code{TRUE}.}

\item{convert}{Should \code{type.convert} be applied to the unnested variable?}
}
\description{
\code{unnest} works by first splitting a string variable by a pattern and 
then the split object is then used to replicate all other variables in a data
 frame. \code{unnest} is useful for making data tidy.
}
\examples{
\dontrun{

# Make tidy data
data_tidy <- unnest(french_verbs, variable = "subject", pattern = ",")

}

}
\author{
Stuart K. Grange
}
\seealso{
\link{tidyr}
}

