% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{upload_to_ftp}
\alias{upload_to_ftp}
\title{Function to upload a file to an ftp or stpf server.}
\usage{
upload_to_ftp(file, url, credentials = "", progress = "text")
}
\description{
Function to upload a file to an ftp or stpf server.
}

