% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_with_scp.R
\name{upload_with_scp}
\alias{upload_with_scp}
\title{Function to upload files locally to a remote system with \code{scp} (secure 
copy).}
\usage{
upload_with_scp(host, file_local, file_remote, user, password,
  compression = FALSE, print = FALSE, progress = "none")
}
\arguments{
\item{host}{Host name of remote system.}

\item{file_local}{A vector of file names which are the destination files for
\code{file_remote}.}

\item{file_remote}{A vector of remote files names to download.}

\item{user}{User name for \code{scp}.}

\item{password}{Password for the user for \code{scp}.}

\item{compression}{Should the files be copied with compression on the fly? 
This can speed up copying time on slower networks but not always.}

\item{print}{Should the file to be uploaded be printed to the console as a
message?}

\item{progress}{Type of progress bar to display.}
}
\value{
Invisible.
}
\description{
\code{upload_with_scp} offers an alternative to the \code{curl} based 
functions which can be troublesome to use with the \code{sftp} protocol on 
Ubuntu systems. \code{upload_with_scp} needs the \code{sshpass} system 
programme to be installed.
}
\examples{
\dontrun{

# Get a file
upload_with_scp(
  host = "192.1.1.1",
  file_local = "~/Desktop/sock_data_copied.rds",
  file_remote = "/network_storage/r_files/sock_data.rds", 
  user = "username",
  password = "password_for_username"
)

}

}
\seealso{
\code{\link{list_files_scp}}, \href{https://gist.github.com/arunoda/7790979}{sshpass},
\code{\link{download_with_scp}}
}
\author{
Stuart K. Grange
}
