% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_utils_re_exports.R
\name{utils functions}
\alias{utils functions}
\title{Pseudo-function to re-export functions from the \strong{utils} package.}
\description{
Pseudo-function to re-export functions from the \strong{utils} package.
}
