% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\alias{uuid_rowwise}
\title{Function to generate randomly generated universally unique identifiers 
(UUIDs). 
 
A UUID is usually considered 32 lowercase hexadecimal digits displayed in 
five groups separated by hyphens (a total of 36 characters). The number of 
characters in the five groups are 8-4-4-4-12. There are about 3.4 x 10 ^ 38 
possible UUID combinations, therefore it is unlikely that two UUIDs will be 
identical.}
\usage{
uuid(uppercase = FALSE, sep = "-")

uuid_rowwise(df)
}
\arguments{
\item{uppercase}{Should the letters be returned as uppercase? Default
is \code{FALSE}.}

\item{sep}{What character should separate the five groups of characters? 
Default is hyphens (-).}

\item{df}{Input data frame for row-wise transformation.}
}
\description{
\code{base::sample} is used as the random number generator. Random-base 
UUIDs are sometimes called version 4 UUIDs.
}
\details{
The \code{uuid_rowwise} function uses \code{replicate} so a variable/column
in a data frame with unique uuids can be generated conveniently.
}
\examples{
\dontrun{
# Get a single identifier
identifier <- uuid()

# Print
identifier
"8e7675d5-4a87-4e0b-ae9b-9c97819aace5"

# Generate unique uuids for every row in a data frame
data_ozone$uuid <- uuid_rowwise(data_ozone)

}

}
\author{
Christopher Bare, modified by Stuart K. Grange
Originally from: \url{http://gist.github.com/cbare/5979354}
}
\seealso{
\code{\link{sample}}, \code{\link{replicate}}
}

