% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_monday.R
\name{week_financial}
\alias{week_financial}
\title{Function to calculate financial week of the year.}
\usage{
week_financial(date, start = "july", floor = TRUE)
}
\arguments{
\item{date}{Vector containing parsed dates.}

\item{start}{The starting month of the financial year. Currently, the only
option is \code{"july"}.}

\item{floor}{Should weeks which belong to the 53 rd week for the year be 
rounded down to the 52 nd week? Default is \code{TRUE}.}
}
\value{
Integer vector.
}
\description{
Function to calculate financial week of the year.
}
\examples{
\dontrun{
data_hours$week <- week_financial(data_hours$date)
}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{wday}}, \code{\link{week_monday}}
}

