% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_monday.R
\name{week_monday}
\alias{week_monday}
\title{Function to get Monday-based week number of a date.}
\usage{
week_monday(date, floor = FALSE)
}
\arguments{
\item{date}{Vector containing parsed dates.}

\item{floor}{Should weeks which belong to the 53 rd week for the year be 
rounded down to the 52 nd week? Default is \code{FALSE}.}
}
\value{
Integer vector.
}
\description{
Function to get Monday-based week number of a date.
}
\seealso{
\code{\link{wday}}, \code{\link{week_financial}}
}
\author{
Stuart K. Grange
}
