% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekday_sequence.R
\name{weekday_sequence}
\alias{weekday_sequence}
\title{Function to generate a sequence of weekdays.}
\usage{
weekday_sequence(n = 7, monday = TRUE, factor = TRUE)
}
\arguments{
\item{n}{Number of days for the sequence to return.}

\item{monday}{Should the sequence start on a Monday? Default is \code{TRUE} 
and set to \code{FALSE} if the sequence should start on Sunday.}

\item{factor}{Should the return be an ordered factor? Default is \code{TRUE}.}
}
\description{
Function to generate a sequence of weekdays.
}
\examples{
\dontrun{

# Generate weekday sequence as ordered factors
weekday_sequence()

# Ordered factors starting on Sunday
weekday_sequence(monday = FALSE)

# As characters
weekday_sequence(factor = FALSE)
 
}

}
\author{
Stuart K. Grange
}
