% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekdays.R
\name{weekdays}
\alias{weekdays}
\title{Function to generate a sequence of weekdays.}
\usage{
weekdays(n = 7, monday = TRUE, factor = TRUE)
}
\arguments{
\item{n}{Number of days for the sequence to return.}

\item{monday}{Should the sequence start on a Monday? Default is \code{TRUE} 
and set to \code{FALSE} if the sequence should start on Sunday.}

\item{factor}{Should the return be an ordered factor? Default is \code{TRUE}.}
}
\description{
Function to generate a sequence of weekdays.
}
\examples{
\dontrun{
# Get all weekdays
weekdays()

# Only weekdays
weekdays(5)
}

}
\author{
Stuart K. Grange
}

