% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{weekend}
\alias{weekend}
\title{Function to test if dates are during the weekend or not.}
\usage{
weekend(x)
}
\arguments{
\item{x}{Date vector.}
}
\value{
Logical vector.
}
\description{
Function to test if dates are during the weekend or not.
}
\author{
Stuart K. Grange
}
