% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_closest.R
\name{which_closest}
\alias{which_closest}
\title{Function to find closest index in a vector.}
\usage{
which_closest(vector, value, na.rm = FALSE)
}
\arguments{
\item{vector}{Vector to find closest value within.}

\item{value}{Value to find closest indices for. Must be a vector with length 
of 1.}

\item{na.rm}{Should \code{NA}s be ommited?}
}
\value{
Integer vector.
}
\description{
Function to find closest index in a vector.
}
\examples{

# A vector with no 17
x <- c(
  15, 15.5, 16, 16, 16, 18, 18, 18, 18, 18, 18.5, 19, 19.5, 20, 20, 20, 20.5, 
  21, 21, 21, 21, 21
)

# Using base which returns an empty integer
which(x == 17)

# All values which are closest
which_closest(x, 17)

# Just a single value which is closest
which_closest(x, 17)[1]

# Get element, use as a filter
x[which_closest(x, 17)[1]]

}
\seealso{
\code{\link{which}}, \code{\link{which.min}}
}
\author{
Stuart K. Grange
}
