% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_direction_from_wind_components.R
\name{wind_direction_from_wind_components}
\alias{wind_direction_from_wind_components}
\alias{wind_speed_from_wind_components}
\title{Functions to calculate wind direction and wind speed from horizontal wind 
components.}
\usage{
wind_direction_from_wind_components(u, v)

wind_speed_from_wind_components(u, v)
}
\arguments{
\item{u}{\code{u} wind component. The wind component which represents flow 
from west to east.}

\item{v}{\code{v} wind component. The wind component which represents flow 
from south to north.}
}
\value{
Numeric vector.
}
\description{
Functions to calculate wind direction and wind speed from horizontal wind 
components.
}
\examples{

# Calculate wind direction in decimal degrees
wind_direction_from_wind_components(3.12, 0.29)

# Calculate wind speed in m s-1
wind_speed_from_wind_components(3.12, 0.29)

}
\author{
Stuart K. Grange
}
