% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_range.R
\name{within_range}
\alias{within_range}
\title{Function to test if a vector is within one, or many ranges.}
\usage{
within_range(x, start, end)
}
\arguments{
\item{x}{Numeric vector to test.}

\item{start}{Start of range.}

\item{end}{End of range.}
}
\value{
Logical vector of length of \code{x}.
}
\description{
\code{within_range} works well for testing Unix time too.
}
\examples{

# Test a single range
within_range(1:10, 8, 9)

# Test many ranges
within_range(1:100, c(1, 5), c(2, 9))
 
}
\seealso{
\code{\link{which}}
}
\author{
Stuart K. Grange
}
