% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.gpx.R
\name{write.gpx}
\alias{write.gpx}
\title{Function to write a data frame to a GPX file with usage analogous to
\code{write.table} and its derivatives.}
\usage{
write.gpx(df, latitude = "latitude", longitude = "longitude", file = "",
  name = NULL, layer = "")
}
\arguments{
\item{df}{Data frame to be written as a GPX file.}

\item{latitude}{df's latitude variable name.}

\item{longitude}{df's longitude variable name.}

\item{file}{File name of exported GPX file}

\item{name}{Name of variable which will be copied and added in the "name"
element of the GPX file.}

\item{layer}{}
}
\description{
\code{write.gpx} uses \code{rgdal::writeOGR} as the GPX writer. Unlike the
standard \code{rgdal} function, write.gpx will automatically expand file
paths and overwrite previous files.
}
\examples{
\dontrun{
write.gpx(data.bus.stations, "~/Desktop/bus_stations.gpx")
}
}
\author{
Stuart K. Grange
}

