% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.json.R
\name{write.json}
\alias{write.json}
\title{Function to write a data frame to a JSON file with usage analogous to
\code{write.table}.}
\usage{
write.json(df, file = "", pretty = TRUE, na.empty = TRUE)
}
\arguments{
\item{df}{Data frame to be written as a JSON file.}

\item{file}{JSON file name.}

\item{pretty}{Should the JSON output be formatted for readability? Default is
TRUE.}

\item{na.empty}{Should empty strings be converted to NA values to they are
not included in the the JSON element?}
}
\description{
\code{write.json} uses \code{jsonlite} as the data frame to JSON parser.
}
\examples{
\dontrun{
write.json(data.car.identifiers, "car_identifiers.json")
}
}
\author{
Stuart K. Grange
}

