% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_threadr.R
\name{write_csv_threadr}
\alias{write_csv_threadr}
\title{Function to export data frame/tibble to a \code{.csv} text file with certain 
defaults.}
\usage{
write_csv_threadr(df, file, format_dates = TRUE, na = "", excel_bom = FALSE)
}
\arguments{
\item{df}{A data frame or tibble to write to disk.}

\item{file}{File to write \code{df} to.}

\item{format_dates}{Should the dates in \code{df} be converted to a 
\code{\%Y-\%m-\%d \%H:\%M:\%S} string?}

\item{na}{String used for missing values.}

\item{excel_bom}{Should a byte-order mark (BOM) be added to the file to 
indicate to Microsoft Excel that the file is \code{UTF-8} encoded.}
}
\value{
Invisible \code{df}.
}
\description{
Function to export data frame/tibble to a \code{.csv} text file with certain 
defaults.
}
\author{
Stuart K. Grange
}
