% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_database_table.R
\name{write_database_table}
\alias{write_database_table}
\title{Function to write an entire database table to disk.}
\usage{
write_database_table(db, table, nrow = 1e+06, file = "")
}
\arguments{
\item{db}{A database connection.}

\item{table}{\code{db}'s table which is to be written to disk.}

\item{nrow}{Number of rows of the \code{db}'s table to query and then write
to disk in each cycle. Default is 1000000 rows.}

\item{file}{File to export the table to.}
}
\description{
\code{write_database_table} breaks up the queries and exports to small pieces
to allow for very large tables to be written to disc with systems with limited
amounts of physical memory. The output is a \code{.csv} file.
}
\author{
Stuart K. Grange
}

