% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json.R
\name{write_json}
\alias{write_json}
\title{Function to write an R object to a JSON file.}
\usage{
write_json(x, file, pretty = TRUE, na = FALSE, auto_unbox = FALSE)
}
\arguments{
\item{x}{R object to be written as a JSON file.}

\item{file}{JSON file name.}

\item{pretty}{Should the JSON output be formatted for readability? Default is
\code{TRUE}.}

\item{na}{Should \code{NA} values be kept in the JSON file? Default is 
\code{FALSE}. If \code{na} is \code{TRUE} then \code{NA}s will transformed to
JSON's \code{null} value.}

\item{auto_unbox}{Should vectors with the length of \code{1} be represented 
as atomic, single element arrays? Default is \code{FALSE}.}
}
\value{
Invisible \code{x}.
}
\description{
\code{write_json} uses \strong{jsonlite} as the data frame to JSON parser.
}
\examples{
\dontrun{

# Export a data frame
write_json(data_car_identifiers, "car_identifiers.json")

}

}
\author{
Stuart K. Grange.
}
