% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json_lines.R
\name{write_json_lines}
\alias{write_json_lines}
\title{Function to export a data frame as new line JSON, also known as \code{ndjson}.}
\usage{
write_json_lines(df, file, append = FALSE, pagesize = 500,
  verbose = FALSE)
}
\arguments{
\item{df}{Data frame to export.}

\item{file}{File name for exported new line JSON file.}

\item{append}{Should the data be appended to \code{file}?}

\item{pagesize}{Number of lines to write to \code{file} per iteration.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible.
}
\description{
Function to export a data frame as new line JSON, also known as \code{ndjson}.
}
\author{
Stuart K. Grange
}
