% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rds.R
\name{write_rds}
\alias{write_rds}
\alias{read_rds}
\title{Function to read and write R's RDS files with some helpful messaging.}
\usage{
write_rds(x, file, compress = TRUE, file_only_print = FALSE, verbose = TRUE)

read_rds(file, file_only_print = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{An R object to write to disc.}

\item{file}{File name of RDS file.}

\item{compress}{Should the RDS file be written with compression?}

\item{file_only_print}{If \code{verbose} is \code{TRUE}, should only the file
name be printed without its path?}

\item{verbose}{Should the function give messages? Note that the default is 
\code{TRUE}.}
}
\value{
Invisible \code{x}.
}
\description{
Function to read and write R's RDS files with some helpful messaging.
}
\author{
Stuart K. Grange
}
