% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml.R
\name{write_xml}
\alias{write_xml}
\title{Function to write a data frame to an XML document.}
\usage{
write_xml(df, file, name = "observation")
}
\description{
\code{write_xml} is not suitable for large data frames.
}

