% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeJavaScriptPointData.R
\name{writeJavaScriptPointData}
\alias{writeJavaScriptPointData}
\title{Function to generate \code{geojson} JavaScript variables from a data frame
containing points.}
\usage{
writeJavaScriptPointData(x = data.done, name = "points_done",
  output = "uk_location_tracking/", latitude = "latitude",
  longitude = "longitude")
}
\description{
Function to generate \code{geojson} JavaScript variables from a data frame
containing points.
}
\author{
Stuart Grange
}

