% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{yesterday}
\alias{yesterday}
\title{Function to return the system's idea of yesterday.}
\usage{
yesterday(as_POSIXct = FALSE, tz = "UTC")
}
\arguments{
\item{as_POSIXct}{Should the return be of \code{POSIXct} data type?}

\item{tz}{If \code{as_POSIXct} is \code{TRUE}, what time zone should the 
return be in?}
}
\value{
\code{Date} or \code{POSIXct} vector with a length of 1.
}
\description{
Function to return the system's idea of yesterday.
}
\examples{

# Date data type
yesterday()

# POSIXct data type
yesterday(as_POSIXct = TRUE)

}
\author{
Stuart K. Grange
}
