% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_zoo.R
\name{zoo_to_data_frame}
\alias{zoo_to_data_frame}
\title{Function to transform a \strong{zoo} time series object to a tibble.}
\usage{
zoo_to_data_frame(zoo_df, clean_names = FALSE)
}
\arguments{
\item{zoo_df}{A \strong{zoo} time series object.}

\item{clean_names}{Should the names be cleaned?}
}
\value{
Tibble.
}
\description{
Function to transform a \strong{zoo} time series object to a tibble.
}
\author{
Stuart K. Grange.
}
