% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_humidity.R
\name{absolute_humidity}
\alias{absolute_humidity}
\title{Function to calculate absolute humidity from temperature and relative 
humidity.}
\usage{
absolute_humidity(air_temp, rh)
}
\arguments{
\item{air_temp}{Air temperature in degrees Celsius.}

\item{rh}{Relative humidity in percentage.}
}
\value{
Numeric vector, absolute humidity in \code{g.m-3}.
}
\description{
Absolute humidity is the amount of amount of water vapour in air, usually
expressed in \code{g.m-3}.
}
\examples{

# Calculate absolute humidity with temperature in degrees C and relative
# humidity in percent
absolute_humidity(10.7, 73.3)

}
\author{
Stuart K. Grange
}
