% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_numbers.R
\name{add_row_numbers}
\alias{add_row_numbers}
\title{Function to add a row number variable to a data frame.}
\usage{
add_row_numbers(df, name = "row_number", zero_based = FALSE)
}
\arguments{
\item{df}{Data frame to be transformed.}

\item{name}{Name of variable to use. \code{add_row_numbers} will silently 
drop an existing variable of the same name if present in \code{df}.}

\item{zero_based}{Should the numbering start at \code{0} rather than \code{1}? 
Default is \code{FALSE}.}
}
\value{
Data frame.
}
\description{
Row numbers will be a variable which is an integer, unlike 
\code{rownames_to_column} which adds a character variable.
}
\examples{
\dontrun{

# Add row numbers
data_zones <- add_row_numbers(data_zones)

# Or change the name
data_zones <- add_row_numbers(data_zones, name = "order")

# Use zero-based numbering
data_zones <- add_row_numbers(data_zones, name = "id", zero_based = TRUE)

}

}
\seealso{
\code{\link{add_rownames}}
}
\author{
Stuart K. Grange
}
