% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_date.R
\name{aggregate_by_date}
\alias{aggregate_by_date}
\title{Function to aggregate time series data by dates.}
\usage{
aggregate_by_date(
  df,
  interval = "hour",
  by = NA,
  summary = "mean",
  threshold = 0,
  round = NA,
  pad = TRUE,
  determine_interval = TRUE,
  warn = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{Input data frame to be aggregated. \code{df} must contain 
\code{"date"} and \code{"value"} variables. The \code{"date"} variable 
must be a \code{POSIXct} date class and the value must be a \code{numeric} or
\code{integer} data type.}

\item{interval}{What interval should the aggregation be? Default is 
\code{"hour"}.}

\item{by}{What variables should \code{df} be grouped by? Common groups are 
\code{"site"} and \code{"variable"}.}

\item{summary}{What summary function should be applied for the aggregation? 
Default is the \code{mean}. The options are: 

\itemize{
  \item{mean}
  \item{median}
  \item{max}
  \item{min}
  \item{sum}
  \item{count}
  \item{sd}
  \item{mode}
  \item{data_capture}
}}

\item{threshold}{What data capture threshold is needed to create a valid 
aggregation. This is an value between \code{0} and \code{1}. Zero would mean 
any number of values will be valid but \code{0.75} would mean \code{75 \%} of
values are needed for a valid average.}

\item{round}{Should the aggregations be rounded? Default is no but \code{3} 
would round to three decimal places.}

\item{pad}{Should the time series be padded before aggregation? Almost always
this will be \code{TRUE} but if you want to speed the function up and have done
this previously, it can be set to \code{FALSE}.}

\item{determine_interval}{Should the input time series be evaluate to find
it's averaging period/interval. This is required for the correct calculation 
of \code{threshold}.}

\item{warn}{Should the function return warnings in certain situations?}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
\code{aggregate_by_date} does a similar job and has the same objectives of
\strong{openair}'s \code{timeAverage}. However, it has been developed 
to perform on "longer" data which is often encountered.
}
\examples{
\dontrun{

# Aggregate to hourly means
aggregate_by_date(data_air, "hour", by = c("site", "variable"))

# Multiple intervals are also supported, such as multiple days 
aggregate_by_date(data_air, "14 days", by = c("site", "variable"))

}

}
\seealso{
\code{\link[openair]{timeAverage}}, \code{\link{time_pad}}
}
\author{
Stuart K. Grange
}
