% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apparent_temperature.R
\name{apparent_temperature}
\alias{apparent_temperature}
\title{Function to calculate apparent temperature.}
\usage{
apparent_temperature(temp, rh, ws)
}
\arguments{
\item{temp}{Temperature (usually air temperature) in degrees Celsius.}

\item{rh}{Relative humidity in percentage.}

\item{ws}{Wind speed in \code{m.s-1}.}
}
\description{
Apparent temperature was defined by Steadman (1994). There are two variants,
one using radiation and one without -- this function uses the non-radiation
variant.
}
\examples{

# Calculate apparent temperature, some real observations from the Bernese 
# Jura in summer 2021 when it felt warm
apparent_temperature(25.7, 64.5, 0.9)

}
\seealso{
\code{\link{heat_index}}.
}
\author{
Stuart K. Grange
}
