% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_left.R
\name{arrange_left}
\alias{arrange_left}
\title{Function to arrange variables to the left of a data frame.}
\usage{
arrange_left(df, variable)
}
\arguments{
\item{df}{Data frame which is to be transformed.}

\item{variable}{Variable(s) to be moved to the left side of \code{df}.}
}
\description{
\code{arrange_left} is useful when some variables/columns want to be moved to
the left-side of a data frame.
}
\author{
Stuart K. Grange
}

