% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_date_summaries.R
\name{calculate_date_summaries}
\alias{calculate_date_summaries}
\title{Function to calculate summaries of dates for a continuous time series.}
\usage{
calculate_date_summaries(
  df,
  ...,
  interval = "hour",
  drop_n = FALSE,
  drop_date_end = FALSE,
  n_to_count = FALSE,
  use_data_table = TRUE,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{df}{Input tibble. \code{df} must contain \code{date} and \code{value} 
variables. The \code{date} variable must be a \code{POSIXct} date class and
the value must be a \code{numeric} data type.}

\item{...}{Grouping variables/specification for the summaries. If location
and variable variables are contained in \code{df}, these will usually be 
used as groups. If a column/variable is used, wind direction (\code{wd}) will
be processed with vector functions.}

\item{interval}{What interval should the aggregation be? English names and 
multiples of time intervals are accepted. For example, a valid \code{interval}
is \code{"5 mins"}.}

\item{drop_n}{Should the count of non-missing elements (\code{n}) be dropped
from the return?}

\item{drop_date_end}{Should the \code{date_end} variable be dropped from the
return?}

\item{n_to_count}{Should the \code{n} variable be renamed to \code{count}?}

\item{use_data_table}{Should the \code{data.table} backend be used for the
aggregation calculations?}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Tibble.
}
\description{
\code{calculate_date_summaries} will pad the time series before aggregation
to ensure all date steps are present in the return.
}
\seealso{
\code{\link{aggregate_by_date}}, \code{\link{time_pad}}, 
\code{\link{calculate_date_summaries_wide}}
}
\author{
Stuart K. Grange
}
