% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pace.R
\name{calculate_pace}
\alias{calculate_pace}
\title{Function for calculating pace in min km-1.}
\usage{
calculate_pace(time, distance)
}
\arguments{
\item{time}{Numeric vector of time in seconds.}

\item{distance}{Numeric vector of kilometers covered within the time period 
in \code{time}.}
}
\value{
Numeric vector of pace in min km-1.
}
\description{
Function for calculating pace in min km-1.
}
\examples{

\dontrun{

# Trasform time and distance to pace in a data frame
df$pace <- calculate_pace(time, distance)

}

}
\author{
Stuart K. Grange
}
\keyword{cycling}
\keyword{pace,}
\keyword{running,}
\keyword{sport,}

