% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantiles.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Function to calculate quantiles for a numeric vector.}
\usage{
calculate_quantiles(
  x,
  probabilities = c(0.025, 0.975),
  type = 7,
  keep_probabilities = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{x}{A numeric vector.}

\item{probabilities}{A numeric vector with the length of two representing
probabilities of the quantiles. These values will always be less than 1.}

\item{type}{Method to use to calculate the quantiles.}

\item{keep_probabilities}{Should the probabilities vector be kept? If 
\code{TRUE}, the return will contain a nested column.}

\item{na.rm}{Should missing values be removed for calculation? The default is
\code{TRUE} because missing values cannot be used for quantile calculations.}
}
\value{
Tibble.
}
\description{
Function to calculate quantiles for a numeric vector.
}
\seealso{
\code{\link{quantile}}
}
\author{
Stuart K. Grange
}
