% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pace.R
\name{calculate_speed}
\alias{calculate_speed}
\alias{calculate_pace}
\title{Functions for calculating speed in m s-1 and pace in min km-1.}
\usage{
calculate_speed(seconds, metres)

calculate_pace(seconds, metres)
}
\arguments{
\item{seconds}{Numeric vector of time in seconds.}

\item{metres}{Numeric vector of metres covered within the time period 
in \code{seconds}.}
}
\value{
Numeric vector of speed in m s-1 or pace in min km-1.
}
\description{
Functions for calculating speed in m s-1 and pace in min km-1.
}
\examples{

\dontrun{

# Calculate some speeds
calculate_speed(9.58, 100)
ms_to_km_h(calculate_speed(9.58, 100))

# and pace
calculate_pace(19.19, 200)
min_km_to_km_h(calculate_pace(19.19, 200))

decimal_minute_to_string(calculate_pace(19.19, 200), hour = FALSE)

calculate_speed(3503, half_marathon() * 1000)
decimal_minute_to_string(calculate_pace(3503, half_marathon() * 1000))

}

}
\author{
Stuart K. Grange
}
\keyword{cycling}
\keyword{pace,}
\keyword{running,}
\keyword{speed,}
\keyword{sport,}
