% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dew_point.R
\name{calculate_vapour_pressure}
\alias{calculate_vapour_pressure}
\title{Function to calculate atmospheric vapour pressure.}
\usage{
calculate_vapour_pressure(temp)
}
\arguments{
\item{temp}{Air temperature or dew point temperature in degrees Celsius.}
}
\value{
Numeric vector, vapour pressure in hectopascals (hPa).
}
\description{
Function to calculate atmospheric vapour pressure.
}
\examples{

# Calculate saturated vapour pressure with air temperature
calculate_vapour_pressure(20)

# Calculate actual vapour pressure with dew point temperature
calculate_vapour_pressure(calculate_dew_point(temp = 20, rh = 65))

}
\author{
Stuart K. Grange
}
