% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vapour_pressure.R
\name{calculate_vapour_pressure}
\alias{calculate_vapour_pressure}
\title{Function to calculate atmospheric vapour pressure.}
\usage{
calculate_vapour_pressure(temp, type = 1L)
}
\arguments{
\item{temp}{Air temperature or dew point temperature in degrees Celsius.}

\item{type}{What type of vapour pressure should be calculated? Type \code{1} 
is the package's original version, type \code{2} is Bolton 1980's version
often used with dew point temperatures, and type \code{3} is Dosseger et 
al. 1992's version that is used for the calculation of apparent temperature}
}
\value{
Numeric vector, vapour pressure in hectopascals (hPa).
}
\description{
Water vapour pressure is the atmospheric pressure which is exerted by water 
vapour (gaseous water).
}
\examples{

# Calculate saturated vapour pressure with air temperature
calculate_vapour_pressure(20)

# Use Bolton 1980's formula
calculate_vapour_pressure(20, type = 2)

# Calculate actual vapour pressure with dew point temperature
calculate_vapour_pressure(calculate_dew_point(temp = 20, rh = 65))
calculate_vapour_pressure(calculate_dew_point(temp = 20, rh = 65), type = 2)

}
\author{
Stuart K. Grange
}
