% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R
\name{clc}
\alias{clc}
\title{Function to clear the R console.}
\usage{
clc()
}
\description{
Function to clear the R console.
}
\examples{
\dontrun{
# Clear console
clc()

# Or without loading package
threadr::clc()

}

}
\seealso{
\url{http://stackoverflow.com/questions/14260340/function-to-clear-the-console-in-r}{stackoverflow.com}
}
\author{
Stuart K. Grange
}
