% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours_lancet.R
\name{colours_lancet}
\alias{colours_lancet}
\alias{colours_icct}
\title{Function to generate a custom colour palettes from Lancet figures.}
\usage{
colours_lancet(palette = c(1, 2))

colours_icct()
}
\arguments{
\item{palette}{Palette number.}
}
\value{
Character vector.
}
\description{
Function to generate a custom colour palettes from Lancet figures.
}
\examples{

colours_lancet()
colours_lancet(palette = 2)

}
\author{
Stuart K. Grange
}
