% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_files.R
\name{compress_files}
\alias{compress_files}
\title{Function to compress files.}
\usage{
compress_files(files, algorithm = "bzip2", keep = TRUE)
}
\arguments{
\item{files}{Files to compress.}

\item{algorithm}{Compression algorithm to use.}

\item{keep}{Should the original file be kept? Default is \code{TRUE}.}
}
\description{
Function to compress files.
}
\author{
Stuart K. Grange
}
