% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_directory.R
\name{create_directory}
\alias{create_directory}
\title{Function to create directories if they are not present.}
\usage{
create_directory(directory)
}
\arguments{
\item{directory}{Name or path of directory to be created if id does not exist.
\code{directory} can take many values.}
}
\description{
If the directory exists, no function will be called.
}
\examples{
\dontrun{

# Create a directory
create_directory("data_dump")

# Create many directories
create_directory(c("data_dump", "/testing/new_directory"))

}
}
\author{
Stuart K. Grange
}

