% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_frame_to_line.R
\name{data_frame_to_line}
\alias{data_frame_to_line}
\title{Function for converting a data frame into a spatial-lines data frame.}
\usage{
data_frame_to_line(df, latitude = "latitude", longitude = "longitude")
}
\arguments{
\item{df}{Data frame to be converted into SpatialLinesDataFrame.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}
}
\description{
\code{data_frame_to_line} returns a SpatialLinesDataFrame with a single line
object. Multiple identifiers are not currently preserved. In general usage,
\code{latitude} and \code{longitude} will be projected in WGS84.
\code{latitude} and \code{longitude} must not contain \code{NA} values
because this is a limitation of the spatial object.
}
\examples{
\dontrun{
sp.lines <- data_frame_to_line(data.gps.track, "latitude", "longitude")
}
}
\author{
Stuart K. Grange
}

