% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_zoo.R
\name{data_frame_to_zoo}
\alias{data_frame_to_zoo}
\title{Function to transform a data frame into a \strong{zoo} time series object.}
\usage{
data_frame_to_zoo(df)
}
\arguments{
\item{df}{Data frame or tibble containing a parsed date variable named 
\code{"date"}.}
}
\value{
\strong{zoo} series.
}
\description{
Function to transform a data frame into a \strong{zoo} time series object.
}
\author{
Stuart K. Grange.
}
