% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_date_formatted.R
\name{date_message}
\alias{date_message}
\alias{message_date_prefix}
\title{Function to get a formatted date string used to prefix messages.}
\usage{
date_message()

message_date_prefix()
}
\value{
Character vector.
}
\description{
Function to get a formatted date string used to prefix messages.
}
\examples{

# The message prefix
date_message()

# Within a message
message(date_message(), "Loading...")

}
\author{
Stuart K. Grange
}
