% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_sequence.R
\name{date_sequence}
\alias{date_sequence}
\title{Function to generate a sequence of past or future dates.}
\usage{
date_sequence(days = -1, character = TRUE, today = TRUE, sep = "-")
}
\arguments{
\item{days}{Number of days of the sequence from system's date. Use negative
integers to go back in time.}

\item{character}{Should the date sequence be returned as a character vector 
rather than a date vector? Default is \code{TRUE}.}

\item{today}{Should system date be included in the returned sequence?}

\item{sep}{What separator should be used between the year, month, and day
pieces of the date? If this is changed from \code{"-"} (hyphen), then the 
returned vector will always be a character.}
}
\description{
\code{date_sequence} is useful for generating patterns which match file names.
}
\examples{
\dontrun{

# Make a date sequence of the last three days and today
pattern <- date_sequence(days = -3)

}

}
\author{
Stuart K. Grange
}
