% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Function to create a database connection with a JSON configuration file.}
\usage{
db_connect(file, database, config = TRUE, foreign_keys = FALSE)
}
\arguments{
\item{file}{\code{JSON} file or string containing database connection 
details. For SQLite databases, use the database's file path.}

\item{database}{The name of the database within \code{file} to use to create
a database connection to. If only one entry is in \code{file}, this argument 
is not needed and will be ignored if used.}

\item{config}{A logical to skip the \code{JSON} file configuration and just
attempt to connect to \code{file} directly. This is used for \code{SQLite}
databases which require no configuration.}

\item{foreign_keys}{A logical for SQLite databases where foreign keys should 
be enforced. Default is \code{FALSE}. For other database types, this will be 
ignored.}
}
\description{
\code{db_connect} uses a \code{JSON} configuration file to create a database
connection. This configuration file will often exist outside a code package 
so database credentials are not accidentally transmitted or shared.
}
\details{
If only one entry is in the \code{JSON} file, the \code{database} argument is
not needed.

MySQL, PostgreSQL, and SQLite connections are currently supported.
}
\examples{
\dontrun{

# Connect to an air quality database
db <- db_connect("connections.json", "air_quality")


# Use a json config file which looks similar to this:
string <- '
{
  "driver": "MySQL",
  "host": "172.31.4.159",
  "database_name": "database_seven",
  "user": "web",
  "password": "read_password"
}'

# Connect, no need for second argument when one connection is present in 
# configuration file
db_seven <- db_connect(string)


# A SQLite connection needs no configuration
con <- db_connect("../databases/air_quality_data.sqlite", config = FALSE)

}

}
\author{
Stuart K. Grange
}

