% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_count_rows.R
\name{db_count_rows}
\alias{db_count_rows}
\title{Function to get row counts from database tables.}
\usage{
db_count_rows(con, table = NA, progress = "none")
}
\arguments{
\item{con}{Database connection}

\item{table}{Table name in \code{con}. If \code{table} is unknown, all tables
in \code{con} will be queried.}

\item{progress}{Type of progress bar to display. Default is \code{"none"}.}
}
\description{
The \code{row_count} variable will be a string rather than an integer to 
avoid integer overflow issues.
}
\author{
Stuart K. Grange
}

