% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_count_rows.R
\name{db_count_rows}
\alias{db_count_rows}
\title{Function to get row counts from database tables.}
\usage{
db_count_rows(db, table = NA, progress = FALSE)
}
\arguments{
\item{db}{Database connection}

\item{table}{Table name in \code{db}. If \code{table} is unknown, all tables
in \code{db} will be queried.}

\item{progress}{Progress bar. Should a progress bar be shown? Default is 
\code{FALSE}.}
}
\description{
Function to get row counts from database tables.
}
\author{
Stuart K. Grange
}

