% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_variables.R
\name{db_head}
\alias{db_head}
\title{Function to get first \emph{n} rows of a database table.}
\usage{
db_head(con, table, n = 5)
}
\description{
Function to get first \emph{n} rows of a database table.
}
\author{
Stuart K. Grange
}

