% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert.R
\name{db_insert}
\alias{db_insert}
\title{Function to insert a data frame into/as a database table.}
\usage{
db_insert(db, table, df, append = TRUE, overwrite = FALSE, rows = FALSE,
  fill = FALSE, increment_reset = FALSE)
}
\arguments{
\item{db}{Database connection.}

\item{table}{Table in, or to be created in \code{db}.}

\item{df}{Data frame to be inserted into \code{db}.}

\item{append}{Should the table be appended? Default is \code{TRUE}.}

\item{overwrite}{Should the table be overwritten? Default is \code{FALSE}.}

\item{rows}{Should the inserted data include row names? Default is 
\code{FALSE}.}

\item{fill}{Should \code{df} be forced to have the same columns and order as
\code{table}?}

\item{increment_reset}{Should the auto-increment column in \code{table} be
reset before insert? Default is \code{FALSE}.}
}
\description{
\code{db_insert} is a wrapper for \code{DBI::dbWriteTable} but uses different 
defaults and has a few enhancements which can be helpful.
}
\details{
\code{db_insert} will not overwrite data or include a "row.names" variable by
default.
}
\author{
Stuart K. Grange
}

