% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_insert.R
\name{db_insert}
\alias{db_insert}
\title{Function to insert a data frame into/as a database table.}
\usage{
db_insert(db, table, df, append = TRUE, overwrite = FALSE, rows = FALSE,
  fill = FALSE, increment_reset = FALSE)
}
\arguments{
\item{db}{A database connection.}

\item{table}{A table in or to be created in \code{db}}

\item{df}{Data frame to be inserted into \code{db}}

\item{append}{Should the table be appended? Default is TRUE.}

\item{overwrite}{Should the table be overwrited? Default is FALSE.}

\item{rows}{Should the inserted data include row names? Default is FALSE.}

\item{fill}{Should \code{df} be forced to have the same columns and order as
\code{table}?}

\item{increment_reset}{Should the auto-increment column in \code{table} be
reset before insert? Default is FALSE.}
}
\description{
\code{db_insert} is a wrapper for \code{dbWriteTable} but uses different
defaults and has a few enhancements which can be helpful.
}
\author{
Stuart K. Grange
}

