% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert.R
\name{db_send}
\alias{db_send}
\title{Function to send a statement to a database.}
\usage{
db_send(con, statement)
}
\arguments{
\item{con}{Database connection.}

\item{statement}{Statement to send to \code{con}.}
}
\description{
\code{db_send} is a wrapper for \code{DBI::dbSendQuery}.
}

