% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_variables.R
\name{db_variables}
\alias{db_variables}
\title{Function to return all variable names in all tables in a SQL database.}
\usage{
db_variables(db)
}
\arguments{
\item{db}{Database connection.}
}
\description{
Function to return all variable names in all tables in a SQL database.
}
\examples{
\dontrun{
table_variables <- db_variables(db)

}
}
\author{
Stuart K. Grange
}

