% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_date_interval.R
\name{detect_date_interval}
\alias{detect_date_interval}
\title{Function to determine averaging period in a date vector.}
\usage{
detect_date_interval(date, skip = 1, n = 100)
}
\arguments{
\item{date}{A date vector.}

\item{skip}{Number of elements in \code{date} to skip before detecting 
interval.}

\item{n}{Number of elements in \code{date} to evaluate to detect the interval.}
}
\description{
Function to determine averaging period in a date vector.
}
\author{
Stuart K. Grange
}

