% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{download_ftp_file}
\alias{download_ftp_file}
\alias{list_files_ftp}
\title{Functions to download files from an FTP or SFTP server.}
\usage{
download_ftp_file(url, credentials, directory = NA)

list_files_ftp(url, credentials)
}
\arguments{
\item{credentials}{Credentials for a FTP or SFTP server.}

\item{directory}{Directory where the files are to be downloaded to.}

\item{url.}{The url of the file which is to be downloaded.}
}
\description{
\code{download_ftp_file} is not vectorised yet.
}
\examples{

\dontrun{

# Set credentials in this format
credentials <- "username:password"

# Get the file "hour_5" in the directory "test" from a FTP server
download_ftp_file("ftp://195.174.23.76/test/hour_5.csv", 
                  credentials, "ftp_data/")
                  

# List files on a ftp server
list_files_ftp("ftp://195.174.23.76/test", credentials)
"ftp://195.174.23.76/test/hour_5.csv"

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{RCurl}}
}

