% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{download_ftp_file}
\alias{download_ftp_file}
\alias{download_ftp_files}
\title{Functions to download files from an FTP or SFTP server.}
\usage{
download_ftp_file(file, credentials = "", directory = NA, curl = TRUE,
  progress = "text")
}
\arguments{
\item{credentials}{Credentials for a FTP or SFTP server. Do not use 
\code{credentials} if the server does not require authentication. 
\code{credentials} take the format: \code{"username:password"}.}

\item{directory}{Directory where the files are to be downloaded to.}

\item{curl}{Should \strong{RCurl} be used to download the files or base R's 
\code{download.file}? Default is \code{TRUE}.}

\item{url}{The url(s) of the file(s) which are to be downloaded.}
}
\description{
Functions to download files from an FTP or SFTP server.
}
\examples{
\dontrun{
# Set credentials in this format
credentials <- "username:password"

# Get the file "hour_5" in the directory "test" from a FTP server
download_ftp_file("ftp://195.174.23.76/test/hour_5.csv", 
                  credentials, "ftp_data/")
                  

# List files on a ftp server
list_files_ftp("ftp://195.174.23.76/test", credentials)
"ftp://195.174.23.76/test/hour_5.csv"

}

}
\author{
Stuart K. Grange
}
\seealso{
\strong{RCurl}
}

