% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{download_ftp_file}
\alias{download_ftp_file}
\title{Function to download files from an FTP or SFTP server.}
\usage{
download_ftp_file(url, file_local = NA, credentials = "", curl = FALSE,
  verbose = FALSE, progress = "none")
}
\arguments{
\item{url}{The url(s) of the file(s) which are to be downloaded.}

\item{credentials}{Credentials for a FTP or SFTP server. Do not use 
\code{credentials} if the server does not require authentication. 
\code{credentials} takes the format: \code{"username:password"}.}

\item{curl}{Should \strong{RCurl} be used to download the files or base R's 
\code{download.file}?}

\item{verbose}{Should the function give messages about download progress? 
Only works when \code{curl = FALSE}.}

\item{progress}{Progress bar type.}

\item{file_output}{Name of downloaded files. \code{file_output} should be the
same length as \code{url}. If not used, the \code{url} basename will be used. 
The directory will be created if it does not exist.}
}
\description{
Function to download files from an FTP or SFTP server.
}
\examples{
\dontrun{

# Download a file from a server which does not need credentials
url <- "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv"
download_ftp_file(url, "~/Desktop/noaa_data.csv")

}

}
\seealso{
\strong{\link{RCurl}}, \code{\link{list_files_ftp}}, 
\code{\link{upload_ftp_file}}
}
\author{
Stuart K. Grange
}
