% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_with_scp.R
\name{download_with_scp}
\alias{download_with_scp}
\title{Function to download files from a remote system with \code{scp} (secure copy).}
\usage{
download_with_scp(
  host,
  file_remote,
  file_local,
  user,
  password,
  compression = FALSE,
  verbose = FALSE,
  basename = FALSE,
  quiet_streams = FALSE
)
}
\arguments{
\item{host}{Host name of remote system.}

\item{file_remote}{A vector of remote files names to download.}

\item{file_local}{A vector of file names which are the destination files for
\code{file_remote}.}

\item{user}{User name for \code{scp}.}

\item{password}{Password for the user for \code{scp}.}

\item{compression}{Should the files be copied with compression on the fly? 
This can speed up copying time on slower networks but not always.}

\item{verbose}{Should the file to be downloaded be printed to the console as 
a message?}

\item{basename}{If \code{verbose}, should only the basename of file being 
handled be printed?}

\item{quiet_streams}{Should the system streams be suppressed?}
}
\value{
Invisible.
}
\description{
\code{download_with_scp} offers an alternative to the \code{curl} based 
functions which can be troublesome to use with the \code{sftp} protocol on 
Ubuntu systems. \code{download_with_scp} needs the \code{sshpass} system 
programme to be installed.
}
\examples{
\dontrun{

# Get a file
download_with_scp(
  host = "192.1.1.1",
  file_remote = "/network_storage/r_files/sock_data.rds", 
  file_local = "~/Desktop/sock_data_copied.rds",
  user = "username",
  password = "password_for_username"
)

}

}
\seealso{
\code{\link{list_files_scp}}, \href{https://gist.github.com/arunoda/7790979}{sshpass},
\code{\link{upload_with_scp}}
}
\author{
Stuart K. Grange
}
