% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_columns.R
\name{drop_na_columns}
\alias{drop_na_columns}
\title{Function to drop columns in a data frame which only contain \code{NA}.}
\usage{
drop_na_columns(df)
}
\arguments{
\item{df}{Data frame with \code{NA} columns.}
}
\description{
Function to drop columns in a data frame which only contain \code{NA}.
}
\author{
teucer
}
\seealso{
\href{http://stackoverflow.com/questions/2643939/remove-columns-from-dataframe-where-all-values-are-na}{stackoverflow.com}
}

