% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_read.R
\name{excel_read}
\alias{excel_read}
\title{Function to wrap \code{readxl::read_excel} and then do some header cleaning.}
\usage{
excel_read(file, sheet = 1, col_names = TRUE, col_types = NULL, na = "",
  skip = 0, clean = TRUE, convert = FALSE)
}
\description{
Function to wrap \code{readxl::read_excel} and then do some header cleaning.
}

