% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{file_metadata}
\alias{file_metadata}
\title{Function to extract file metadata with \code{exiftool}.}
\usage{
file_metadata(file, json = FALSE, progress = "none")
}
\arguments{
\item{json}{Should the return be a pretty printed JSON string?}

\item{progress}{What type of progress bar should the funciton display? Default
is \code{"none"}.}

\item{files}{A vector of files to extract metadata from.}
}
\description{
\code{file_metadata} uses Phil Harvey's Perl \code{exiftool} to extract a 
file's metadata and then formats the output into a table which can be 
conveniently used within R.
}
\details{
\code{exiftool} is called as a system command and this function has not been
tested on other non-Debian systems. I do not know how \code{exiftool} is 
implemented on Windows or OS X and therefore I cannot guarantee that this 
function will work.

\code{file_metadata} can be useful for media files and the exploration of 
metadata.

All variables which are supported and returned by \code{exiftool} will be 
present. Watch out for very large numbers of variables when imputing large
XML or JSON documents with a large numbers of nested elements.
}
\examples{
\dontrun{
data_metadata <- file_metadata("music.mp3")
}

}
\author{
Stuart K. Grange
}
\seealso{
\link{http://www.sno.phy.queensu.ca/~phil/exiftool/}
}

