% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_volume_to_mass.R
\name{gas_volume_to_mass}
\alias{gas_mass_to_volume}
\alias{gas_volume_to_mass}
\title{Functions to convert gas concentrations between volume and mass unit systems.}
\usage{
gas_volume_to_mass(volume, gas, molecular_mass = NA, unit_input = "ppb",
  unit_output = "ug_m3", temperature = 0, pressure = 101325)

gas_mass_to_volume(mass, gas, molecular_mass = NA, unit_input = "ug_m3",
  unit_output = "ppb", temperature = 0, pressure = 101325)
}
\arguments{
\item{volume}{Gas concentration in ppb or ppm.}

\item{gas}{A string such as \code{"o3"}, \code{"co"}, or \code{"co2"} which 
represents the input gas. This is used to find the molecular mass of a gas. 
Every gas-species will never be supported, therefore, use the 
\code{molecular_mass} argument if this needs to be overridden.}

\item{molecular_mass}{Molecular mass of gas in g mol-1. For example, for 
ozone, this is 48.}

\item{unit_input}{}

\item{unit_output}{}

\item{temperature}{Default is 0 degrees Celsius. However, values such as 20
or 25 degrees Celsius are often used too.}

\item{pressure}{Default is 101325 Pa (a standard atmosphere).}

\item{mass}{Gas concentration in ug m-3 or mg m-3.}
}
\description{
Use \code{gas_volume_to_mass} when a gas concentration is in ppb or ppm and 
ug m-3 or mg m-3 is desired. 
Use \code{gas_mass_to_volume} when a gas concentration is in ug m-3 or mg m-3
and ppb or ppm is desired.
}
\details{
These functions allow for input of temperatures and pressures which change the
coefficients used for the conversion between the unit systems.

To-do: Clean up documentation!
}
\examples{
\dontrun{
gas_volume_to_mass(400, "o3")

gas_mass_to_volume(800, "o3")

}

}
\author{
Stuart K. Grange
}

