% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Function get exchange rates from Oanda.}
\usage{
get_exchange_rates(currencies, from = NA, rename = TRUE)
}
\arguments{
\item{currencies}{Currency-pairs to get. See examples for format.}

\item{from}{Date from when exchange rates should be returned from. The date
format should be \code{yyyy-mm-dd}.}

\item{rename}{If \code{from} is used, should the exchange rate variable be
renamed as a generic "exchange_rate" rather than the currency pair string?
Default is TRUE.}
}
\description{
\code{get_exchange_rates} is a simple wrapper for \code{quantmod::getFX}. In
default usage, \code{get_exchange_rates} returns a single numeric value for
the current exchange rate from Oanda (\url{http://www.oanda.com/}). If 
a \code{from} date is specified, a data frame is returned with daily exchange
rates to the current day. Oanda only supports the retrieval of data up to 
five years from the current day.
}
\examples{

\dontrun{
# Get exhange rate of the New Zealand Dollar to the British Pound
nzd_to_gbp <- get_exchange_rates("NZD/GBP")

# On 2015-07-08
nzd_to_gbp
0.4289


# Get exchange rates since 2015-07-08
data_exchange_rates <- get_exchange_rates("NZD/GBP", from = "2015-07-08")

head(data_exchange_rates, 3)
      date exchange_rate
2015-07-08        0.4289
2015-07-09        0.4339
2015-07-10        0.4375
}

}

