% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_remote_file.R
\name{get_remote_file}
\alias{get_remote_file}
\title{Function to get/download file from an \code{URL}.}
\usage{
get_remote_file(file_remote, file_local, verbose = TRUE, mode = "w",
  method = "auto", sleep = FALSE)
}
\arguments{
\item{file_remote}{Vector of \code{URL}s/remote file names.}

\item{file_local}{Vector of local file names, the destination files for 
\code{file_remote}.}

\item{verbose}{Should the function give messages about download progress?}

\item{mode}{Mode with which to write the file.}

\item{method}{Method for downloading.}

\item{sleep}{If a number, this is the number of seconds to sleep between 
download iterations. This can help with keeping some webservers happy.}
}
\description{
Function to get/download file from an \code{URL}.
}
\seealso{
\code{\link{download.file}}
}
\author{
Stuart K. Grange
}
