% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_colours.R
\name{ggplot2_colours}
\alias{ggplot2_colours}
\title{Function to generate a vector of colours which emulate ggplot2's 
colours. 
 
http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
\usage{
ggplot2_colours(n = 2)
}
\description{
Function to generate a vector of colours which emulate ggplot2's 
colours. 
 
http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
\author{
John Colby
}

