% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme_threadr.R
\name{ggtheme_threadr}
\alias{ggtheme_threadr}
\alias{theme_super_minimal}
\alias{colour_byzantine_blue}
\alias{colour_crimson}
\alias{colour_inferno_purple}
\alias{colour_inferno_orange}
\alias{colour_inferno_peach}
\alias{colours_ggpubr}
\alias{colours_plotly}
\title{Function for "tufte-like" \strong{ggplot2} appearance.}
\usage{
ggtheme_threadr(grid_lines = FALSE)

theme_super_minimal(grid_lines = FALSE)

colour_byzantine_blue()

colour_crimson()

colour_inferno_purple()

colour_inferno_orange()

colour_inferno_peach()

colours_ggpubr()

colours_plotly()
}
\arguments{
\item{grid_lines}{Should the plot have grid lines?}
}
\value{
Invisible, modification to a \strong{ggplot2} plot.
}
\description{
Function for "tufte-like" \strong{ggplot2} appearance.
}
\examples{

\dontrun{

# Load package
library(ggplot2)

# Create data
data_example <- tibble(
  x = seq(1:3), 
  y = sample(1:10, 3)
)

# Plot with a tufte theme
ggplot(data_example, aes(x, y)) + 
  geom_point() + 
  ggtheme_threadr()
  
# Plot with a super minimal
ggplot(data_example, aes(x, y)) + 
  geom_point() + 
  theme_super_minimal()
  
}

}
\author{
Stuart K. Grange
}
