% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/join_many.R
\name{join_many}
\alias{join_many}
\title{Function for joining many data frames together. Also called a multi-merge.}
\usage{
join_many(dfs)
}
\arguments{
\item{dfs}{A \emph{list} of data frames.}
}
\description{
Function for joining many data frames together. Also called a multi-merge.
}
\examples{
\dontrun{
# Join three data frames together by a common key variable
data.join <- join_many(list(data.dates, data.min, data.max))
}
}
\author{
Stuart K. Grange
}

