% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_delta.R
\name{lag_delta}
\alias{lag_delta}
\alias{lead_delta}
\title{Functions to calculate the delta between a value and the lagged or leading 
value.}
\usage{
lag_delta(x)

lead_delta(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Functions to calculate the delta between a value and the lagged or leading 
value.
}
\examples{

lag_delta(sample(seq(0:10)))

}
\author{
Stuart K. Grange.
}
