% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{list_files_ftp}
\alias{list_files_ftp}
\title{Function to list files on an FTP or SFTP server.}
\usage{
list_files_ftp(
  url,
  credentials = "",
  sleep = NA,
  sort = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{url}{The url of remote directory.}

\item{credentials}{Credentials for a FTP or SFTP server. Do not use 
\code{credentials} if the server does not require authentication. 
\code{credentials} takes the format: \code{"username:password"}.}

\item{sleep}{Number of seconds to wait between querying server.}

\item{sort}{Should the files be sorted alphabetically?}

\item{verbose}{Should the function give messages? 

\code{\link{download_ftp_file}}, \code{\link{upload_to_ftp}}}
}
\description{
Function to list files on an FTP or SFTP server.
}
\examples{
\dontrun{

# Set credentials in this format
credentials <- "username:password"

# List files on a ftp server
list_files_ftp("ftp://195.174.23.76/test", credentials)

}

}
\author{
Stuart K. Grange
}
