% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locf.R
\name{locf}
\alias{locf}
\title{Function for carring last observation fowards.}
\usage{
locf(x, na.rm = FALSE)
}
\description{
Used to replace each NA with the most recent non-NA prior to it.
}
\details{
A wrapper for zoo::na.locf with na.rm defaulting to FALSE.
}
\author{
Stuart K. Grange
}

