% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locf.R
\name{locf}
\alias{locf}
\title{Function for carrying last observation forwards.}
\usage{
locf(x, na.rm = FALSE)
}
\description{
Used to replace each \code{NA} with the most recent non-\code{NA} prior to it.
}
\details{
A simple wrapper for \code{zoo::na.locf} with \code{na.rm} defaulting to 
\code{FALSE}
}
\examples{
\dontrun{
# Push observations forwards
data_income$income_chunk <- locf(data_income$income_chunk)
}

}
\author{
Stuart K. Grange
}

