% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{miles_to_km}
\alias{bar_to_psi}
\alias{celsius_to_fahrenheit}
\alias{cubic_centimetres_to_cubic_inches}
\alias{cubic_inches_to_cubic_centimetres}
\alias{fahrenheit_to_celsius}
\alias{foot_pound_to_newton_meter}
\alias{foot_to_metre}
\alias{heat_index}
\alias{hp_to_kw}
\alias{imperial_gallon_to_litres}
\alias{inch_to_mm}
\alias{kg_to_pound}
\alias{kg_to_stone}
\alias{km_h_to_km_min}
\alias{km_h_to_min_km}
\alias{km_h_to_ms}
\alias{km_to_miles}
\alias{km_to_nmile}
\alias{kw_to_hp}
\alias{l_100_km_to_km_l}
\alias{l_100km_to_mpg}
\alias{lb_to_kg}
\alias{litres_to_imperial_gallon}
\alias{litres_to_us_gallon}
\alias{metre_to_foot}
\alias{miles_to_km}
\alias{min_km_to_km_h}
\alias{mm_to_inch}
\alias{mpg_to_km_l}
\alias{mpg_to_l_100_km}
\alias{ms_to_km_h}
\alias{n_mile_to_km}
\alias{newton_metre_to_foot_pound}
\alias{newton_metre_to_kg_force}
\alias{pound_to_kg}
\alias{psi_to_bar}
\alias{stone_to_kg}
\alias{us_gallon_to_litres}
\title{Functions to convert unit systems.}
\usage{
miles_to_km(x)

km_to_miles(x)

n_mile_to_km(x)

km_to_nmile(x)

ms_to_km_h(x)

km_h_to_ms(x)

km_h_to_km_min(x)

km_h_to_min_km(x)

min_km_to_km_h(x)

inch_to_mm(x)

mm_to_inch(x)

foot_to_metre(x)

metre_to_foot(x)

imperial_gallon_to_litres(x)

litres_to_imperial_gallon(x)

us_gallon_to_litres(x)

litres_to_us_gallon(x)

mpg_to_l_100_km(mpg, type = "imperial")

mpg_to_km_l(mpg, type = "imperial")

l_100_km_to_km_l(x)

l_100km_to_mpg(x, type = "imperial")

cubic_inches_to_cubic_centimetres(x)

cubic_centimetres_to_cubic_inches(x)

bar_to_psi(x)

psi_to_bar(x)

fahrenheit_to_celsius(x)

celsius_to_fahrenheit(x)

heat_index(temp, rh, unit = "c")

kw_to_hp(x, metric = FALSE)

hp_to_kw(x, metric = FALSE)

newton_metre_to_kg_force(x)

newton_metre_to_foot_pound(x)

foot_pound_to_newton_meter(x)

pound_to_kg(x)

lb_to_kg(x)

kg_to_pound(x)

stone_to_kg(x)

kg_to_stone(x)
}
\description{
Functions to convert unit systems.
}
\author{
Stuart K. Grange
}

