% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{miles_to_km}
\alias{miles_to_km}
\alias{km_to_miles}
\alias{n_mile_to_km}
\alias{km_to_n_mile}
\alias{ms_to_km_h}
\alias{km_h_to_ms}
\alias{km_h_to_km_min}
\alias{km_h_to_min_km}
\alias{min_km_to_km_h}
\alias{inch_to_mm}
\alias{mm_to_inch}
\alias{foot_to_metre}
\alias{feet_to_metres}
\alias{metre_to_foot}
\alias{metres_to_feet}
\alias{yard_to_metre}
\alias{metre_to_yard}
\alias{gallon_to_litre}
\alias{litre_to_gallon}
\alias{mpg_to_km_l}
\alias{mpg_to_l_100_km}
\alias{l_100_km_to_km_l}
\alias{l_100km_to_mpg}
\alias{cubic_inches_to_cubic_centimetres}
\alias{cubic_centimetres_to_cubic_inches}
\alias{bar_to_psi}
\alias{psi_to_bar}
\alias{inch_hg_to_mb}
\alias{mb_to_inch_hg}
\alias{pascal_to_psi}
\alias{psi_to_pascal}
\alias{fahrenheit_to_celsius}
\alias{celsius_to_fahrenheit}
\alias{heat_index}
\alias{kw_to_hp}
\alias{hp_to_kw}
\alias{newton_metre_to_kg_force}
\alias{newton_metre_to_foot_pound}
\alias{foot_pound_to_newton_meter}
\alias{pound_to_kg}
\alias{lb_to_kg}
\alias{kg_to_pound}
\alias{stone_to_kg}
\alias{kg_to_stone}
\alias{acre_to_metre}
\alias{metre_to_acre}
\alias{marathon}
\alias{half_marathon}
\alias{absolute_zero}
\alias{standard_atmosphere}
\alias{miles_h_to_ms}
\alias{miles_h_s_2_to_ms_2}
\title{Functions to convert unit systems.}
\usage{
miles_to_km(x)

km_to_miles(x)

n_mile_to_km(x)

km_to_n_mile(x)

ms_to_km_h(x)

km_h_to_ms(x)

km_h_to_km_min(x)

km_h_to_min_km(x)

min_km_to_km_h(x)

inch_to_mm(x)

mm_to_inch(x)

foot_to_metre(x)

feet_to_metres(x)

metre_to_foot(x)

metres_to_feet(x)

yard_to_metre(x)

metre_to_yard(x)

gallon_to_litre(x, type = "imperial")

litre_to_gallon(x, type = "imperial")

mpg_to_km_l(x, type = "imperial")

mpg_to_l_100_km(x, type = "imperial")

l_100_km_to_km_l(x)

l_100km_to_mpg(x, type = "imperial")

cubic_inches_to_cubic_centimetres(x)

cubic_centimetres_to_cubic_inches(x)

bar_to_psi(x)

psi_to_bar(x)

inch_hg_to_mb(x)

mb_to_inch_hg(x)

pascal_to_psi(x)

psi_to_pascal(x)

fahrenheit_to_celsius(x)

celsius_to_fahrenheit(x)

heat_index(temp, rh, unit = "c")

kw_to_hp(x, metric = FALSE)

hp_to_kw(x, metric = FALSE)

newton_metre_to_kg_force(x)

newton_metre_to_foot_pound(x)

foot_pound_to_newton_meter(x)

pound_to_kg(x)

lb_to_kg(x)

kg_to_pound(x)

stone_to_kg(x)

kg_to_stone(x)

acre_to_metre(x)

metre_to_acre(x)

marathon()

half_marathon()

absolute_zero()

standard_atmosphere()

miles_h_to_ms(x)

miles_h_s_2_to_ms_2(x)
}
\arguments{
\item{x}{Input vector.}

\item{type, metric}{Type options.}

\item{temp, rh, unit}{Specific options for some functions.}
}
\description{
No rounding is ever done.
}
\author{
Stuart K. Grange
}
