% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_to_string.R
\name{parse_no_hour_time}
\alias{parse_no_hour_time}
\title{Function to parse a time with greater than 60 minuntes to a period.}
\usage{
parse_no_hour_time(x)
}
\arguments{
\item{x}{Vector of times in no hour format.}
}
\value{
Vector of periods; a \strong{lubridate} data type.
}
\description{
\code{parse_no_hour_time} will format and parse a time such as \code{124.17}
(124 minutes and 17 seconds) to a period with hour, minutes, and seconds.
}
\examples{
\dontrun{

# Parse a times with no hour
parse_no_hour_time(c(124.17, "89:22"))

}

}
\seealso{
\code{\link{period_to_date}}, \code{\link{period_to_string}}
}
\author{
Stuart K. Grange
}
