% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_change.R
\name{percent_lost}
\alias{percent_lost}
\title{Function to calculate the percentage lost between two values.}
\usage{
percent_lost(value, value_two)
}
\arguments{
\item{value}{Numeric vector of values.}

\item{value_two}{Numeric vector of values.}
}
\value{
Numeric vector with length of \code{value}.
}
\description{
Function to calculate the percentage lost between two values.
}
\examples{

# Check the calculation, loss 80 from 100
percent_lost(100, 80)

# Take 1693.2 from 33864
percent_lost(33864, 1693.2)

}
\author{
Stuart K. Grange
}
