% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_change.R
\name{percentage_difference}
\alias{percentage_difference}
\title{Function to calculate the percentage difference between two values.}
\usage{
percentage_difference(value, value_two)
}
\arguments{
\item{value}{Numeric vector of values.}

\item{value_two}{Numeric vector of values.}
}
\value{
Numeric vector with length of \code{value}.
}
\description{
Function to calculate the percentage difference between two values.
}
\examples{

# Calculate some differences
percentage_difference(100, 80)
percentage_difference(104, 130)

# A percentage difference's sign changes when the order of the inputs change,
# but not the value
percentage_difference(104, 130)
percentage_difference(130, 104)

}
\seealso{
\code{\link{percentage_change}}
}
\author{
Stuart K. Grange
}
