% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_variation.R
\name{plot_time_variation}
\alias{plot_time_variation}
\title{Function to plot hourly, weekday, and monthly averages.}
\usage{
plot_time_variation(df, by = NA, colours = NA, ylim = c(NA, NA))
}
\arguments{
\item{df}{Input data frame with at least two variables, two of which are 
named \code{value} and \code{date}.}

\item{by}{A grouping variable to separately plot different variables.}

\item{colours}{A vector of colours to use for the plotting.}

\item{ylim}{Limits for y-axes.}
}
\value{
A \strong{ggplot2} list object containing four plots.
}
\description{
Function to plot hourly, weekday, and monthly averages.
}
\author{
Stuart K. Grange
}
