% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bibtex.R
\name{read_bibtex}
\alias{read_bibtex}
\alias{read_bibtex_aux}
\title{Function to load a \code{BibTeX} file to a data frame.}
\usage{
read_bibtex(file, skip = 0, article = NA, progress = "none")

read_bibtex_aux(file)
}
\arguments{
\item{file}{File name of a \code{BibTeX} file. URLs work too.}

\item{skip}{How many lines to skip before starting reading? This is useful
if \code{file} contains a non-commented preamble.}

\item{article}{A vector of \code{BibTeX} keys to filter \code{file} to. If 
not used, all entries will be returned.}

\item{progress}{Progress bar to display, only useful for large files.}
}
\value{
Data frame.

Character vector.
}
\description{
Function to load a \code{BibTeX} file to a data frame.

\code{read_bibtex_aux} reads \code{.aux} files created by \code{BibTeX} and 
\code{BibLaTeX}.
}
\examples{
\dontrun{

# A file from the web
url <- "http://www.andy-roberts.net/res/writing/latex/sample.bib"
data_bib <- read_bibtex(url)

# To json
to_json(data_bib)

}

}
\author{
Stuart K. Grange

Stuart K. Grange
}
