% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bibtex.R
\name{read_bibtex}
\alias{read_bibtex}
\title{Function to load a BibTeX file to a data frame.}
\usage{
read_bibtex(file, article = NA, progress = "none")
}
\arguments{
\item{file}{File name of a BibTeX file.}

\item{article}{A vector of BibTeX keys to filter \code{file} to. If not used,
all entries will be returned.}

\item{progress}{Progress bar to display.}
}
\value{
Data frame.
}
\description{
Function to load a BibTeX file to a data frame.
}
\author{
Stuart K. Grange
}

