% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_html_tables.R
\name{read_html_tables}
\alias{read_html_tables}
\title{Function to read HTML tables from an URL.}
\usage{
read_html_tables(url, n = NA, character = FALSE)
}
\arguments{
\item{url}{A URL which contains HTML tables.}

\item{n}{Table-number to return. Optional, but this will return a single
data frame rather than a list of data frames which is usually what is 
desired.}

\item{character}{Is \code{url} a character string? This is useful when the 
HTML has been read and stored previously before being used by this function.}
}
\description{
Function to read HTML tables from an URL.
}
\examples{
\dontrun{
# A url
url <- "https://en.wikipedia.org/wiki/List_of_London_Underground_stations"

# A list
list_tables <- read_html_tables(url)
length(list_tables)

# A data frame
df <- read_html_tables(url, n = 1)

}

}
\author{
Stuart K. Grange
}

