% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_json}
\alias{read_json}
\title{Function to read JSON files.}
\usage{
read_json(file, flatten = TRUE)
}
\arguments{
\item{file}{File name of JSON file.}

\item{flatten}{Should the JSON file be flattened? Default is \code{TRUE}.}
}
\description{
Very simple wrapper for \code{jsonlite::fromJSON}.
}
\author{
Stuart K. Grange
}

