% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lines.R
\name{read_lines}
\alias{read_lines}
\title{Function to wrap \code{\link{readLines}} with different defaults.}
\usage{
read_lines(connection, n = -1, collapse = FALSE, warn = FALSE)
}
\arguments{
\item{connection}{File name, URL, or connection to load.}

\item{n}{Number of lines to read. Default is all (\code{-1}).}

\item{collapse}{Should the vector which is loaded be collapsed to a length of
\code{1}.}

\item{warn}{Should \code{readLines} give warning? Default is \code{FALSE}.}
}
\description{
Function to wrap \code{\link{readLines}} with different defaults.
}
\author{
Stuart K. Grange
}
