% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round_date_interval.R
\name{round_date_interval}
\alias{round_date_interval}
\title{Function to round dates to arbitrary time intervals such as 5-, 10-, 15-, or
30-minute periods.}
\usage{
round_date_interval(date, interval = "5 min")
}
\arguments{
\item{date}{A vector of dates to be rounded. This variable must be named
\code{'date'}.}

\item{interval}{A numeric or string value which represents the time interval
in minutes.}
}
\description{
round_date_interval rounds dates to the nearest arbitrary time interval
unlike xts::align.time which always pushes dates to the next interval. Dates
that fall in the centre of the time interval are rounded down.
}
\details{
round_date_interval works by first calculating an origin which is the first
instant of the date's hour. round_date_interval then extracts the
seconds-past-the-origin of the date and rounds this value to the interval's
multiple. The rounded value and the origin are then summed and returned.

round_date_interval does not make any adjustments for duplicated dates in the
input or output.

round_date_interval can be used for rounding dates to the nearest whole
minute or hour, however, lubridate::round_date and related functions may be
more convenient.

Currently round_date_interval cannot deal with multiples other than minutes.
Intervals such as 2 hours will not work yet.

The \code{interval} argument can take 'friendly' forms such as
\code{'5 mins'} or \code{'15 min'} as well as \code{5} or \code{15}.
}
\examples{
\dontrun{
}
}
\author{
Stuart K. Grange
}
\seealso{
See \code{\link{time_pad}}, \code{\link{lubridate::round_date}}
}

