% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{season}
\alias{season}
\title{Function to determine season of a date.}
\usage{
season(date, hemisphere = "northern", as.factor = FALSE,
  as.title = FALSE)
}
\arguments{
\item{date}{\code{POSIXct} or \code{Date} vector.}

\item{hemisphere}{Which hemisphere to use? Can be \code{"northern"} or 
\code{"southern"}.}

\item{as.factor}{Should the return be an ordered factor, not a numeric 
vector?}

\item{as.title}{When \code{as.factor} is \code{TRUE}, should the labels be
in "title-case"?}
}
\value{
Numeric or ordered factor vector with the length of \code{date}.
}
\description{
The season coding is based on months of year.
}
\author{
Stuart K. Grange
}
