% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_humidity.R
\name{specific_humidity}
\alias{specific_humidity}
\title{Function to calculate specific humidity.}
\usage{
specific_humidity(temp, rh, pressure = 1013.25)
}
\arguments{
\item{temp}{Temperature degrees Celsius.}

\item{rh}{Relative humidity in percentage.}

\item{pressure}{Pressure in hectopascals (hPa).}
}
\value{
Numeric vector, specific humidity in \code{g.kg-1}.
}
\description{
Specific humidity is the mass of water vapour in grams per kilogram of moist 
air.
}
\examples{

# Calculate specific humidities at different relative humidities
specific_humidity(20, 95:100)

}
\seealso{
\code{\link{absolute_humidity}}, \code{\link{calculate_dew_point}},
\code{\link{calculate_vapour_pressure}}, \code{\link{calculate_mixing_ratio}}
}
\author{
Stuart K. Grange
}
