% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_nrow.R
\name{split_nrow}
\alias{split_nrow}
\title{Function to split a data frame into a list with \emph{n} elements based on 
number of rows.}
\usage{
split_nrow(df, rows)
}
\arguments{
\item{df}{Data frame to be split.}

\item{rows}{Number of rows for \code{df} to be split by.}
}
\value{
list with \emph{n} elements.
}
\description{
Function to split a data frame into a list with \emph{n} elements based on 
number of rows.
}
\examples{
\dontrun{
# Split a large data frame into chunks
data_large_list <- split_nrow(data_large, 500000)

}

}
\author{
Stuart K. Grange
}

