% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_date.R
\name{str_date}
\alias{str_date}
\title{Function to return the system's idea of the date as a character string.}
\usage{
str_date(time = TRUE, tz = TRUE, underscore = FALSE)
}
\description{
Function to return the system's idea of the date as a character string.
}
\author{
Stuart K. Grange
}

