% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_functions.R
\name{str_proper_case}
\alias{str_date}
\alias{str_proper_case}
\alias{str_rm_non_ascii}
\alias{str_sentence_case}
\alias{str_trim_length}
\title{Functions to preform string operations which do not currently exist in
\code{stringr}.}
\usage{
str_proper_case(x)

str_date(time = TRUE, tz = TRUE, underscore = FALSE)

str_rm_non_ascii(x)

str_trim_length(string, length)

str_sentence_case(x)
}
\description{
\code{str_proper_case} capitalises the first letter of every word in a string.
\code{str_proper_case} is a wrapper for \code{stringi::stri_trans_totitle}.
}
\details{
\code{str_rm_non_ascii} removes all non-ASCII characters from a string.

\code{str_date} returns the system's idea of the date as a character string.

\code{str_trim_length} trims strings to certain character lengths.

\code{str_sentence_case} capitalises the first letter in a string and makes
all other characters lowercase.
}
\examples{
\dontrun{
# Some strings
string <- c("auckland", "wellington", "berlin")

# Make proper case
str_proper_case(string)
"Auckland"   "Wellington" "Berlin"
}
}
\author{
Stuart K. Grange
}

