% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_functions.R
\name{str_proper_case}
\alias{str_chop}
\alias{str_date}
\alias{str_drop_xml_tags}
\alias{str_proper_case}
\alias{str_rm_brackets_and_contents}
\alias{str_rm_non_ascii}
\alias{str_sentence_case}
\alias{str_trim_length}
\alias{str_trim_many_spaces}
\alias{str_underscore}
\title{Functions to preform string operations which do not currently exist in
\code{stringr}.}
\usage{
str_proper_case(x)

str_date(time = TRUE, tz = TRUE, underscore = FALSE)

str_rm_non_ascii(x)

str_trim_length(string, length)

str_sentence_case(x)

str_underscore(x)

str_trim_many_spaces(x)

str_chop(string, n)

str_drop_xml_tags(string)

str_rm_brackets_and_contents(x)
}
\description{
\code{str_proper_case} capitalises the first letter of every word in a string. 
\code{str_proper_case} is a wrapper for \code{stringi::stri_trans_totitle}.
}
\details{
\code{str_rm_non_ascii} removes all non-ASCII characters from a string.

\code{str_date} returns the system's idea of the date as a character string.

\code{str_trim_length} trims strings to a certain length, for example 30 
characters.

\code{str_sentence_case} capitalises the first letter in a string and makes
all other characters lowercase. This is different than \code{str_proper_case}.

\code{str_underscore} converts CamelCase and period.separated strings to
lower-case underscore_separated strings.

\code{str_trim_many_spaces} will remove excess spaces between words in a 
string.

\code{str_chop} will chop a string into a vector of fixed-width lengths.

\code{str_drop_xml_tags} removes XML tags from strings.

\code{str_rm_brackets_and_contents} will erase brackets (\code{(} and \code{)})
and characters within them.
}
\author{
Stuart K. Grange
}

