% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_functions.R
\name{str_proper_case}
\alias{str_chop}
\alias{str_create_na}
\alias{str_date}
\alias{str_drop_xml_tags}
\alias{str_extract_digits}
\alias{str_nth_character}
\alias{str_proper_case}
\alias{str_rm_brackets_and_contents}
\alias{str_rm_non_ascii}
\alias{str_sentence_case}
\alias{str_sql_quote}
\alias{str_thousands_separator}
\alias{str_trim_length}
\alias{str_trim_many_spaces}
\alias{str_underscore}
\alias{str_unique}
\title{Functions to preform string operations which do not currently exist in
\code{stringr}.}
\usage{
str_proper_case(x)

str_date(time = TRUE, tz = TRUE, underscore = FALSE)

str_rm_non_ascii(x)

str_trim_length(string, length)

str_sentence_case(x)

str_underscore(x)

str_trim_many_spaces(x)

str_chop(string, n)

str_drop_xml_tags(string)

str_rm_brackets_and_contents(x, type = "round")

str_extract_digits(x, as.numeric = TRUE)

str_sql_quote(x, collapse = TRUE)

str_unique(x)

str_nth_character(x, n)

str_create_na(x)

str_thousands_separator(x, sep = " ")
}
\description{
\code{str_proper_case} capitalises the first letter of every word in a string.
}
\details{
\code{str_rm_non_ascii} removes all non-ASCII characters from a string.

\code{str_date} returns the system's idea of the date as a character string.

\code{str_trim_length} trims strings to a certain length of characters.

\code{str_sentence_case} capitalises the first letter in a string and makes
all other characters lowercase. This is different than \code{str_proper_case}.

\code{str_underscore} converts CamelCase and period.separated strings to
lower-case underscore_separated strings.

\code{str_trim_many_spaces} will remove excess spaces between words in a 
string.

\code{str_chop} will chop a string into a vector of fixed-width lengths.

\code{str_drop_xml_tags} removes XML tags from strings.

\code{str_rm_brackets_and_contents} will erase brackets (\code{(} and \code{)}
or \code{[} and \code{]}) and characters within them.

\code{str_extract_digits} will extract digits/numbers from a string and convert
to a numeric data class if desired.

\code{str_sql_quote} will add single quotes can collapse string vectors, a 
common step for building SQL statements.

\code{str_unique} will find unique characters in a string.

\code{str_nth_character} will return a single character(s) from position(s) 
for a string.

\code{str_create_na} will convert "NA" and "" into \code{NA}.

\code{str_thousands_separator} will add thousands separators to a string or 
numeric vector.
}
\author{
Stuart K. Grange
}

