% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streak.R
\name{streak}
\alias{streak}
\title{Function to determine streaks/consecutive events.}
\usage{
streak(event, only_true = TRUE)
}
\arguments{
\item{event}{A logical vector indicating if an event occurred or not.}

\item{only_true}{Should only events that are \code{TRUE} be returned as 
streaks?}
}
\value{
Numeric vector with the length of \code{event}.
}
\description{
Function to determine streaks/consecutive events.
}
\seealso{
\code{\link{rle}}
}
\author{
Stuart K. Grange
}
