% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_to_observations.R
\name{strings_to_observations}
\alias{strings_to_observations}
\title{Functions to transform a delimited string into observations (rows) in a data
frame.}
\usage{
strings_to_observations(df, variable, pattern)
}
\arguments{
\item{df}{Data frame which contains a delimited string variable.}

\item{variable}{Name of variable in \code{df} which is, or contains delimited 
strings.}

\item{pattern}{Delimiter to split the string. Common values are \code{","} and
\code{"/"}.}
}
\description{
\code{strings_to_observations} works by first splitting a string variable by a 
pattern. The split object is then used to replicate all other variables in a
data frame. \code{strings_to_observations} is useful for making data tidy.
}
\examples{
\dontrun{
# Load data
french_verbs

# Make tidy data
data_tidy <- string_to_observations(french_verbs, variable = "subject", 
                                    pattern = ",")

# Print
data_tidy

}

}
\author{
Stuart K. Grange
}

