% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_less_minimal.R
\name{theme_less_minimal}
\alias{theme_less_minimal}
\alias{theme_narrow_strips}
\alias{theme_x_label_rotate}
\title{Function for a minimal \strong{ggplot2} appearance but with bounding lines 
for plot area.}
\usage{
theme_less_minimal(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  narrow_strips = FALSE,
  x_label_rotate = FALSE
)

theme_narrow_strips(strip_margin = 2)

theme_x_label_rotate(angle = 45)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base size for line elements.}

\item{base_rect_size}{Base size for rectangle elements.}

\item{narrow_strips}{Should the plot have narrow strips for facet headings?}

\item{x_label_rotate}{Should the plot have rotated x-axis ticks?}

\item{strip_margin}{Margins for strip headings.}

\item{angle}{Angle of text.}
}
\value{
Invisible, modification to a \strong{ggplot2} plot.
}
\description{
Function for a minimal \strong{ggplot2} appearance but with bounding lines 
for plot area.
}
\examples{

\dontrun{

# Load package
library(ggplot2)

# Create data
data_example <- tibble(
  x = seq(1:3), 
  y = sample(1:10, 3),
  facet_heading = "The plot"
)

# Plot with the theme
ggplot(data_example, aes(x, y)) + 
  geom_point() + 
  facet_wrap("facet_heading") + 
  theme_less_minimal()
  
# Plot with the theme with some options
ggplot(data_example, aes(x, y)) + 
  geom_point() + 
  facet_wrap("facet_heading") + 
  theme_less_minimal(narrow_strips = TRUE, x_label_rotate = TRUE)

}

}
\author{
Stuart K. Grange
}
