% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_less_minimal.R
\name{theme_less_minimal}
\alias{theme_less_minimal}
\alias{theme_narrow_strips}
\alias{theme_x_label_rotate}
\alias{ggplot_caption_format}
\title{Function for a minimal \strong{ggplot2} appearance but with bounding lines 
for plot area.}
\usage{
theme_less_minimal(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  narrow_strips = FALSE,
  x_label_rotate = 0,
  legend_position = NA
)

theme_narrow_strips(x_strip_margin = 2.2, y_strip_margin = 2.5)

theme_x_label_rotate(angle = 45)

ggplot_caption_format(location = "left", size = NULL)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base size for line elements.}

\item{base_rect_size}{Base size for rectangle elements.}

\item{narrow_strips}{Should the plot have narrow strips for facet headings?}

\item{x_label_rotate}{Angle for x-axes labels.}

\item{legend_position}{The position of the legend ("none", "left", "right", 
"bottom", "top", or two-element numeric vector).}

\item{angle}{Angle of text.}

\item{location, size}{Specific arguments for \code{ggplot_caption_format}.}

\item{strip_margin}{Margins for strip headings.}
}
\value{
Invisible, modification to a \strong{ggplot2} plot.
}
\description{
Function for a minimal \strong{ggplot2} appearance but with bounding lines 
for plot area.
}
\examples{

# Load package
library(ggplot2)

# Create data
data_example <- tibble::tibble(
  x = seq(1:3), 
  y = sample(1:10, 3),
  facet_heading = "The plot"
)

# Plot with the theme
ggplot(data_example, aes(x, y)) + 
  geom_point() + 
  facet_wrap("facet_heading") + 
  theme_less_minimal()
  
# Plot with the theme with some options
ggplot(data_example, aes(x, y)) + 
  geom_point() + 
  facet_wrap("facet_heading") + 
  theme_less_minimal(narrow_strips = TRUE, x_label_rotate = 45)
  
# Plot and move legend
ggplot(data_example, aes(x, y, colour = facet_heading)) + 
  geom_point() + 
  theme_less_minimal(legend_position = "bottom")
  
# Plot and move legend
ggplot(data_example, aes(x, y, colour = facet_heading)) + 
  geom_point() + 
  theme_less_minimal() + 
  labs(caption = "A caption") + 
  ggplot_caption_format(size = 10)

}
\author{
Stuart K. Grange
}
