% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidy_data_to_line.R
\name{tidy_data_to_line}
\alias{tidy_data_to_line}
\title{Function for converting tidy data into spatial lines.}
\usage{
tidy_data_to_line(df, latitude = "lat", longitude = "long")
}
\description{
Function for converting tidy data into spatial lines.
}
\author{
Stuart K. Grange
}

