% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_summary.R
\name{tidy_summary}
\alias{tidy_summary}
\title{Function to create tidy data descriptives from a data frame.}
\usage{
tidy_summary(df)
}
\arguments{
\item{df}{Data frame for descriptives to be calculated from.}
}
\description{
Only numeric variables are currently returned.
}
\examples{
\dontrun{
data_summary <- tidy_summary(data_air)

}

}
\author{
Stuart K. Grange
}

