% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dygraph.R
\name{time_dygraph}
\alias{time_dygraph}
\title{Function for plotting interactive time-series plots in a similar way to
\code{timePlot}.}
\usage{
time_dygraph(df, variable = "value", colour = "dodgerblue",
  range = TRUE, step = FALSE, points = FALSE, fill = FALSE,
  color = colour, ylab = NA, legend_width = 400, mouse_label = NA,
  tz = NA, window = NULL, group = NULL, y_reference = NA)
}
\arguments{
\item{df}{Data frame with a variable/column named \code{"date"}.}

\item{variable}{Variable in \code{df} to plot. Default is \code{"value"}. 
\code{variable} can be a vector but faceting is not yet supported.}

\item{colour, color}{Colour of plotted geometry. Default is 
\code{"dodgerblue"}.}

\item{range}{Should a range selector be plotted? Default is \code{TRUE}.}

\item{step}{Should the plot take a stepped look? Default is \code{FALSE}.}

\item{points}{Should the plot include points where observations are recorded? 
Default is \code{FALSE}.}

\item{fill}{Should the plot fill between the x-axis and the geometry? Default
is \code{FALSE}.}

\item{ylab}{String for y-axis label.}

\item{legend_width}{Width of legend in pixels. Default is \code{400}.}

\item{mouse_label}{String for legend label. If not used, label will take the
name of \code{variable}.}

\item{tz}{Timezone to display plot with. If not used, \code{tz} will use the
time-zone of the \code{df}'s \code{"date"} variable.}

\item{window}{An optional range window. Uses a date vector with a length of 
two.}

\item{group}{An optional grouping string for when multiple plots are to be 
connected when rendered in an HTML document.}

\item{y_reference}{Location of a y-axis reference line.}
}
\value{
Invisible, an interactive dygraph.
}
\description{
Function for plotting interactive time-series plots in a similar way to
\code{timePlot}.
}
\seealso{
\code{timePlot}, \code{\link{dygraph}}
}
\author{
Stuart K. Grange
}
