% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dygraph.R
\name{time_dygraph}
\alias{time_dygraph}
\title{Function for plotting interactive time-series plots in a similar way to
\code{\link{timePlot}}.}
\usage{
time_dygraph(df, variable = "value", colour = "red", range = TRUE,
  step = FALSE, points = FALSE, fill = FALSE, color = colour,
  ylab = NA, legend_width = 400, mouse_label = NA, tz = NA,
  window = NULL)
}
\arguments{
\item{df}{Data frame with a variable/column named \code{"date"}.}

\item{variable}{Variable(s) in \code{df} to plot. Default is \code{"value"}.}

\item{colour}{Coluor of plotted geometry.}

\item{range}{Should a range selector be plotted? Default is \code{TRUE}.}

\item{step}{Should the plot take a stepped look?}

\item{points}{Should the plot include points where observations are recorded?}

\item{fill}{Should the plot fill between the x-axis and the geometry?}

\item{ylab}{String for y-axis label.}

\item{legend_width}{Width of legend in pixels. Default is \code{400}.}

\item{mouse_label}{String for legend label. If not used, label will take the
name of \code{variable}.}

\item{tz}{Timezone to display plot with. If not used, \code{tz} will use the
time-zone of the \code{df}'s \code{"date"} variable.}

\item{window}{Default range window. Uses a date vector with a length of two.}
}
\description{
Function for plotting interactive time-series plots in a similar way to
\code{\link{timePlot}}.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{timePlot}}
}

