% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{time_zone}
\alias{time_zone}
\title{Function to extract time-zone from a POSIXct date.}
\usage{
time_zone(date)
}
\arguments{
\item{date}{Date vector.}
}
\value{
Vector with a length of one.
}
\description{
Function to extract time-zone from a POSIXct date.
}
\examples{
\dontrun{

# A date vector
time_zone(date_parsed)

# A variable in data frame
time_zone(data_test$date)

}

}
\author{
Stuart K. Grange
}
