% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twos_complement.R
\name{twos_complement}
\alias{twos_complement}
\title{Function to convert Two's Complement integer to a standard integer.}
\usage{
twos_complement(x)
}
\arguments{
\item{x}{Vector of integers.}
}
\value{
Numeric vector.
}
\description{
\code{twos_complement} is vectorised and will convert integers to-and-from
(8-bit) binary form to decode the Two's Complement rule.
}
\examples{
\dontrun{

# A hexadecimal number
hex <- "fff4d"

# Hexadecimal as a base 10 decimal
integer <- strtoi(hex, 16)
integer
1048397

# Decode the Two's Complement rule
twos_complement(integer)
-179

}

}
\author{
Stuart K. Grange
}
