% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ftp_file.R
\name{upload_to_ftp}
\alias{upload_to_ftp}
\title{Function to upload a files to an FTP or SFTP server.}
\usage{
upload_to_ftp(file, url, credentials = "", basename = FALSE,
  progress = "text")
}
\arguments{
\item{file}{File to upload.}

\item{url}{The url of remote directory.}

\item{credentials}{Credentials for a FTP or SFTP server. Do not use 
\code{credentials} if the server does not require authentication. 
\code{credentials} takes the format: \code{"username:password"}.}

\item{basename}{Should only the basename be used for transferring the files? This is
useful when files to be uploaded are outside of the working directory, but the
directory structure will be lost on the remote server.}

\item{progress}{Progress bar type. Default is \code{"text"}, for no bar, use
\code{"none"}.}
}
\description{
\code{upload_to_ftp} will create directories if necessary.
}
\examples{
\dontrun{

# Get file list
file_list <- list.files(recursive = TRUE, full.names = TRUE)

# Exclude some files
file_list <- grep("cache|.Rmd$", file_list, value = TRUE, invert = TRUE)

# Upload files to server
upload_to_ftp(file_list, url, "user:pass")

}

}
\seealso{
\code{\link{download_ftp_file}}, \code{\link{list_files_ftp}}
}
\author{
Stuart K. Grange
}
