% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_supplements.R
\name{wday_monday}
\alias{wday_monday}
\title{Function to get weekday number from a date where \code{1} is Monday and 
\code{7} is Sunday.}
\usage{
wday_monday(x)
}
\arguments{
\item{x}{Date vector.}
}
\value{
Numeric vector.
}
\description{
Function to get weekday number from a date where \code{1} is Monday and 
\code{7} is Sunday.
}
\author{
Stuart K. Grange
}
