% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_range.R
\name{within_range}
\alias{within_range}
\title{Function to test if a vector is within one, or many ranges.}
\usage{
within_range(vector, start, end)
}
\arguments{
\item{vector}{Vector to test.}

\item{end}{End of range.}

\item{begin}{Start of range.}
}
\value{
Logical vector of length of \code{vector}.
}
\description{
\code{within_range} works well for testing unix time dates.
}
\examples{
\dontrun{

# Test a single range
within_range(1:10, 8, 9)


# Test many ranges
# A vector
vector <- 1:10

# Build look-up table
look_up <- data.frame(
  start = c(1, 5),
  end = c(2, 9)
)

# Test vector
within_range(vector, look_up$start, look_up$end)

}
 
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{which}}
}

