% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.gpx.R
\name{write.gpx}
\alias{write.gpx}
\title{Function to write a data frame to a GPX file with usage analogous to
\code{write.table}.}
\usage{
write.gpx(df, file, latitude = "latitude", longitude = "longitude",
  name = NULL, layer = "points")
}
\arguments{
\item{df}{Data frame to be written to a GPX file.}

\item{file}{File name of GPX file.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}

\item{name}{Name of variable which will be added in the "name" element of the
GPX file.}

\item{layer}{Type of layer to be written to GPX file. Can either be
"points" or "lines". Default is "points".}
}
\description{
\code{write.gpx} uses \code{rgdal::writeOGR} as the GPX writer. Unlike the
standard \code{writeOGR} function, \code{write.gpx} will automatically expand
file paths and can overwrite previous files if necessary.
}
\examples{
\dontrun{
# Export a GPX file containing points
write.gpx(data.bus.stations, "~/Desktop/bus_stations.gpx")

# Export GPX file with lines
write.gpx(data.gpx.track, "~/Desktop/drive_to_bath.gpx", layer = "points")

}
}
\author{
Stuart K. Grange
}

