% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_database_table.R
\name{write_database_table}
\alias{write_database_table}
\title{Function to write an entire database table to disk.}
\usage{
write_database_table(db, table, nrow = 1e+06, file = "")
}
\arguments{
\item{db}{A database connection.}
}
\description{
write_database_table breaks up the queries and exports to small pieces to
allow for very large tables to be writen to disc with systems with limited
amounts of physical memory.
}
\examples{
# Pad time-series so every minute is present
\dontrun{
data.nelson.pad <- time_pad(data.nelson, pad.time = 'min', start.unit = 'day')
}
}
\author{
Stuart K. Grange
}

