% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_json.R
\name{write_json}
\alias{write_json}
\title{Function to write a data frame to a JSON file with usage analogous to
\code{write.table}.}
\usage{
write_json(df, file, pretty = TRUE, na = FALSE)
}
\arguments{
\item{df}{Data frame to be written as a JSON file.}

\item{file}{JSON file name.}

\item{pretty}{Should the JSON output be formatted for readability? Default is
\code{TRUE}.}

\item{na}{Should \code{NA} values be kept in the JSON file? Default is
\code{FALSE}. If \code{na} is \code{TRUE} then \code{NA}s will transformed to
JSON's \code{null} value.}
}
\description{
\code{write_json} uses \code{jsonlite} as the data frame to JSON parser.
}
\examples{
\dontrun{
write_json(data_car_identifiers, "car_identifiers.json")

}
}
\author{
Stuart K. Grange
}

