% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rds.R
\name{write_rds}
\alias{write_rds}
\alias{read_rds}
\title{Function to read and write R's RDS files with some helpful messaging.}
\usage{
write_rds(x, file, compress = TRUE, verbose = TRUE)

read_rds(file, verbose = TRUE)
}
\arguments{
\item{x}{An R object to write to disc.}

\item{file}{File name of RDS file.}

\item{compress}{Should the RDS file be written with compression?}

\item{verbose}{Should the function give messages? Note that the default is 
\code{TRUE}.}
}
\value{
Invisible x.
}
\description{
Function to read and write R's RDS files with some helpful messaging.
}
\author{
Stuart K. Grange
}
